Ficha3

Description

Este módulo contém as definições Haskell, para o cálculo de funções recursivas sobre listas e sobre inteiros, pedidas na Ficha 3 de Laboratórios de Informática I.

Documentation

remove

Arguments

:: [String]
argumento

-> Char
argumento

-> [String]
resultado

A função remove recebe uma lista de strings e remove todas as strings iniciadas por um dado caractere.

Notas:

Expand

Esta função é sensível a maiúsculas e minúsculas;

Esta função funciona para caracteres numéricos;

Esta função funciona para caracteres especiais;

Esta função funciona para listas vazias.

Exemplos de utilização:

Expand

>>> remove ["Ola", "Amor", "Comida"] 'A'
["Ola", "Comida"]
>>> remove ["Ola", "Amor", "Comida"] 'a'
["Ola", "Amor", "Comida"]
>>> remove [] 'a'
[]

Propriedades:

Expand

remove [] _ = []

addTuple

Arguments

:: [(Int, Int)]
argumento

-> Int
argumento

-> [(Int, Int)]
resultado

A função addTuple recebe uma lista de pares de inteiros e adiciona um valor dado à primeira componente de cada par.

Exemplos de utilização:

Expand

>>> addTuple [(1, 2), (-1, 1), (0, 0)] 5
[(6, 2), (4, 1), (5, 0)]
>>> addTuple [(1, 2), (-1, 1), (0, 0)] (-1)
[(0, 2), (-2, 1), (-1, 0)]
>>> addTuple [(1, 2), (-1, 1), (0, 0)] 0
[(1, 2), (-1, 1), (0, 0)]
>>> addTuple [] 9
[]

Propriedades:

Expand

addTuple [] _ = []

maxSndTuple

Arguments

:: [(Int, Int)]
argumento (lista não vazia)

-> Int
resultado

A função maxSndTuple recebe uma lista, não vazia, de pares de inteiros e calcula qual o maior valor da segunda componente.

Exemplos de utilização:

Expand

>>> maxSndTuple [(1, 2), (-1, 1), (0, 0)]
2
>>> maxSndTuple [(1, 2), (-1, 1), (0, 3)]
3

Propriedades:

Expand

maxSndTuple [(a, b)] = b