Generator

Documentation

generateRandoms :: Int -> Int -> [Int]

Given a seed returns a list of n integer randomly generated

nrAleatorio :: Int -> Int

Given a seed returns an integer randomly generated

subLista :: Int -> [a] -> [[a]]

convertePeca :: Int -> Piece

Converts an integer number into a Peca 3 = Comida Grangre 0 n < 7 <= Comida Pequena 7 <= 9 <= Parede

printCorridor :: Corridor -> String

Converts a Corredor to a string

converteCorredor :: [Int] -> Corridor

Converts a list of integers into a Corredor

converteLabirinto :: [[Int]] -> Maze

Converts a list of lists of integers into a Labirinto

geraLabirinto :: Int -> Int -> Int -> Maze

imprimeLabirinto :: Maze -> IO ()