Interface.Sprites

Documentation

data Sprite

Constructors

Sprite [Frame] Float Float

type Frame = Picture

type Rect = (Int, Int, Int, Int)

type SpriteData = (Float, [Rect])

Representa toda a informação necessária para criar um Sprite, exprimida num par (duração, frames), onde frames é uma lista de Rect para todas as frames deste Sprite.

getSprite :: Sprite -> Picture

Devolve a representação de um Sprite em Picture, que pode ser desenhada pelo Gloss.

makeSprite :: SpriteData -> DynamicImage -> Sprite

Gera um Sprite a partir de uma DynamicImage e de um SpriteData

updateSprite :: Float -> Sprite -> Sprite

Atualiza as animações dos Sprites

crop :: DynamicImage -> Rect -> DynamicImage

Recorta uma DynamicImage