Interface.ThreeD

Documentation

type Ponto3 = (Float, Float, Float)

Representa um ponto tridimensional, utilziado para desenhar a imagem

add3 :: Ponto3 -> Ponto3 -> Ponto3

Soma dois Ponto3

>>> (0,0,0) `add3` (2,3,5)
(2,3,5)

rotateX :: Float -> Ponto3 -> Ponto3

rotateY :: Float -> Ponto3 -> Ponto3

rotateZ :: Float -> Ponto3 -> Ponto3

data Objeto

Representa um objeto tridimensional

Constructors

Forma Color [Ponto3]

LineO Color [Ponto3]

Imagem Ponto3 Picture

Combinado [Objeto]

translateO :: Ponto3 -> Objeto -> Objeto

scale3 :: Float -> Float -> Float -> Objeto -> Objeto

scaleP :: Float -> Float -> Float -> Ponto3 -> Ponto3

projetarPonto :: Ponto3 -> Ponto3 -> Point

Projeta o ponto num plano 2d utilizando uma projeção ortogonal. https://en.wikipedia.org/wiki/Orthographic_projection

drawObjeto :: Ponto3 -> Objeto -> Picture

Converte um Objeto numa Picture que pode ser desenhada pelo Gloss.

pisoColor :: Piso -> Int -> Color

Devolve a Picture de um Piso

pecaToObjeto :: Peca -> Objeto

pistaToObjeto :: Pista -> Objeto

mapaToObjeto :: Mapa -> Objeto