Main

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2021/22.

Documentation

type Mundo = (Estado, Save, Imagens)

Esta tarefa consiste em criar os gráficos e a interação do jogo.

Neste Jogo inicialmente temos 3 opções. Jogar, Sair ou Instruções. Instruções - mostra quais as teclas que podemos usar e para que servem. Sair - abandonamos o jogo. Se selecionarmos Jogar vai levar nos para um sub Menu com duas opções. Níveis que está dividido em 3. Fácil, Médio e Díficil. Cada um deles tem 3 jogos, podemos fazer restart em qualquer um deles e também podemos salvar e continuar mais tarde. No final de concluir qualquer um dos modos de dificuldade aparece uma mensagem para terminar e sair. Contra-Relogio que existe um temporizador que vai decrementando. Por cada jogo que passamos ganhámos mais 60 segundos de cada vez. E quando o temporizador chegar a 0 aparece uma mensagem com a nossa pontuação e a melhor pontuação até ao momento e nesta mesma mensagem existe a opção de jogar de novo ou sair. Tanto na opção Níveis ou Contra-Relógio temos à escolha 8 jogadores diferentes e cada um deles com um fundo de jogo diferente.

estado global que estamos a usar

data Estado

estado mais particular

Constructors

Menu String

Ins String

Menu2 String

N String Int

J String Int

CR (Int, Jogo, Coordenadas, [Jogo], Int, Int, Int)

Mensagem2 String Int

Game Jogo Coordenadas [Jogo] Int Int

Mensagem String

type Save = (Int, Int, Int, Int)

para saber estados e valores acumulados

data Imagens

imagens divididas por categorias

Constructors

Imagens

Fields

pecas :: [(Peca, Picture)]

jogador :: [(Float, Picture)]

fundo :: [(Int, Picture)]

palavras :: [(String, Picture)]

niveis :: [(String, Picture)]

estadoInicial :: Estado

função que fornece o estado inicial

rectangulo1 :: Picture

dimensões do retângulo exterior usado no Menu e no Menu2 nos Níveis

rectangulo2 :: Picture

dimensões do retângulo interior usado no usado no Menu e no Menu2 nos Níveis

rectVermelho :: Picture

rectângulo vermelho usado como selecionador no Menu de Jogar e Sair

rectBranco :: Picture

rectângulo branco usado como selecionador no Menu de Jogar e Sair

rectVermelhoIns :: Picture

rectângulo vermelho usado como selecionador no Menu de Instruções

rectBrancoIns :: Picture

rectângulo branco usado como selecionador no Menu de Instruções

cimaJ1 :: Picture

a selecionar a palavra Jogar

cimaJ2 :: Picture

a não selecionar a palavra Jogar

baixoS1 :: Picture

a selecionar a palavra Sair

baixoS2 :: Picture

a não selecionar a palavra Sair

ladoS1 :: Picture

a selecionar a palavra Instruções

ladoS2 :: Picture

a não selecionar a palavra Instruções

rectIns1 :: Picture

dimensões do retângulo exterior usado no Ins

rectIns2 :: Picture

dimensões do retângulo interior usado no Ins

rectPretoIns :: Picture

rectângulo preto usado no Ins

ins1 :: Picture

rectângulo principal no Ins

rectangulo3 :: Picture

dimensões do retângulo exterior usado no Menu2 no CR

rectangulo4 :: Picture

dimensões do retângulo interior usado no Menu2 no CR

poligono :: Picture

desenho preenchido da seta

contorno1 :: Picture

desenho linha da seta

rectVermelho2 :: Picture

atribuição da cor vermelha ao rectângulo exterior no Menu2 de Níveis

rectBranco1 :: Picture

atribuição da cor branco ao rectângulo interior no Menu2 de Níveis

rectVermelho3 :: Picture

atribuição da cor vermelha ao rectângulo exterior no Menu2 de "Contra-Relógio"

rectPreto3 :: Picture

atribuição da cor preto ao rectângulo interior no Menu2 de "Contra-Relógio"

rectBranco2 :: Picture

atribuição da cor branco ao rectângulo interior no Menu2 de "Contra-Relógio"

setaV :: Picture

atribuição da cor vermelha da seta preenchida

contVermelho1 :: Picture

atribuição da cor vermelha da seta linha

cimaN1 :: Picture

rectângulo vermelho usado como selecionador no Menu2 de Níveis

cimaN2 :: Picture

rectângulo branco usado como não selecionador no Menu2 de Níveis

meioC1 :: Picture

rectângulo vermelho usado como selecionador no Menu2 de "Contra-Relógio"

meioC2 :: Picture

rectângulo preto e branco usado como não selecionador no Menu2 de "Contra-Relógio"

seta1 :: Picture

seta preenchida como selecionador

seta2 :: Picture

seta linha como selecionador

rectangulo5 :: Picture

dimensões do retângulo exterior usado no N de Fácil, Médio e Difícil

rectangulo6 :: Picture

dimensões do retângulo interior usado no N de Fácil, Médio e Difícil

rectVermelho4 :: Picture

atribuição da cor vermelha ao retângulo exterior usado no N de Fácil, Médio e Difícil

rectPreto4 :: Picture

atribuição da cor preto ao retângulo interior usado no N de Fácil, Médio e Difícil

rectN1 :: Picture

rectângulo a dizer Níveis

cimaN :: Picture

selecionador do nível Fácil

meioN :: Picture

selecionador do nível Médio

baixoN :: Picture

selecionador do nível Difícil

quadrado1 :: Picture

dimensões do quadrado exterior usado na Mensagem e Mensagem2

quadrado2 :: Picture

dimensões do quadrado interior usado na Mensagem e Mensagem2

contorno4 :: Picture

dimensões do selecionador usado na Mensagem e Mensagem2 de Sair

contorno5 :: Picture

dimensões do selecionador usado na Mensagem2 de "Novo Jogo"

mPreto :: Picture

atribuição da cor preta ao quadrado exterior usado na Mensagem e Mensagem2

mBranco :: Picture

atribuição da cor branco ao quadrado interior usado na Mensagem e Mensagem2

contVermelho4 :: Picture

atribuição da cor vermelha ao selecionador usado na Mensagem e Mensagem2 de Sair

contVermelho5 :: Picture

atribuição da cor vermelha ao selecionador usado na Mensagem2 de "Novo Jogo"

selecionadorSair :: Picture

selecionador de Sair usado na Mensagem e Mensagem2 de Sair

selecionadorNovo :: Picture

selecionador de "Novo Jogo" usado na Mensagem2 de "Novo Jogo"

segundosPmin :: Int -> String

função que converte segundos para uma string em minutos e segundos (m:s)

encontraPorta :: Jogo -> Coordenadas

função que descobre as coordenadas de uma porta

mapaToPicture :: Mundo -> [Picture]

função que transforma um Game e um CR para uma única imagem

descobreCentro :: Mapa -> Coordenadas

função que descobre as coordenadas centrais de um mapa

saveOrNot :: Mundo -> IO Mundo

função que verifica através dos inteiros s1, s2 e s3 se já foi dado o save nos Níveis Fácil, Médio e Difícil respetivamente e caso tenho dado retorna o estado onse se encontrava antes

descobreListaMapas :: Jogo -> [Jogo]

função utilizada na função "saveOrNot" para descobrir a lista de mapas que restava quando este deu save

escolheNivel :: Mundo -> IO Mundo

função que cria a 1ª vez que jogamos os Níveis e o CR

ajudaRestart :: Jogo -> Int

função que descobre o jogo em que se encontra para poder comecçar de novo (apenas no CR)

seguinte :: Mundo -> Mundo

função que fornece dependendo do jogo em que se encontra, o fim ou o próximo jogo (nos Níveis e no CR)

restart :: Mundo -> Mundo

função que através de um inteiro retorno as posiçoes iniciais de tudo no jogo

desenhaEstado :: Mundo -> IO Picture

função que converte um estado para uma imagem

reageEvento :: Event -> Mundo -> IO Mundo

função que através de um evento gera ou não um novo estado

reageTempo :: Float -> Mundo -> IO Mundo

função que através de um fr e gera ou não um novo estado

fr :: Int

função recebe um frame

dm :: Display

função que fornece o tamanho da janela

main :: IO ()

main

jogo1 :: Jogo

Jogo 1

jogo2 :: Jogo

Jogo 2

jogo3 :: Jogo

Jogo 3

jogo4 :: Jogo

Jogo 4

jogo5 :: Jogo

Jogo 5

jogo6 :: Jogo

Jogo 6

jogo7 :: Jogo

Jogo 7

jogo8 :: Jogo

Jogo 8

jogo9 :: Jogo

Jogo 9