Miniteste3

Description

Este módulo Haskell contém a resolução do Miniteste 3.

Documentation

type Coordenadas = (Float, Float)

Coordenadas = (abcissa, ordenada)

type Velocidade = (Float, Float)

Velocidade = (velocidade x, velocidade y)

type Tempo = Float

Tempo = segundos

type Circulos = [(Coordenadas, Float)]

Circulos = [(Coordenadas, raio)]

type Estado = (Coordenadas, Velocidade, Circulos, Tempo)

Representa o Estado do objeto jogável.

estadoInicial :: Estado

A função estadoInicial define o Estado inicial de todos os objetos do Jogo.

desenhaCirculos :: Circulos -> [Picture]

A função desenhaCirculos é uma função auxiliar que escreve todas as figuras circulares definidas numa lista Circulos.

Exemplos de utilização:

Expand

>>> desenhaCirculos [((0, 0), 15),((130, 80), 15)]
[translate 0.0 0.0 $ yellow circleSolid 15.0, translate 130.0 80.0 $ yellow circleSolid 15.0]

Propriedades:

Expand

desenhaCirculos [] = []

desenhaEstado :: Estado -> Picture

A função desenhaEstado desenha o Estado atual de Jogo:

Objeto jogável (cuja cor depende da velocidade y);

Circulos (recorrendo à função auxiliar desenhaCirculos);

Tempo (arredondado, no canto superior direito).

reageEvento :: Event -> Estado -> Estado

A função reageEvento reage aos Eventos que ocorram, nomeadamente às teclas KeyUp, KeyDown, KeyRight e KeyLeft.

Esta função faz com que o objeto jogável se mova, unicamente, dentro da janela.

Exemplos de utilização:

Expand

>>> reageEvento (EventKey (SpecialKey KeyUp) Down _ _) Estado
((x, min (y+5) 190),_,_,_)

Propriedades:

Expand

Caso o Evento registado não seja nenhum dos referidos, nada acontece.

reageEvento _ Estado = Estado

reageTempo :: Float -> Estado -> Estado

A função reageTempo reage à passagem do Tempo.

Esta função faz com que o objeto jogável se mova automaticamente dentro da janela, e verifica se o mesmo colide com algum objeto de Circulos (recorrendo à função auxilair colide).

colide :: Estado -> Estado

A função colide é uma função auxiliar que verifica se o objeto jogável colide com alguma das figuras circulares definidas numa lista Circulos e a remove se for verdade.

Exemplos de utilização:

Expand

>>> desenhaCirculos ((0,35), v, [((0, 0), 15),((130, 80), 15)], t)
((0,35), v, [((130, 80), 15)], t)

fr :: Int

É uma função que define a frame rate, ou seja, a quantidade de quadros, que serão desenhados por segundo.

dm :: Display

É uma função que desenha a Janela de Jogo e, portanto, o seu nome, tamanho e posição no ecrã.

corFundo :: Color

É uma função constante que define a cor do fundo de dm.

main :: IO ()

A função main é responsável por iniciar o programa e executar as ações principais.

Esta função configura a janela, lida com entrada/saída de dados, e chama outras funções para realizar as tarefas do programa.