Tarefa1

Documentation

generateRandoms

Arguments

:: Int
Número de inteiros a ser gerado.

-> Int
seed de números aleatórios.

-> [Int]
Lista dos números aleatórios.

Dada uma seed, retorna a lista dos n inteiros gerados aleatoriamente.

subList

Arguments

:: Int
Tamanho n das sublistas.

-> [a]
Lista que vai ser dividida.

-> [[a]]
Lista de listas de tamanho n resultante da divisão da lista pedida.

Converte uma lista numa lista de listas de tamanho n.

convertPiece

Arguments

:: Int
Inteiro entre 0 e 99 que vai ser convertido.

-> Piece
Peça resultante da conversão.

Converte um número inteiro numa Piece.

printCorridor

Arguments

:: Corridor
Conjunto de peças.

-> String
String em que está representado simbolicamente o corredor.

Converte um corredor numa String.

mazeToString

Arguments

:: Maze
Conjunto de corredores.

-> String
String em que está representado simbolicamente o labirinto.

Converte um labirinto numa String.

convertCorridor

Arguments

:: [Int]
Lista de inteiros que irão ser convertidos em peças.

-> Corridor
Lista de peças convertidas.

Converte uma lista de inteiros num corredor.

convertMaze

Arguments

:: [[Int]]
Lista de listas de inteiros que irão ser convertidas em corredores.

-> Maze
Lista de corredores convertidos.

Converte uma lista de inteiros num labirinto.

justWalls

Arguments

:: Int
Tamanho da lista pretendido.

-> Corridor
Lista com tamanho n só com paredes.

Pede um número e constrói uma lista de paredes com esse tamanho.

wallsUpDown

Arguments

:: Maze
Labirinto que vai sofrer as alterações.

-> Maze
Labirinto com paredes em cima e em baixo.

Substitui os limites superior e inferior do labirinto por paredes.

sideWalls

Arguments

:: Maze
Labirinto que vai sofrer as alterações.

-> Maze
Labirinto com os limites laterais substituídos por paredes.

Substitui os limites laterais do labirinto por paredes.

closeMaze

Arguments

:: Maze
Labirinto que vai sofrer as alterações.

-> Maze
Labirinto cercado por paredes.

Substitui todos os limites do labirinto por paredes, utilizando as funções auxiliares anteriores.

getMiddle :: Maze -> Int

Dá o número da posição do corredor central do labirinto.

replaceInCorridorAB

Arguments

:: Maze
Labirinto que vai sofrer as alterações.

-> Int
Indice do corredor que vai ser alterado.

-> Piece
Peça pela qual queremos substituir.

-> Maze
Labirinto com o primeiro e último elemento de um corredor substituído por uma determinada peça.

Substitui o primeiro e último elemento de um corredor por uma peça.

replaceInList

Arguments

:: Int
Indice no corredor que quer subsituir.

-> Piece
Peça pela qual quer substituir.

-> Corridor
Corredor que vai sofrer as alterações.

-> Corridor
Corredor com os indices 0 i substituídos por uma determinada peça.

Substitui num corredor o Indice 0 i (i :: Int), por uma determinada peça.

makeTunnel

Arguments

:: Maze
Labirinto em que vai ser construído o túnel.

-> Maze
Labirinto com o túnel implementado.

Faz o túnel no labirinto tendo em conta a sua altura, utilizando as funções auxiliares anteriores.

mazeTunel

Arguments

:: Maze
Labirinto que vai ser fechado por paredes, sendo de seguida implementado o túnel.

-> Maze
Labirinto rodeado por paredes mas com o túnel.

Função conjunta que é aplicada no generateMaze, que em primeiro fecha o labirinto e de seguida constrói o túnel.

midCorridor :: Corridor -> Int

Função que nos dá o número da posicão da peça central dum corredor.

constroiCasa

Arguments

:: Maze
Labirinto cujas dimensões vão ser analisadas.

-> Maze
Constituição parte central que contém a casa dos fantasmas, relativamente ao labirinto pedido.

Constituição da parte central do labirinto que irá conter a casa dos fantasmas, tendo em conta as dimensões do labirinto.

ghostHouse

Arguments

:: Maze
Labirinto no qual irá ser implementada a casa dos fantasmas.

-> Maze
Labirinto com a casa dos fantasmas construída.

Constrói a casa dos fantasmas no labirinto, tendo em conta a paridade da sua altura.

validM

Arguments

:: Int
Largura do labirinto.

-> Int
Altura do labirinto.

-> Bool
Bool que nos diz se o labirinto cumpre os requisitos mínimos das suas dimensões.

Função auxiliar para avaliar a validade do labirinto quanto às suas dimensões.

generateMaze

Arguments

:: Int
Largura do labirinto.

-> Int
Altura do labirinto.

-> Int
seed do labirinto.

-> Maze
Labirinto válido com todos os requisitos estruturais e com as dimensões mínimas de 15x10.

Pede as dimensões do labirinto e a sua seed e faz um labirinto válido, dando erro caso as suas dimensões não cumpram os requisitos mínimos.

imprimeMaze :: Maze -> IO ()

Mostra cada peça do labirinto nos seus respetivos símbolos atribuidos para uma visualização mais gráfica do labirinto.

genTest :: Int -> Int -> Int -> IO ()

Dado a sua largura, altura e uma seed, a função mostra o labirinto com essas características, de uma maneira mais gráfica para uma melhor visualização.

Funções-teste

testCasesGenerateMaze :: [(Int, Int, Int)]

Casos de argumentos para a função generateMaze.

testItT1 :: [(Int, Int, Int)] -> [Maze]

Testa os casos para a função generateMaze.

validMeasures :: Maze -> Bool

Verifica se a largura e comprimento do labirinto gerado cumprem os requisitos mínimos.

checkCorr :: Corridor -> Bool

Função auxiliar que verifica se um corredor tem só peças Wall.

checkWallAB :: Maze -> Bool

Verifica se o primeiro e o último corredor estão com uma lista de Wall.