Tarefa1_2018li1g101

Description

Este módulo define funções comuns da Tarefa 1 do trabalho prático.

Testes

testesT1 :: [Instrucoes]

Testes unitários da Tarefa 1.

Cada teste é uma sequência de Instrucoes.

Funções principais da Tarefa 1.

instrucao :: Instrucao -> Editor -> Editor

Aplica uma Instrucao num Editor.

Move - move numa dada Direcao.

MudaTetromino - seleciona a Peca seguinte (usar a ordem léxica na estrutura de dados), sem alterar os outros parâmetros.

MudaParede - muda o tipo de Parede.

Desenha - altera o Mapa para incluir o Tetromino atual, sem alterar os outros parâmetros. ^ A Instrucao a aplicar. ^ O Editor anterior. ^ O Editor resultante após aplicar a Instrucao.

desenhaTetromino :: Editor -> Editor

Desenha tetromino

atualizaMapa :: Matriz Bool -> (Parede, Int) -> Posicao -> Mapa -> Mapa

funcao "atualizaMapa" : recebe a matriz bool a implemententar, o tipo de parede, um contador para as linhas, a posiçao do editor e o mapa a modificar

atualizaLista2 :: [Bool] -> (Parede, Int) -> Int -> [Peca] -> [Peca]

atualizaLista2 recebe uma lista de bools a implementar , o tipo de parede , um contador para as colunas, a posicao na linha que deve ser implementado e a lista de pecas a modificar

rodaMatrizTetromino :: Matriz Bool -> Direcao -> Matriz Bool

Roda a Matriz Bool de acordo com a direcao

trocaDirecao :: Direcao -> Direcao

Troca a direção

trocaTetromino :: Tetromino -> Tetromino

Troca o Tetronimo

trocaParede :: Parede -> Parede

Troca o tipo de Parede

instrucoes :: Instrucoes -> Editor -> Editor

Aplica uma sequência de Instrucoes num Editor.

NB: Deve chamar a função instrucao. ^ As Instrucoes a aplicar ^ O Editor anterior. ^ O Editor resultante após aplicar as Instrucoes.

mapaInicial :: Dimensao -> Mapa

Cria um Mapa inicial com Paredes nas bordas e o resto vazio. ^ O Mapa resultante com a Dimensao dada ^ A Dimensao do Mapa a criar.

bordaIndestrutivel :: Matriz Peca -> Int -> Posicao -> Mapa

"bordaIndestrutivel" aualiza a matriz vazia com blocos indestrutiveis nas bordas

listaIndestrutivel :: [Peca] -> [Peca]

"listaIndestrutivel" recebe uma lista de peças e retorna uma lista de blocos indestrutiveis

editorInicial :: Instrucoes -> Editor

Cria um Editor inicial.

NB: Deve chamar as funções mapaInicial, dimensaoInicial, e posicaoInicial. ^ Uma sequência de Instrucoes de forma a poder calcular a dimensaoInicial e a posicaoInicial. ^ O Editor inicial, usando a Peca I Indestrutivel voltada para C.

constroi :: Instrucoes -> Mapa

Constrói um Mapa dada uma sequência de Instrucoes.

NB: Deve chamar as funções Instrucoes e editorInicial. ^ Uma sequência de Instrucoes dadas a um Editor para construir um Mapa. ^ O Mapa resultante.