Tarefa1_2018li1g127

Description

Este módulo define funções comuns da Tarefa 1 do trabalho prático.

Testes

testesT1 :: [Instrucoes]

Testes unitários da Tarefa 1.

Cada teste é uma sequência de Instrucoes.

Funções principais da Tarefa 1.

instrucao

Arguments

:: Instrucao
A Instrucao a aplicar.

-> Editor
O Editor anterior.

-> Editor
O Editor resultante após aplicar a Instrucao.

Aplica uma Instrucao num Editor.

Move - move numa dada Direcao.

MudaTetromino - seleciona a Peca seguinte (usar a ordem léxica na estrutura de dados), sem alterar os outros parâmetros.

MudaParede - muda o tipo de Parede.

Desenha - altera o Mapa para incluir o Tetromino atual, sem alterar os outros parâmetros.

auxdesenha :: Posicao -> Posicao -> Mapa -> Mapa -> Mapa

recebe uma auxiliar de posição (0,0) , a posição a iniciar o desenho (a,b), o mapa do tetromino e o mapa a substituir | se a linha coincidir com a auxiliar, então manda para uma auxiliar para encontrar a coluna e começar a desenhar | se a linha não coincidir com a auxiliar, escreve a linha do mapa, e repete o processo para as restantes linhas até achar a linha a substituir

auxdesenhalinha :: Int -> [Peca] -> [Peca] -> [Peca]

Recebe a quantidade de peças que vai ter de substituir, uma linha do tetromino, e a linha do mapa a substituir | Se a peça do tetromino a substituir for vazia, não substitui e passa à peça seguinte | Se a peça do tetromino a substituir não for vazia, então substitui no mapa, independentemente do tipo de peça que está no mapa

desenhaMapa :: Matriz Bool -> Parede -> Mapa

Dada a matriz de Bools do tetromino e o tipo de parede, transforma a Matriz de Bool num Mapa

desenhaLinha :: [Bool] -> Parede -> [Peca]

recebe uma linha de bool, o tipo de parede, e transforma a linha numa linha de Peças

rodaT :: Direcao -> Matriz Bool -> Matriz Bool

Recebe uma direção, e a Matriz de Bools do tetromino | Se for cima, dá a matriz; Se for Direita, roda uma vez; Se for Baixo, roda 2 vezes; Se for Esquerda, roda 1 vez para a esquerda.

instrucoes

Arguments

:: Instrucoes
As Instrucoes a aplicar.

-> Editor
O Editor anterior.

-> Editor
O Editor resultante após aplicar as Instrucoes.

Aplica uma sequência de Instrucoes num Editor.

NB: Deve chamar a função instrucao.

mapaInicial

Arguments

:: Dimensao
A Dimensao do Mapa a criar.

-> Mapa
O Mapa resultante com a Dimensao dada.

Cria um Mapa inicial com Paredes nas bordas e o resto vazio.

auxauxMapaI :: Int -> Posicao -> Mapa

recebe uma variável auxiliar e a dimensão do mapa (a,b) | se for só 1 linha, então repete b vezes a peça Bloco Indestrutivel | se a variável auxiliar for igual ao número de linhas, então faz uma linha só de Blocos Indestrutiveis | caso contrário, escreve um Bloco Indestrutivel, seguido de (b-2) blocos vazios e mais um bloco indestrutivel, acabando assim a linha e repetindo para o resto das linhas

editorInicial

Arguments

:: Instrucoes
Uma sequência de Instrucoes de forma a poder calcular a dimensaoInicial e a posicaoInicial.

-> Editor
O Editor inicial, usando a Peca I Indestrutivel voltada para C.

recebe uma auxiliar de posição auxmapaI :: Posicao -> Dimensao -> [Peca] auxmapaI (c,d) (a,b) | d > b = [] | c == 1 || d == 1 || c == a || d == b = Bloco Indestrutivel : auxmapaI (c,d+1) (a,b) | otherwise = Vazia : auxmapaI (c,d+1) (a,b)

Cria um Editor inicial.

NB: Deve chamar as funções mapaInicial, dimensaoInicial, e posicaoInicial.

constroi

Arguments

:: Instrucoes
Uma sequência de Instrucoes dadas a um Editor para construir um Mapa.

-> Mapa
O Mapa resultante.

Constrói um Mapa dada uma sequência de Instrucoes.

NB: Deve chamar as funções Instrucoes e editorInicial.

constroiM :: Editor -> Mapa

recebe um editor e devolve o mapa