Tarefa1_2021li1g056

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2021/22.

Documentation

validaPotencialMapa :: [(Peca, Coordenadas)] -> Bool

Esta tarefa consiste em validar um potencial mapa. Seguindo os seguintes critérios:

Todas as coordenadas devem ser diferentes entre si.

Só deve existir exatamente uma porta.

Nenhuma caixa deve estar a "flutuar".

Existir no mínimo um espaço vazio.

Deve existir um chão ao longo do mapa.

No 1º tópico foi utilizada a função diferentesCoordenadas para verificar se todas são diferentes.

No 2º tópico foi utilizada a função soUmaPorta para verificar se apenas existe uma porta. A função contaPortas foi utilizada como auxiliar para determinar o número total de portas.

No 3º tópico foi utilizada a função posicaoCaixas para verificar se todas as caixas estão bem posicionadas. A função procuraCaixas constrói uma lista apenas de caixas com as suas coordenadas. A função deBaixoDaCaixa verifica se existe um bloco ou uma caixa de baixo de uma caixa.

No 4º tópico foi utilizada a função haVazios para verificar se o nº total de peças era diferente de o nº total da matriz. A função maiorXeY serve para tirar as dimensões da matriz. A função conta serve para enumerar o nº total de peças.

No 5º tópico foi utilizada a função validaChao para verificar a partir do bloco mais à esquerda se existe uma ligação de blocos até ao bloco mais à direita. A função procuraBlocos constói uma lista só de blocos. A função blocoMaisBaixoEsq determina o bloco mais em baixo e mais à esquerda de todos. A função blocoMaisBaixoDir determina o bloco mais em baixo e mais à direita de todos. A função haSeguinte verifica se existe um bloco com uma determinada coordenada.

diferentesCoordenadas

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Bool
verifica se as peças são todas diferentes

função que verifica se todas as peças têm diferentes coordenadas

soUmaPorta

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Bool
verifica se existe apenas uma porta

função que verifica se apenas existe uma porta

contaPortas

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Int
retorna o nº total de portas

função que conta o nº total de portas

posicaoCaixas

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Bool
verifica se as caixas todas estão bem posicionadas

função que verifica o posicionamento das caixas

procuraCaixas

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> [(Peca, Coordenadas)]
constrói uma lista só de caixas e as suas coordenadas

função que constroi uma lista de caixas

deBaixoDaCaixa

Arguments

:: (Peca, Coordenadas)
recebe uma caixa e as suas coordenadas

-> [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Bool
verifica se existe algo de baixo da caixa

função que verifica se existe algo de baixo de uma caixa

haVazios

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Bool
verifica se existe pelo menos um espaço vazio

função que verifica se existe pelo menos um vazio

maiorXeY

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Coordenadas
determina o maior X e o maior Y

função que determina o maior X e Y

conta

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Int
determina o nº total de peças

função que determina o nº total de peças

validaChao

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Bool
verifica se o chão está corretamente construído

função que verifica se o chão está corretamente construído

procuraBlocos

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> [Coordenadas]
constrói uma lista só de coordenadas de blocos

função que constrói uma lista apenas com as coordenadas dos blocos

blocoMaisBaixoEsq

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Coordenadas
determina as coordenadas do bloco mais em baixo e à esquerda

função que determina as coordenadas do bloco mais em baixo e à esquerda

blocoMaisBaixoDir

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Coordenadas
determina as coordenadas do bloco mais em baixo e à direita

função que determina as coordenadas do bloco mais em baixo e à direita

haSeguinte

Arguments

:: Coordenadas
recebe um par de coordenadas

-> [Coordenadas]
recebe uma lista com as coordenadas de todos os blocos

-> Bool
verifica se existe um bloco com uma determinada coordenada

função que verifica se existe um bloco com uma determinada coordenada