Tarefa1_2021li1g085

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2021/22.

Documentation

validaPotencialMapa

Arguments

:: [(Peca, Coordenadas)]
argumento

-> Bool
validade do Mapa

A função validaPotencialMapa junta funções secundárias de modo a poder validar o mapa.

Esta é definida da seguinte forma:

validaPotencialMapa :: [(Peca, Coordenadas)] -> Bool
validaPotencialMapa [] = False
validaPotencialMapa a = vPecaPorPosicao x xs && unicaPorta a 0 && nflutua a && existeEspVazio a && blocoPisoChao a
    where
     (x:xs) = pecaCoorEmCoor a

Exemplos de utilização:

>>> validaPotenciaMapa m1
True

pecaCoorEmCoor

Arguments

:: [(Peca, Coordenadas)]
lista de Peca e Coordenadas

-> [Coordenadas]
lista de Coordenadas

A função pecaCoorEmCoor transforma o argumento para uma lista de Coordenadas, de modo a poderem ser comparadas.

pecaCoorEmCoor :: [(Peca, Coordenadas)] -> [Coordenadas]
pecaCoorEmCoor [] = []
pecaCoorEmCoor ((_,(b,c)):t) = (b,c) : pecaCoorEmCoor t

Exemplos de utilização:

>>> pecaCoorEmCoor [(Bloco,(0,2)),(Caixa,(3,3)),(Porta,(0,2))]
[(0,2),(3,3),(0,2)]

vPecaPorPosicao :: Coordenadas -> [Coordenadas] -> Bool

A função vPecaPorPosicao recorre à função pecaCoorEmCoor para verificar se existem duas peças a ocupar a mesma posição.

vPecaPorPosicao :: Coordenadas -> [Coordenadas] -> Bool 
vPecaPorPosicao _ [] = True
vPecaPorPosicao a (b:t) | elem a (b:t) = False 
                        | otherwise = vPecaPorPosicao b t

unicaPorta :: [(Peca, Coordenadas)] -> Int -> Bool

A função unicaPorta soma 1 ao acumulador sempre que aparece Porta na lista.

unicaPorta :: [(Peca,Coordenadas)] -> Int -> Bool 
unicaPorta [] ac | ac == 0 = False
                 | otherwise = True
unicaPorta ((a,(b,c)):t) ac | ac > 1 = False
                            | a == Porta = unicaPorta t (ac+1)
                            | otherwise  = unicaPorta t ac

Exemplos de utilização:

>>> unicaPorta [(Porta,(0,0)),(Porta,(1,0))]
False

>>> unicaPorta m1
True

nflutua

Arguments

:: [(Peca, Coordenadas)]
lista

-> Bool
inexistência de Caixa a flutuar

A função nflutua certifica de que não há nenhuma Caixa a flutuar.

- | Recorre-se a procuraCaixa

nflutua :: [(Peca,Coordenadas)] -> Bool 
nflutua [] = True 
nflutua l = procuraCaixa l l

procuraCaixa

Arguments

:: [(Peca, Coordenadas)]
lista

-> [(Peca, Coordenadas)]
lista inicial completa

-> Bool
inexistência de Caixa a flutuar

- * Quando se encontra uma Caixa na lista, procura-se a base da mesma.

procuraCaixa :: [(Peca,Coordenadas)]-> [(Peca,Coordenadas)] -> Bool 
procuraCaixa [] _ = True
procuraCaixa ((a,b):t) l | a == Caixa = procuraBase (a,b) l && procuraCaixa t l
                         | otherwise  = procuraCaixa t l

procuraBase

Arguments

:: (Peca, Coordenadas)
Coordenadas de Caixa

-> [(Peca, Coordenadas)]
lista original com todos os elementos

-> Bool
inexistência de Caixa a flutuar

- * Verifica a existencia de uma Caixa ou Bloco numa coordenada a baixo do eixo y.

procuraBase :: (Peca,Coordenadas) -> [(Peca,Coordenadas)] -> Bool 
procuraBase _ [] = False
procuraBase (n,(x,y)) ((p,(a,b)):ts) | x == a && (y+1) == b && p == Bloco = True
                                     | x == a && (y+1) == b && p == Caixa = True
                                     | otherwise = procuraBase (n,(x,y)) ts

dimensaoMatriz

Arguments

:: [(Peca, Coordenadas)]
argumento

-> Coordenadas
dimensão da Matriz

A função dimensaoMatriz busca as maiores coordenadas da lista, e soma se 1 no fim a cada eixo, de modo a compensar as coordenadas nulas.

dimensaoMatriz :: [(Peca,Coordenadas)] -> (Coordenadas)
dimensaoMatriz [] = (0,0)
dimensaoMatriz [(a,(c,d))] = (c+1,d+1)
dimensaoMatriz ((p1,(a1,b1)):(p2,(a2,b2)):t) | a1 >= a2 && b1 >= b2 = dimensaoMatriz ((p1,(a1,b1)):t)
                                             | b1 > b2   = dimensaoMatriz ((p1,(a2,b1)):t)
                                             | a1 > a2   = dimensaoMatriz ((p1,(a1,b2)):t)
                                             | otherwise = dimensaoMatriz ((p1,(a2,b2)):t)

espacoVazio

Arguments

:: [(Peca, Coordenadas)]
lista de elementos do Mapa

-> Bool
existencia de Peca que seja Vazio

Quando existe Peca que seja Vazio a função devolve True.

espacoVazio :: [(Peca,Coordenadas)] -> Bool
espacoVazio [] = False
espacoVazio ((a,(b,c)):t) | a == Vazio = True 
                          | otherwise = espacoVazio t

existeEspVazio

Arguments

:: [(Peca, Coordenadas)]
lista de elementos

-> Bool
existência de espaço vazio

Caso não exista Peca que seja Vazio, a função existeEspVazio recorre à comparação da dimensão da Matriz com o número de elementos que esta possui.

existeEspVazio :: [(Peca,Coordenadas)] -> Bool
existeEspVazio [] = False
existeEspVazio l | a == False = length l < (x*y)
                 | a == True  = True
   where 
       (x,y) = dimensaoMatriz l
       a     = espacoVazio l

soBlocos

Arguments

:: [(Peca, Coordenadas)]
elementos do Mapa

-> [Coordenadas]
coordenadas dos elementos

A função soBlocos gera uma lista so com as cooredenadas de blocos.

soBlocos :: [(Peca,Coordenadas)] -> [Coordenadas]
soBlocos [] = []
soBlocos ((a,(b,c)):t) | a == Bloco = (b,c): soBlocos t
                       | otherwise = soBlocos t

firstBloco

Arguments

:: [Coordenadas]
lista só de Bloco

-> Coordenadas
assume-se a existência de Bloco na coluna 0

-> Coordenadas
bloco com maior y (situado mais a baixo)

A função firstBloco procura pelo Bloco mais a baixo e mais a Oeste da lista.

firstBloco :: [Coordenadas] -> Coordenadas -> Coordenadas
firstBloco [] ac = ac
firstBloco ((b,c):t) (x,y) | b == 0 && c > y = firstBloco t (b,c)
                           | otherwise = firstBloco t (x,y)

quickBlocos

Arguments

:: [Coordenadas]
Coordenadas de Bloco

-> [Coordenadas]
Coordenadas ordenadas

Esta função ordena as coordenadas de baixo para cima, de esquerda para a direira, utilizando o algoritmo 'quick sort'.

quickBlocos :: [Coordenadas] -> [Coordenadas]
quickBlocos [] = []
quickBlocos (x:xs) = (quickBlocos a1) ++ [x] ++ (quickBlocos a2)

    where
        (a1,a2) = parteBlocos x xs

parteBlocos

Arguments

:: Coordenadas
Coordenada que é utilizada para comparar

-> [Coordenadas]
lista que vai ser ordenada

-> ([Coordenadas], [Coordenadas])
listas de Coordenadas maiores e menores à de comparação

A função divide os elementos da lista em duas listas recorrendo a Tupling:

- * elementos mais a Baixo e a Esquerda

- * elementos mais a Cima e a Direita

parteBlocos :: Coordenadas -> [Coordenadas] -> ([Coordenadas],[Coordenadas])
parteBlocos _ [] = ([],[])
parteBlocos (b,c) ((y,z):t) | b > y || (b == y && c < z) = ((y,z):as,bs)
                            | otherwise = (as,(y,z):bs)
    where
        (as,bs) = parteBlocos (b,c) t

blocoPiso

Arguments

:: [Coordenadas]
lista de Coordenadas de Bloco

-> Coordenadas
recebe o Bloco mais a Este e Sul

-> Coordenadas
recebe a dimensão da Matriz

-> Bool
veracidade da existência de um chão contínuo

A função blocoPiso recebe as coordenadas ordenadas e utiliza um acumulador para comparar a corrente coordenada com a anterior.

blocoPiso :: [Coordenadas] -> Coordenadas -> Coordenadas -> Bool
blocoPiso [] _ _ = False
blocoPiso ((h,z):ts) (a,b) (x,y) | a == x = True
                                 | h < a      = blocoPiso ts (a,b) (x,y)
                                 | z == (b+1) = blocoPiso ts (h,z) (x,y)
                                 | z == b     = blocoPiso ts (h,z) (x,y)
                                 | z == (b-1) = blocoPiso ts (h,z) (x,y)

blocoPisoChao

Arguments

:: [(Peca, Coordenadas)]
lista dos elementos do Mapa

-> Bool
veracidade da existência de um chão contínuo

A função blocoPisoChao é utilizada para unir as funções definidas anteriormente.

blocoPisoChao :: [(Peca,Coordenadas)] -> Bool 
blocoPisoChao l = blocoPiso td (a,b) (x-1,y-1) 
    where
        (a,b) = firstBloco coor (0,0)
        (x,y) = dimensaoMatriz l
        coor  = soBlocos l
        td    = quickBlocos coor