Tarefa1_2022li1g046

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

validarObstaculos :: Mapa -> Bool

A função validarObstaculos avalia se certos obstaculos são validos no terreno escolhido

validarObstaculoRelva :: Mapa -> Bool

A função validarObstaculoRelva avalia se certos obstaculos são validos para o terreno Relva

validarObstaculoRio :: Mapa -> Bool

A função validarObstaculoRelva avalia se certos obstaculos são validos para o terreno Rio

validarObstaculoEstrada :: Mapa -> Bool

A função validarObstaculoRelva avalia se certos obstaculos são validos para o terreno Estrada

contadorDeObstaculos :: [Obstaculo] -> Int

Esta função conta o número de obstáculos que existe numa lista.

contadorDeObstaculosF :: Mapa -> Bool

Esta função faz a comparação final se existe tantos obstaculos como o número da largura do mapa.

existirNenhum :: [Obstaculo] -> Bool

A função existirNenhum confirma se existe pelo menos um obstáculo Nenhum na lista de Obstaculos.

existirNenhumL :: [[Obstaculo]] -> Bool

A função existirNenhumL confirma se existe pelo menos um obstáculo Nenhum numa lista de listas de Obstaculos.

existirNenhumF :: Mapa -> Bool

A função existirNenhumF confirma se num Mapa existe pelo menos um obstáculo Nenhum.

validarRio :: [Terreno] -> Bool

A função validarRio recebe uma lista de Terrenos, mais especificamente de Rios, e verifica se cada Rio tem um sentido diferente

separarPares :: [(Terreno, [Obstaculo])] -> [Terreno]

A função separarPares será uma função auxiliar à função validarRio para criar compatibilidade de dados, neste caso separar a parte da lista do Terreno

separarPares2 :: [(Terreno, [Obstaculo])] -> [[Obstaculo]]

A função separarPares2 será uma função auxiliar para criar compatibilidade de dados, neste caso separar a parte da lista de lista de Obstaculos.

riosContrarios :: Mapa -> Bool

A função riosContrarios verifica se num mapa rios consecutivos têm sentidos opostos

carros :: [Obstaculo] -> [Int]

A função carros recebe uma lista de Obstaculos e devolve uma lista de interios em que 1 corresponde aos Carros e o resto dos Obstaculos corresponde a 0.

exemplo :: [[Obstaculo]] -> [[Int]]

A função exemplo recebe uma lista de listas de Obstaculos e devolde uma lista de listas baseado na funcionalidade da função carros.

primenta :: [[Int]] -> [[[Int]]]

A função primenta recebe uma lista de listas de inteiros e devolve uma lista de listas de listas de inteiros que estão dividos entre 1's e 0's

somaDeListas :: [[[Int]]] -> [[Int]]

A função somaDeListas recebe uma lista de listas de listas e devolve uma lista de listas com as somas feitas entre os elementos das listas.

quaseFinal :: [[Int]] -> [Int]

A função quaseFinal recebe uma lista de listas e devolve uma lista de inteiros com o maior número de cada lista.

maiorNumero :: [Int] -> Int

A função maiorNumero recebe uma lista de números e devolve o maior

group :: [Int] -> [[Int]]

A função group junta elementos iguais e consecutivos numa lista.

soma2 :: [[Int]] -> [Int]

A função soma2 recebe uma lista de listas e soma os elementos dentro da lista

validaTamanhoCarro :: Mapa -> Bool

A função validaTamanhoCarro valida se o tamanho do Carro pode ou não existir

troncos :: [Obstaculo] -> [Int]

A função carros recebe uma lista de Obstaculos e devolve uma lista de 1's e 0's em que 1 corresponde aos Troncos e 0 corresponde aos outros Obstaculos.

exemplo2 :: [[Obstaculo]] -> [[Int]]

A função exemplo recebe uma lista de listas de Obstaculos e devolde uma lista de listas baseado na funcionalidade da função troncos.

validaTamanhoTronco :: Mapa -> Bool

A função validaTamanhoCarro valida se o tamanho do Tronco pode ou não existir

numeroDeRios :: [Terreno] -> [Int]

A função numeroDeRios recebe uma lista de Terrenos e devolve uma lista de interios em que 1 corresponde aos Rios e o resto dos Terrenos corresponde a 0.

validaTamanhoRio :: Mapa -> Bool

A função validaTamanhoRio valida se o número de Rios consecutivos pode existir

numeroDeRelvas :: [Terreno] -> [Int]

A função numeroDeRios recebe uma lista de Terrenos e devolve uma lista de interios em que 1 corresponde às Relva e o resto dos Terrenos corresponde a 0.

validaTamanhoRelva :: Mapa -> Bool

A função validaTamanhoRelva valida se o número de Relvas consecutivos pode existir

numerodeEstrada :: [Terreno] -> [Int]

A função numeroDeRios recebe uma lista de Terrenos e devolve uma lista de interios em que 1 corresponde às Estradas e o resto dos Terrenos corresponde a 0.

validaTamanhoEstrada :: Mapa -> Bool

A função validaTamanhoEstrada valida se o número de Estradas consecutivos pode existir

mapaValido :: Mapa -> Bool

A função mapaValido valida se o Mapa recebido é ou não jogável.