Tarefa1_2022li1g061

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

FUNÇÃO QUE DETERMINA SE O MAPA É VÁLIDO Esta função recebe um Mapa e retorna um Bool Se todas as funções auxiliares confirmarem válidos os seus parâmetros, a função retornará True Caso contrário retornará False

validMapa :: Mapa -> Bool

FUNÇÃO QUE VERIFICA SE O TERRENO E OS OBSTÁCULOS COMBINAM Esta função recebe um Mapa e retorna um Bool O resultado será determinado por recurso à função auxiliar "validMap"

validMap :: Terreno -> [Obstaculo] -> Bool

FUNÇÃO QUE INVOCA AS AUXILIARES NECESSÁRIAS TENDO EM CONTA O TERRENO Esta função recebe um Terreno e uma lista de Obstáculos e retorna um Bool Se a lista for vazia, significa que chegou ao fim e retorna True Se o terreno for relva, invoca a função "validRelva" Se o terreno for rio, invoca a função "validRio" Se o terreno for estrada, invoca a função "validEstrada"

validRelva :: [Obstaculo] -> Bool

FUNÇÃO QUE VERIFICA SE OS OBSTÁCULOS SÃO VÁLIDOS PARA O TERRENO RELVA Esta função recebe uma lista de obstáculos e retorna um Bool Se a lista for vazia, significa que chegou ao fim e retorna True Se o obstáculo actual for Nenhum ou Árvore a função avança Caso Contrário retorna False

validRio :: [Obstaculo] -> Bool

FUNÇÃO QUE VERIFICA SE OS OBSTÁCULOS SÃO VÁLIDOS PARA O TERRENO RIO Esta função recebe uma lista de obstáculos e retorna um Bool Se a lista for vazia, significa que chegou ao fim e retorna True Se o obstáculo actual for Nenhum ou Tronco a função avança Caso Contrário retorna False

validEstrada :: [Obstaculo] -> Bool

FUNÇÃO QUE VERIFICA SE OS OBSTÁCULOS SÃO VÁLIDOS PARA O TERRENO ESTRADA Esta função recebe uma lista de obstáculos e retorna um Bool Se a lista for vazia, significa que chegou ao fim e retorna True Se o obstáculo actual for Nenhum ou Carro a função avança Caso Contrário retorna False

checkRio :: Mapa -> Bool

FUNÇÃO QUE VERIFICA SE EXISTEM RIOS CONTÍGUOS COM A MESMA DIRECÇÃO Esta função recebe um Mapa e retorna um Bool Se a lista for vazia, significa que chegou ao fim e retorna True Se ambos os terrenos contíguos são Rio e as suas direcções são opostas, retorna True, caso esta última condição não se verifique retorna False Se os terrenos contíguos não forem ambos Rio, a função avança para o resto do mapa

checkVelocidade :: Terreno -> Terreno -> Bool

FUNÇÃO QUE VERIFICA SE 2 RIOS TÊM DIRECÇÕES OPOSTAS Esta função recebe dois terrenos e retorna um Bool Se a velocidade do primeiro for positiva e a do segundo também, retorna Falso, caso a do segundo seja negativa retorna True Se a velocidade do primeiro for negativa e a do segundo também, retorna Falso, caso a do segundo seja positiva retorna True

isRio :: Terreno -> Bool

FUNÇÃO QUE VERIFICA SE UM DADO TERRENO É UM RIO Esta função recebe um terreno e retorna um Bool Se o terreno for Rio retorna True Caso contrário retorna False

validTronco :: [Obstaculo] -> Int -> Bool

FUNÇÃO QUE VERIFICA SE UM TRONCO TEM O COMPRIMENTO VÁLIDO Esta função recebe uma lista de obstáculos e utiliza um contador recursivo Se a lista for vazia, significa que chegou ao fim e retorna True Se o contador for menor ou igual que 5 e o obstáculo for um tronco, a função avança e o contador é incrementado caso a cabeça seja um tronco Caso contrário retorna False

troncosValidos :: [Obstaculo] -> Int -> Bool

FUNÇAO QUE VERIFICA SE EXISTEM TRONCOS SEGUIDOS EM LADOS OPOSTOS DA LISTA Esta funçao recebe a lista de obstaculos e verifica se existemtroncos seguidos em lados opostos da lista, coisa que a validTronco nao faz Lista vazia vai retornar True pois é valido Contador menor que 4 continua a verificar pois pode ser valido Contador igual a 4 ele ja altera pois nao podem existir 6. Entao caso exista nos 2 lados o Rio é invalido.

validTail :: [Obstaculo] -> Int -> Bool

FUNÇAO QUE VERIFICA SE EXISTEM TRONCOS SEGUIDOS NA CAUDA Verifica se ha troncos seguidos na causa caso eles nao sejam iguais na funçao troncosValidos

checkTronco :: Mapa -> Bool

FUNÇÃO QUE VERIFICA SE OS TRONCOS DO MAPA SÃO VÁLIDOS Esta função recebe um Mapa e retorna um Bool Se a lista for vazia, significa que chegou ao fim e retorna True Se o terreno for Rio e o comprimento do tronco for válido a função avança, caso esta última condição não se verifique retorna False Caso contrário a função avança

existeTronco :: [Obstaculo] -> Bool

FUNÇAO QUE VERIFICA SE EXISTE TRONCO NO RIO Para o Rio ser ultrapassavel deve existir pelo menos um tronco nele. Esta função verifica se o elemento Tronco é elemento da lista de obstaculos no Rio.

isEstrada :: Terreno -> Bool

FUNÇÃO QUE VERIFICA SE UM DADO TERRENO É UMA ESTRADA Esta função recebe um terreno e retorna um Bool Se o terreno for de facto Estrada, retorna True Caso contrário retorna False

validCarro :: [Obstaculo] -> Int -> Bool

FUNÇÃO QUE VERIFICA SE UM CARRO TEM O COMPRIMENTO VÁLIDO Esta função recebe uma lista de obstáculos e utiliza um contador recursivo Se a lista for vazia, significa que chegou ao fim e retorna True Se o contador for menor ou igual que 3 e o obstáculo for um Carro, a função avança e o contador é incrementado caso a cabeça seja um Carro Caso contrário retorna False

carrosValidos :: [Obstaculo] -> Int -> Bool

FUNÇAO QUE VERIFICA SE EXISTEM CARROS SEGUIDOS EM LADOS OPOSTOS DA LISTA Esta funçao recebe a lista de obstaculos e verifica se existem carros seguidos em lados opostos da lista, coisa que a validCarro nao faz Lista vazia vai retornar True pois é valido Contador menor que 2 continua a verificar pois pode ser valido Contador igual a 2 ele ja altera pois nao podem existir 4. Entao caso exista nos 2 lados a Estrada é invalida.

validTailC :: [Obstaculo] -> Int -> Bool

FUNÇAO QUE VERIFICA SE EXISTEM CARROS SEGUIDOS NA CAUDA Verifica se ha carros seguidos na causa caso eles nao sejam iguais na funçao carrosValidos

checkCarro :: Mapa -> Bool

FUNÇÃO QUE VERIFICA SE OS CARROS DO MAPA SÃO VÁLIDOS Esta função recebe um Mapa e retorna um Bool Se a lista for vazia, significa que chegou ao fim e retorna True Se o terreno for Estrada e o comprimento do Carro for válido a função avança, caso esta última condição não se verifique retorna False Caso contrário a função avança

linhaValida :: Mapa -> Bool

FUNÇÃO QUE VERIFICA SE TODAS AS LINHAS DO MAPA POSSUEM O ELEMENTO NENHUM Esta função recebe um Mapa e retorna um Bool Se a lista for vazia, significa que chegou ao fim e retorna True Se a linha actual possuir o elemento Nenhum a função avança Caso contrário retorna False

validLine :: [Obstaculo] -> Bool

FUNÇÃO QUE VERIFICA SE NENHUM É UM ELEMENTO DE UMA DETERMINADA LINHA Esta função recebe uma lista de obstáculos e retorna um Bool Se Nenhum pertencer à linha retorna True Caso contrário retorna False

compLista :: Mapa -> Bool

FUNÇÃO QUE VERIFICA SE O COMPRIMENTO DA LISTA DE OBSTÁCULOS É IGUAL À LARGURA Esta função recebe um Mapa e retorna um Bool Se a lista for vazia significa que chegou ao fim e retorna True Se a largura for igual ao comprimento da lista, a função avança para o resto do Mapa Caso contrário retorna False

terrenosSeguidos :: Mapa -> Bool

FUNÇÃO QUE DETERMINA SE EXISTEM DEMASIADOS TERRENOS IGUAIS CONTÍGUOS Esta função recebe um Mapa e retorna um Bool Se a lista for vazia significa que chegou ao fim e retorna True Se o terreno for Rio, irá verificar se o número de Rios contíguos não excede o limite, caso não exceda a função avança para o resto do Mapa Se o terreno for Estrada, irá verificar se o número de Estradas contíguos não excede o limite, caso não exceda a função avança para o resto do Mapa Se o terreno for Relva, irá verificar se o número de Relvas contíguos não excede o limite, caso não exceda a função avança para o resto do Mapa

riosSeguidos :: [(Terreno, [Obstaculo])] -> Int -> Bool

FUNÇÃO QUE VERIFICA SE NÃO EXISTEM DEMASIADOS RIOS CONTÍGUOS Esta função recebe uma lista de tuplos terreno e lista de obstáculos, utiliza um contador recursivo e retorna um Bool Se a lista for vazia significa que chegou ao fim e retorna True Se o contador for inferior a 4 e os terrenos forem iguais, a função avança, caso esta última condição não se verique retorna True Caso contrário retorna False

estradasSeguidas :: [(Terreno, [Obstaculo])] -> Int -> Bool

FUNÇÃO QUE VERIFICA SE NÃO EXISTEM DEMASIADAS ESTRADAS CONTÍGUAS Esta função recebe uma lista de tuplos terreno e lista de obstáculos, utiliza um contador recursivo e retorna um Bool Se a lista for vazia significa que chegou ao fim e retorna True Se o contador for inferior a 5 e os terrenos forem iguais, a função avança, caso esta última condição não se verique retorna True Caso contrário retorna False

relvaSeguida :: [(Terreno, [Obstaculo])] -> Int -> Bool

FUNÇÃO QUE VERIFICA SE NÃO EXISTEM DEMASIADAS RELVAS CONTÍGUAS Esta função recebe uma lista de tuplos de Terreno e lista de Obstáculos, utiliza um contador recursivo e retorna um Bool Se a lista for vazia significa que chegou ao fim e retorna True Se o contador for inferior a 5 e os terrenos forem iguais, a função avança, caso esta última condição não se verique retorna True Caso contrário retorna False