Tarefa1_2022li1g115

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

A função "mapaValido" verifica se um mapa fornecido reúne todas as condições para que seja considerado válido

mapaValido m = (obstaculosValidos m) && (larguraV m) && (tamanhoV m) && (nenhumV m) && (contaCarros m) && (contaTroncos m) && (terrenoV m) && (rioPossivelV m)

obstaculosValidos :: Mapa -> Bool

A função "obstaculosValidos" verifica se os obstáculos de um dado mapa são validos para um dado terreno

obstaculosValidos (Mapa _ []) = True
obstaculosValidos (Mapa l ((tr,[]):t)) = obstaculosValidos (Mapa l t)
obstaculosValidos (Mapa l ((Relva,(o:os)):t)) | o == Arvore || o == Nenhum = obstaculosValidos (Mapa l ((Relva,os):t))
                                              | otherwise = False
obstaculosValidos (Mapa l ((Estrada v,(o:os)):t)) | o == Carro || o == Nenhum = obstaculosValidos (Mapa l ((Estrada v,os):t))
                                                  | otherwise = False
obstaculosValidos (Mapa l ((Rio v,(o:os)):t)) | o == Tronco || o == Nenhum = obstaculosValidos (Mapa l ((Rio v,os):t))
                                              | otherwise = False  

larguraV :: Mapa -> Bool

A função "larguraV" verifica se 2 rios contiguos tem direçao oposta

larguraV (Mapa l ((tr1,l1):(tr2,l2):t)) = larguraVaux (Mapa l ((tr1,l1):(tr2,l2):t)) 

larguraVaux :: Mapa -> Bool

A função "larguraVaux" é uma auxiliar da função "larguraV" que valida a direçao de 2 rios contiguos

larguraVaux (Mapa _ []) = True
larguraVaux (Mapa l [(tr,_)]) = True
larguraVaux (Mapa l ((Rio v1,l1):(Rio v2,l2):t)) = (v1*v2) < 0 && larguraVaux (Mapa l ((Rio v2,l2):t))    
larguraVaux (Mapa l ((tr1,l1):(tr2,l2):t)) = larguraVaux (Mapa l ((tr2,l2):t))

tamanhoV :: Mapa -> Bool

A função "tamanhoV" valida o tamanho de um terreno

tamanhoV (Mapa x []) = True
tamanhoV (Mapa x ((_,l):t)) = (x == (length l))  && tamanhoV (Mapa x t)                                           

nenhumV :: Mapa -> Bool

A função "nenhumV" valida se existe pelo menos um espaço livre em cada terreno

nenhumV (Mapa x []) = True
nenhumV (Mapa x ((_,l):t)) = elem Nenhum l && nenhumV (Mapa x t)                      

contaCarros :: Mapa -> Bool

A função "contaCarros" valida se o tamanho dos carros é valido

contaCarros (Mapa _ []) = True
contaCarros (Mapa x ((tr,[x1,x2,x3]):t)) = contaCarros (Mapa x t)
contaCarros (Mapa x ((tr,[x1]):t)) = contaCarros (Mapa x t)
contaCarros (Mapa x ((tr,[x1,x2]):t)) = contaCarros (Mapa x t)
contaCarros (Mapa x ((tr,(x1:x2:x3:x4:xs)):t)) | (x1 == x2) && (x2 == x3) && (x3 == x4) && (x4 == Carro) = False 
                                               | otherwise  = contaCarros (Mapa x ((tr,x2:x3:x4:xs):t))

contaTroncos :: Mapa -> Bool

A função "contaTroncos" valida o tamanho de um troco

contaTroncos (Mapa _ []) = True
contaTroncos (Mapa x ((tr,[x1]):t)) = contaTroncos (Mapa x t)
contaTroncos (Mapa x ((tr,[x1,x2]):t)) = contaTroncos (Mapa x t)
contaTroncos (Mapa x ((tr,[x1,x2,x3]):t)) = contaTroncos (Mapa x t)
contaTroncos (Mapa x ((tr,[x1,x2,x3,x4]):t)) = contaTroncos (Mapa x t)
contaTroncos (Mapa x ((tr,[x1,x2,x3,x4,x5]):t)) = contaTroncos (Mapa x t)
contaTroncos (Mapa x ((tr,(x1:x2:x3:x4:x5:x6:xs)):t)) 
    | (x1 == x2) && (x2 == x3) && (x3 == x4) && (x4 == x5) && (x5 == x6) && (x6 == Tronco) = False 
    | otherwise  = contaTroncos (Mapa x ((tr,x2:x3:x4:x5:x6:xs):t))

terrenoV :: Mapa -> Bool

A função "terrenoV" valida se os numeros de terrenos contiguos nao é excedido para os diferentes tipos de terreno

terrenoV (Mapa _ []) = True
terrenoV (Mapa l ((Rio v1,l1):(Rio v2,l2):(Rio v3,l3):(Rio v4,l4):(Rio v5,l5):t)) = False 
terrenoV (Mapa l ((Estrada v1,l1):(Estrada v2,l2):(Estrada v3,l3):(Estrada v4,l4):(Estrada v5,l5):(Estrada v6,l6):t)) = False
terrenoV (Mapa l ((Relva,l1):(Relva,l2):(Relva,l3):(Relva,l4):(Relva,l5):(Relva,l6):t)) = False
terrenoV (Mapa l (x:xs)) = terrenoV (Mapa l xs)

rioPossivelV :: Mapa -> Bool

A função "rioPossivelV" valida se existe pelo menos um tronco em cada rio

rioPossivelV (Mapa x []) = True
rioPossivelV (Mapa x ((Rio _,os):t)) | elem Tronco os && rioPossivelV (Mapa x t)  
rioPossivelV (Mapa x ((_,_):t)) = rioPossivelV (Mapa x t)