Tarefa2

Funções de movimento do Pacman

movePlayer

Arguments

:: Player
Um determinado Player.

-> Orientation
A orientação dada.

-> Player
Player com as devidas alterações na sua posição ou orientação.

Faz o Pacman se mover, consoante a orientação dada e a orientação do Pacman.

moveInTunnel

Arguments

:: Maze
Labirinto onde se encontra o Player.

-> Player
O Player selecionado.

-> Orientation
Orientação ordenada.

-> Player
O Player selecionado com a respetiva mudança na sua posição ou orientação.

Faz o Pacman passar pelo túnel, se estiver para entrar nele, caso contrário faz o seu movimento normal básico, através da função anterior.

thePlayer

Arguments

:: Play
Ordem Play dirigida a um determinado Player com uma certa ID.

-> [Player]
Lista de Player.

-> Player
Player cuja Play era dirigida, ou seja, o que tem o mesmo ID da Play.

Função auxiliar que devolve o Player ao qual é destinado o Play, ou seja, identifica o jogador com o ID da Play.

pieceInCoords

Arguments

:: Coords
Coordenadas do labirinto.

-> Maze
Respetivo labirinto.

-> Piece
Peça que se encontra nas coordenadas do respetivo labirinto.

Função auxiliar que dá a peça que se encontra nas coordenadas do labirinto pedidas.

testWall

Arguments

:: Play
Jogada.

-> State
Estado.

-> Bool
No caso de ter uma Wall devolve False, caso contrário devolve True.

Verifica se a posição para onde o Pacman vai, não tem uma peça Wall.

moveID

Arguments

:: Play
Jogada.

-> State
Estado.

-> State
Devolve um State em que é feita a alteração no estado do respetivo Player pretendida com o Play.

Dá uma lista de players com a alteração Play destinada ao Player com o respetivo ID.

movePac

Arguments

:: Play
Jogada.

-> State
Estado.

-> State
Estado em que o Player tem o seu estado alterado consoante a jogada, e que caso haja uma parede no caminho, ele não vá para ela.

Função final que se aplicam todas as leis de movimento do Pacman.

Funções de atualização do score do Pacman

nextPieceSF

Arguments

:: Play
Jogada.

-> State
Estado.

-> Bool
Devolve True se nas coordenadas do Player existir uma comida pequena, caso contrário devolve False.

Verifica se tem uma comida pequena nas coordenadas do Player com o ID do Play.

nextPieceBF

Arguments

:: Play
Jogada.

-> State
Estado.

-> Bool
Devolve True se nas coordenadas do Player existir uma comida grande, caso contrário devolve False.

Verifica se tem uma comida grande nas coordenadas do Player com o ID do Play.

nextPieceGh

Arguments

:: Play
Jogada.

-> State
Estado.

-> Bool
Devolve True se nas coordenadas do Player existir um fantasma morto, caso contrário devolve False.

Verifica se tem um fantasma morto nas coordenadas do Player com o ID do Play, através das coordenadas do Player e das coordenadas dos Ghosts Dead, verificando se a coordenada do Player coincide com a coordenada de algum Ghost, com a função "elem" pré-definida do Prelude.

score1ID

Arguments

:: Play
Jogada com um determinado ID.

-> [Player]
Conjunto de jogadores.

-> [Player]
Conjunto de jogadores em que o jogador com o ID da Play tem +1 ponto.

Aumenta o score do Player com o ID do Play em 1 ponto.

score5ID

Arguments

:: Play
Jogada com um determinado ID.

-> [Player]
Conjunto de jogadores.

-> [Player]
Conjunto de jogadores em que o jogador com o ID da Play tem +5 pontos.

Aumenta o score do Player com o ID do Play em 5 pontos.

score10ID

Arguments

:: Play
Jogada com um determinado ID.

-> [Player]
Conjunto de jogadores.

-> [Player]
Conjunto de jogadores em que o jogador com o ID da Play tem +10 pontos.

Aumenta o score do Player com o ID do Play em 10 pontos.

makeScore

Arguments

:: Play
Jogada.

-> State
Estado.

-> State
Estado atualizado, relativamente ao Score, tendo em conta o que o Pacman comeu.

Atualiza os pontos do State após uma jogada, verificando o que ele come para de seguida atribuir a respetiva pontuação a esse Pacman através das funções auxiliares anteriores.

Função de remoção de peça do labirinto

clearFood

Arguments

:: Play
Jogada.

-> State
Estado.

-> State
Estado atualizado relativamente à remoção de peças do labirinto, comidas pelo Pacman.

Retira a comida do labirinto, após ser comida pelo Pacman através da função no ficheiro Types, que dado uma coordena, uma peça, e um labirinto, ela substitui a peça na coordenada pela peça pedida, nesse determinado labirinto. Para remover a peça comida damos a coordenada do Player, a peça Empty e o labirinto do State, para assim remover a peça.

Funções que controlam o nº de vidas do Pacman, bem como a sua atualização para o estado Dying.

isAliveGhost

Arguments

:: Play
Jogada.

-> State
Estado.

-> Bool
Devolve True se algum fantasma vivo, está nas coordenadas do Pacman, caso contrário devolve False.

Verifica se tem um fantasma vivo nas coordenadas do Player com o ID do Play, através das coordenadas do Player e das coordenadas dos Ghosts Alive, verificando se a coordenada do Player coincide com a coordenada de algum Ghost com a função pré-definida "elem" do Prelude.

takeJgdrLife

Arguments

:: Play
Jogada.

-> [Player]
Conjunto de jogadores.

-> [Player]
Conjunto de jogadores, em que é retirado uma vida ou passa a Dying se não tiver mais nenhuma, ao Player com a mesma ID da Play.

Função auxiliar que tira uma vida a um Player ou que passa o Pacman para Dying se não tiver vidas restantes.

takeLife

Arguments

:: Play
Jogada.

-> State
Estado.

-> State
Estado em que se no seu [Player] houver algum Ghost vivo nas coordenadas do Pacman, então é retirado uma vida ou passa a Dying se não tiver vidas restantes.

Retira uma vida ao Pacman ou passa o Pacman a Dying se não tiver vidas restantes, quando o Pacman entra em contacto com um Ghost Alive.

Funções de transformação de estado dos Players

turnToMega

Arguments

:: Play
Jogada com um determinado ID.

-> [Player]
Conjunto de jogadores.

-> [Player]
Conjunto de jogadores em que o Player com o ID da Play, passa a Mega.

Dado uma Play e um conjunto de players, torna o Pacman com o ID da Play em Mega.

oposta :: Orientation -> Orientation

Função auxiliar para obter a orientação oposta, necessária para a Tarefa 5.

transGhDead

Arguments

:: [Player]
Conjunto de jogadores.

-> [Player]
Conjunto de jogadores em que os fantasmas que estavam vivos passaram a mortos.

Função auxiliar que transforma um conjunto de fantasmas vivos em fantasmas mortos e dando a orientação oposta à que tinham (a pedido da Tarefa 5), não fazendo alterações aos Players que não são fantasmas mortos.

makeMega

Arguments

:: Play
Jogada.

-> State
Estado.

-> State
Estado em que se o Pacman entrou em contacto com uma Big Food, então transforma-se em Mega e transforma todos os fantasmas vivos do [Player] em fantasmas mortos. Caso contrário, não há alterações relativamente ao PacMode do Pacman nem ao estado dos fantasmas.

Transforma o Pacman em Mega (caso coma Big Food), consequentemente transformando todos os fantasmas vivos em mortos.

Funções de teleporte dos fantasmas para a sua casa

resetSpawn

Arguments

:: Coords
Coordenadas pedidas.

-> State
Estado com um determinado labirinto.

-> [Player]
Conjunto de jogadores em que se um Ghost Dead estiver nas coordenadas pedidas, ele é mandado para as coordenadas obtidas através da função auxiliar anterior (que o mandam para a casa dos fantasmas) e o seu estado volta a ser Normal.

Função auxiliar que manda um Ghost comido pelo Pacman para a casa dos fantasmas, retornando o seu estado para Alive e a sua velocidade para o normal (multiplicando por 2).

goHome

Arguments

:: Play
Jogada com um determinado ID.

-> State
Estado.

-> State
Estado em que se o Pacman com o determinado ID, se encontra nas mesmas coordenadas de um Ghost Dead, então esse Ghost é enviado para a casa dos fantasmas e o seu estado e velocidade voltam ao normal.

Dado um Play e um State, dá um State com a alteração nas coordenadas e no estado do Ghost comido caso haja algum Pacman que tenha comido um Ghost Dead, caso contrário não há alterações nas coordenadas nem no estado de nenhum Ghost.

Função final

play

Arguments

:: Play
Jogada.

-> State
Estado.

-> State
Estado final em que dependendo da jogada realizada, o estado é atualizado, tendo em conta todas as ações a realizar.

Função final com todas as ações que o Pacman tem de realizar no decurso de uma jogada.

Funções complementares para a 2ª fase

ismouth :: [Player] -> [Player]

Fecha a boca aos Pacmans com boca aberta e vice-versa numa lista de jogadores. (talvez meter um matching para caso o pacman esteja dying)

moveMouth :: State -> State

Função que aplica na lista de jogadores de um State, a função anterior de fechar e abrir a boca aos Pacmans, consoante o necessário.

areMega :: [Player] -> Bool

Verifica se há algum Pacman no modo Mega.

turnToAlive :: [Player] -> [Player]

Transforma todos os fantasmas mortos em fantasmas vivos.

checkMegaPl :: [Player] -> [Player]

Transforma os fantasmas mortos em vivos se não houver nenhum Pacman no modo Mega.

checkMega :: State -> State

Função que aplica na lista de jogadores de um State, a função anterior que transforma os fantasmas mortos em fantasmas vivos se não houver nenhum Pacman em modo Mega.

checkTimeAux :: [Player] -> [Player]

Verifica se um Pacman tem MegaTime <= 0 estando em modo Mega, transformando-o para o estado Normal nesse caso. (TALVEZ METER AQUI O TIRAR O TEMPO DO MEGATIME POR PLAY)

checkTimeMg :: State -> State

Função que aplica na lista de jogadores de um State, a função anterior que transforma um Pacman para o estado Normal se um Pacman tem MegaTime <= 0 estando em modo Mega.

Funções-teste.

m1 :: Maze

Exemplo de labirinto.

testCasesPlay :: [(Play, State)]

Casos de Plays e States para a função play.

testItT2 :: [(Play, State)] -> [State]

Testa os casos para a função play.

isOnWall :: State -> Bool

Testa se algum Player de um State, se encontra indevidamente nas coordenadas de uma Wall, dando True caso haja.