Tarefa2_2018li1g101

Description

Este módulo define funções comuns da Tarefa 2 do trabalho prático.

Teste

testesT2 :: [(Int, Jogada, Estado)]

Testes unitários da Tarefa 2.

Cada teste é um triplo (identificador do Jogador,Jogada a efetuar,Estado anterior). testesT2 :: [(Int,Jogada,Estado)] testesT2 = [(0, Dispara Canhao,jogada(0,Movimenta C,jogada(0,Movimenta D,jogada(0,Movimenta E,jogada(0,Movimenta B,jogada (0,Dispara Laser,--jogada (0,Dispara Choque,estadoTeste)))))))]

mapaT4 :: Mapa

Mapa Tarefa 4

Estados para Testar

estadoTeste :: Estado

Estado Teste 1

estadoTeste2 :: Estado

Estado Teste 2

estadoTeste3 :: Estado

Estado Teste 3

estadoTeste4 :: Estado

Estado Teste 4

Funções principais da Tarefa 2.

jogada :: Int -> Jogada -> Estado -> Estado

Efetua uma jogada.

podeDisparar :: Arma -> Jogador -> Bool

verifica se o jogador tem disparos para poder disparar uma certa arma

decrementaDisparo :: Arma -> Jogador -> Jogador

decrementa o numero de disparos depois de efetuar um

direcaoIgualMovimento :: Direcao -> Jogador -> Bool

verifica se a direcao do jogador é igual a direcao do moviemento

criaDisparo :: Arma -> Int -> Estado -> Disparo

"criaDisparo" que recebe uma Arma o numero do jogador que dispara Int e um Estado e retorna o Disparo efetuado

moverJogador :: Jogador -> Estado -> Direcao -> Jogador

"moverJogador" move uma Jogador numa dada Direcao e retorna o Jogador movido com essa "DireçaomoverJogador :: Jogador -> Direcao -> Jogador

podeMover :: Jogador -> Estado -> Direcao -> Bool

verifica se um jogador pode mover numa determinada direcao,ou seja se essa posicao está vazio

verificaChoque :: Jogador -> Estado -> Bool

Valida se o jogador não está a ser afetado pelo choque para ver se é possivel mover ou não

verificaChoquesOcc :: Jogador -> [Jogador] -> Estado -> Bool

Verifica se o Jogador é afetado pelo Choque

validaChoque :: [Posicao] -> Posicao -> Bool

Recebe a Posicao do Jogador e a posição do Jogador em Choque e vê se está em choque

jogadorEmChoque :: Estado -> [Jogador]

Atraves do estado da retorno da lista de jogadores com Disparos Choque ativos

constroiListaDeJog :: [Int] -> Estado -> [Jogador]

Atraves da lista de indices de jogadores com Disparos Choque ativos cria uma lista de Jogadores

disparosParaIndJog :: [Disparo] -> [Int]

Constroi a lista de indices de jogadores com Disparos Choque ativos

verificaMoverValido :: Posicao -> Jogador -> Estado -> Bool

Valida se a posicao é valida para um jogador se mover para la

mapaComTanques :: [Jogador] -> Jogador -> Mapa -> Mapa

Substiti os tanks por blocos indestrutiveis no mapa

meteBlocoIndestrutivelNosTanques :: Posicao -> Mapa -> Mapa

Mete blocos indestrutiveis no mapa no lugar dos tanques