Tarefa2_2018li1g127

Description

Este módulo define funções comuns da Tarefa 2 do trabalho prático.

Testes

testesT2 :: [(Int, Jogada, Estado)]

Testes unitários da Tarefa 2.

Cada teste é um triplo (identificador do Jogador,Jogada a efetuar,Estado anterior).

jogada

Arguments

:: Int
O identificador do Jogador que efetua a jogada.

-> Jogada
A Jogada a efetuar.

-> Estado
O Estado anterior.

-> Estado
O Estado resultante após o jogador efetuar a jogada.

Efetua uma jogada.

choqueffect :: Int -> [Jogador] -> [Disparo] -> Bool

pegando no identificador do jogador, na lista de jogadores e na lista de disparos, verifica se o jogador x se encontra dentro da área de um choque senão o dele

auxjogcerto :: [Jogador] -> Int -> Jogada -> Mapa -> [Jogador] -> [Jogador]

devolve lista com jogador alterado

auxjogadorchoque :: Int -> [Jogador] -> Direcao -> [Jogador]

altera direcao jogador afetado com choques

auxjogador :: [Jogador] -> Jogada -> Mapa -> Jogador -> Jogador

altera jogador

auxdispara :: Arma -> Int -> Jogador -> Disparo

cria disparo

posvalida :: Posicao -> Direcao -> Mapa -> Bool

verifica se o jogador se pode mover para a direçao que pretende

jogcerto :: Int -> [Jogador] -> Jogador

jogador que efetua a jogada

verfdisp :: Jogador -> Arma -> Bool

verifica se o jogador pode efetuar o disparo

dispjog :: Int -> Arma -> [Jogador] -> [Jogador]

altera muniçoes

listpjog :: [Jogador] -> [PosicaoGrelha]

lista posiçoes jogadores

postanks :: PosicaoGrelha -> [PosicaoGrelha] -> Bool

posiçoes tanks

verchoque :: [Disparo] -> Bool

choque |verifica se ha choques no mapa

auxchoque :: Posicao -> Posicao -> Bool

posicao jogador disparou choque