Tarefa2_2019li1g085

Description

Este módulo define funções comuns da Tarefa 2 do trabalho prático.

Testes

testesT2

Arguments

:: [(Int, Jogada, Estado)]
triplo de teste

Testes unitários da Tarefa 2.

Cada teste é um triplo (identificador do Jogador,Jogada a efetuar,Estado anterior).

t1 :: (Int, Jogada, Estado)

Teste de Acelerar no Chao False.

t2 :: (Int, Jogada, Estado)

Teste de Acelerar no Ar.

t3 :: (Int, Jogada, Estado)

Teste de Disparar numa Rampa.

t4 :: (Int, Jogada, Estado)

Teste de Disparar numa Recta.

t5 :: (Int, Jogada, Estado)

Teste de Disparar sem Disparos.

t6 :: (Int, Jogada, Estado)

Teste de Desacelerar no Chao False.

t7 :: (Int, Jogada, Estado)

Teste de Desacelerar no Ar.

t8 :: (Int, Jogada, Estado)

Teste de Movimenta D no Ar (Inclinação 0).

t9 :: (Int, Jogada, Estado)

Teste de Movimenta E no Ar (Inclinação 0).

t10 :: (Int, Jogada, Estado)

Teste de Movimenta D no Ar (Inclinação -90).

t11 :: (Int, Jogada, Estado)

Teste de Movimenta D no Ar (Inclinação 90).

t12 :: (Int, Jogada, Estado)

Teste de Movimenta D no Chao.

t13 :: (Int, Jogada, Estado)

Teste de Movimenta E no Chao.

t14 :: (Int, Jogada, Estado)

Teste de Movimenta B (No limite).

t15 :: (Int, Jogada, Estado)

Teste de Movimenta C (No limite).

t16 :: (Int, Jogada, Estado)

Teste de Movimenta B no Chao (mesma altura).

t17 :: (Int, Jogada, Estado)

Teste de Movimenta C no Chao (mesma altura).

t18 :: (Int, Jogada, Estado)

Teste de Movimenta B no Chao (Jogador cai/salta).

t19 :: (Int, Jogada, Estado)

Teste de Movimenta C no Chao (Jogador cai/salta).

t20 :: (Int, Jogada, Estado)

Teste de Movimenta B no Chao (Jogador cai/morre).

t21 :: (Int, Jogada, Estado)

Teste de Movimenta C no Chao (Jogador cai/morre).

t22 :: (Int, Jogada, Estado)

Teste de Disparar na primeira Peca do Mapa.

t23 :: (Int, Jogada, Estado)

Teste de Movimenta B.

t24 :: (Int, Jogada, Estado)

Teste de Movimenta B (Jogador Morto).

t25 :: (Int, Jogada, Estado)

Teste de Disparo numa Peca que já é Cola.

t26 :: (Int, Jogada, Estado)

Teste de Movimenta B (Morto).

t27 :: (Int, Jogada, Estado)

Teste de Movimenta C (Ar).

t28 :: (Int, Jogada, Estado)

Teste de Desacelerar no Chao True.

t29 :: (Int, Jogada, Estado)

Teste de Desacelerar no Chao True.

t30 :: (Int, Jogada, Estado)

Teste de Acelerar no Chao True.

Funções principais da Tarefa 2.

jogada

Arguments

:: Int
O identificador do Jogador que efetua a jogada.

-> Jogada
A Jogada a efetuar.

-> Estado
O Estado anterior.

-> Estado
O Estado resultante após o jogador efetuar a jogada.

Efetua uma jogada.

Funções relativas a disparos

colocaCola

Arguments

:: Peca
Peca que se pretende modificar.

-> Peca
Peca modifcada.

Modifica uma Peca depois de um Dispara.

adicionaCola

Arguments

:: Mapa
Mapa por modificar.

-> Int
Pista que se encontra o Jogador.

-> Int
distanciaJogador.

-> Mapa
Mapa modificado.

Modifica o Mapa colocando a 'Peca Cola' na respetiva posição (Reparar que na posicao inicial não é suposto alterar nada).

retiraCola

Arguments

:: Int
Indice do Jogador.

-> [Jogador]
Lista de Jogador atual.

-> [Jogador]
Lista dos Jogador após um Dispara.

Retira uma Cola após um Dispara.

temCola

Arguments

:: Jogador
Jogador que Dispara.

-> Bool
Pode efetuar um Dispara?

Verifica se o Jogador se pode efetuar um Dispara.

Funções relativas aos movimentos horizontais

esquerda

Arguments

:: EstadoJogador
EstadoJogador do Jogador que 'Movimenta E'.

-> EstadoJogador
EstadoJogador após 'Movimenta E'.

Altera a inclinacaoJogador (Movimenta E).

modEstadoEsquerda

Arguments

:: Int
Indice do Jogador 'Movimenta E'.

-> [Jogador]
Lista de Jogador atual.

-> [Jogador]
Lista dos Jogador após um 'Movimenta E'.

Modifica o 'Estado Jogador' no jogador desejado ('Movimenta E').

direita

Arguments

:: EstadoJogador
EstadoJogador do Jogador que 'Movimenta D'.

-> EstadoJogador
EstadoJogador após 'Movimenta D'.

Altera a inclinacaoJogador (Movimenta D).

modEstadoDireita

Arguments

:: Int
Indice do Jogador 'Movimenta D'.

-> [Jogador]
Lista de Jogador atual.

-> [Jogador]
Lista dos Jogador após um 'Movimenta D'.

Modifica o 'Estado Jogador' no jogador desejado ('Movimenta D').

Funções relativas aos movimentos verticais

declive

Arguments

:: Int
Altura inicial.

-> Int
Altura final.

-> Double
Declive.

Devolve o declive de uma 'Recta'/'Rampa' - Declive == (y1-y2)/1 pois x apenas altera 1 (peça em peça).

percentagemPercorrida

Arguments

:: Double
distanciaJogador.

-> Double
Fração da Peca percorrida.

Devolve a "percentagem" percorrida de uma Peca.

alturaPista

Arguments

:: Double
distanciaJogador.

-> Int
Altura inicial da Peca.

-> Int
Altura inicial da Peca.

-> Double
Altura "atual".

Devolve a altura "atual" que depende da percentagem da Peca já percorrida pelo Jogador.

alturaInicial

Arguments

:: Peca
Peca analisada.

-> Int
Altura inicial.

Devolve a altura inicial de uma Peca.

alturaFinal

Arguments

:: Peca
Peca analisada.

-> Int
Altura final.

Devolve a altura final de uma Peca.

sobePista

Arguments

:: Jogador
Jogador que 'Movimenta C'.

-> Mapa
Mapa atual.

-> Jogador
Jogador após o 'Movimenta C'.

Devolve o resultado de alterar de pista para cima ('Movimenta C').

modEstadoCima

Arguments

:: Int
Indice do Jogador que 'Movimenta C'.

-> [Jogador]
Lista de Jogador atual.

-> Mapa
Mapa atual.

-> [Jogador]
Lista de Jogador após um 'Movimenta C'.

Modifica o 'Estado Jogador' no jogador desejado ('Movimenta C').

descePista

Arguments

:: Jogador
Jogador que 'Movimenta B'.

-> Mapa
Mapa atual.

-> Jogador
Jogador após o 'Movimenta B'.

Devolve o resultado de alterar de pista para baixo ('Movimenta B').

modEstadoBaixo

Arguments

:: Int
Indice do Jogador que 'Movimenta B'.

-> [Jogador]
Lista de Jogador atual.

-> Mapa
Mapa atual.

-> [Jogador]
Lista de Jogador após um 'Movimenta B'.

Modifica o 'Estado Jogador' no jogador desejado ('Movimenta B').

acelera

Arguments

:: Int
Indice do Jogador que Acelera.

-> [Jogador]
Lista de Jogador atual.

-> [Jogador]
Lista de Jogador após um Acelera.

Modifica o 'Estado Jogador' no jogador desejado (Acelera).

desacelera

Arguments

:: Int
Indice do Jogador que Desacelera.

-> [Jogador]
Lista de Jogador atual.

-> [Jogador]
Lista de Jogador após um Desacelera.

Modifica o 'Estado Jogador' no jogador desejado (Desacelera).

podeSubir

Arguments

:: Jogador
Jogador analisado .

-> Bool
Pode 'Movimenta C'?

Verifica se o Jogador não se encontra na 1ª Pista.

podeDescer

Arguments

:: Jogador
Jogador analisado.

-> Mapa
Mapa atual.

-> Bool
Pode 'Movimenta B'?

Verifica se o Jogador não se encontra na ultima Pista.

Funções secundárias

verificaChao

Arguments

:: Int
Indice do Jogador.

-> [Jogador]
Lista de Jogador.

-> Bool
Está no Chao?

Verifica se o Jogador está no Chao.

encontraJogador

Arguments

:: Jogador
Jogador que pretende encontrar.

-> (Int, Int)
Posicão encontrada do Jogador.

Dá a posição do jogador em forma de par ( nº da Pista , nº da Peca ).

encontraPistaJogador

Arguments

:: Jogador
Jogador da Pista que pretende encontrar.

-> Int
Numero da Pista encontrada.

Devolve a pistaJogador.

encontraDistanciaJogador

Arguments

:: Jogador
Jogador que pretende avaliar.

-> Int
Indice da Peca do Jogador.

Devolve o indice da Peca segundo o estado do Jogador.

distancia2indice

Arguments

:: Double
distanciaJogador.

-> Int
Indice da Peca.

Devolve o indice da Peca segundo a distancia a que Jogador se encontra.