Tarefa2_2021li1g056

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2021/22.

Documentation

constroiMapa :: [(Peca, Coordenadas)] -> Mapa

Esta tarefa consiste em construir e desconstruir um mapa.

Construir o mapa está dividido em três principais funções:

A função acrescentaVazios que cria uma lista de peças e coordenadas acrescentando todos os vazios à lista. Como auxiliares temos:

A função maiorXeY que determina o maior X e Y.

A função existeCoordenada que verifica se existe uma coordenada numa lista.

A função buscaPeca que através de umas coordenadas e de uma lista determina uma peça.

A função agrupa que agrupa listas pelos seus Ys.

A função buscaMatriz que seleciona as peças numa lista de [(Pecas,Coordenadas)]. Como auxiliar temos:

A função busca que seleciona as pecas numa lista.

Desconstruir o mapa está dividido em duas principais funções:

A função medidas que determina o valor máximo de X e Y.

A função desconstroiLinha que descontroi uma linha de peças para uma lista de [(Peca,Coordenadas)].

desconstroiMapa :: Mapa -> [(Peca, Coordenadas)]

função que desconstói um mapa para uma lista de peças e coordenadas

acrescentaVazios

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> [(Peca, Coordenadas)]
acrescenta vazios à lista recebida

função que acrescenta vazios à lista recebida

maiorXeY

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Coordenadas
determina o maior X e Y

função que determina o maior X e Y

existeCoordenada

Arguments

:: Coordenadas
recebe um par de coordenadas

-> [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Bool
verifica se uma coordenada existe numa lista

função que verifica se uma coordenada existe numa lista

buscaPeca

Arguments

:: Coordenadas
recebe um par de coordenadas

-> [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> Peca
determina a peça que tem essas coordenadas

função que determina a peça que tem essas coordenadas

agrupa

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> [[(Peca, Coordenadas)]]
constrói listas de listas guiando-se pelos Ys

função que constrói listas de listas guiando-se pelos Ys

buscaMatriz

Arguments

:: [[(Peca, Coordenadas)]]
recebe uma lista de peças e coordenadas

-> Mapa
constrói uma lista de lista de peças

função que constrói uma lista de lista de peças

busca

Arguments

:: [(Peca, Coordenadas)]
recebe uma lista de peças e coordenadas

-> [Peca]
constrói uma lista de peças

função que constrói uma lista de peças

medidas

Arguments

:: Mapa
recebe um mapa

-> Coordenadas
determina o máximo do X e do Y

função que determina o máximo do X e Y

desconstroiLinha

Arguments

:: [Peca]
recebe uma lista de peças

-> Coordenadas
um par de coordenadas que inicialmente vão ser (0,0)

-> [(Peca, Coordenadas)]
constrói uma lista de peças e coordenadas

função que constrói uma lista de peças e coordenadas