Tarefa2_2021li1g085

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2021/22.

Documentation

constroiMapa :: [(Peca, Coordenadas)] -> Mapa

A função constroiMapa é uma função que dada uma determinada lista de peças e coordenadas produz um mapa, isto é, uma lista de listas de peças, sendo estas respetivamente compostas pelas linhas da matriz que constituem o mapa. Enquanto que a lista de peças e coordenadas não apresenta os espaços Vazio do mapa, este exibe essas mesmas peças.

Quanto à função desconstroiMapa, esta realiza o inverso da função constroiMapa, ou seja, dado um determinado mapa, esta função é capaz de atribuir a cada peça as suas coordenadas, ocultando da sua lista as peças correspondentes a Vazio.

desconstroiMapa :: Mapa -> [(Peca, Coordenadas)]

retiracoor :: [(Peca, Coordenadas)] -> [Coordenadas]

Criamos uma função que, dada uma lista de peças e coordenadas, retribui apenas uma lista das coordenadas.

Exemplo: Tendo uma lista de peças e coordenadas, como por exemplo, [(Bloco,(0,2)),(Vazio,(1,1)),(Caixa,(1,2)),(Vazio,(1,0))], através da função retiracoor é possível obter uma lista constituída apenas pelas coordenadas, que seria, neste caso, [(0,2),(1,1),(1,2),(1,0)].

constroiVazios :: [Coordenadas] -> [(Peca, Coordenadas)] -> [Peca]

Tal como comentado previamente, a lista de peças e coordenadas que é atribuída à função constroiMapa não apresenta as peças Vazio, pelo que, é assim necessário construir uma função que irá atribuir essas mesmas peças. Deste modo, através da função constroiVazios, dada uma lista de coordenadas (lista 1) e uma outra lista de peças e coordenadas (lista 2), esta função irá verificar se as coordenadas da lista 1 pertecem às coordenadas da lista 2. Para isso, recorremos à função auxiliar retiracoor, que constroi uma lista de coordenadas a partir da lista 2. Dado isto, caso a _head_ da lista 1 pertença à lista 2, a função constroiVazios recorre à função auxiliar atribuiPeca que retribui a peça que corresponde às coordenadas da _head_ da lisa 1. Caso não se verifique que a _head_ da lista 1 pertence à lista 2, a função atribui a essa coordenada um Vazio e verifica o resto da lista.

atribuiPeca :: Coordenadas -> [(Peca, Coordenadas)] -> Peca

constroilistaPecas :: [(Peca, Coordenadas)] -> [Peca]

Dada uma lista de peças e coordenadas, a função constroilistaPecas retribui uma lista de peças. Esta função recorre a várias funções auxiliares, sendo a primeira a função retiracoor, que irá criar uma lista de coordenadas a partir da lista de peças e coordenadas introduzida. Tendo a lista de coordenadas, a função recorre à função parcoormax, que retribui um par de coordenadas, sendo este par as coordenadas máximas da matriz que, de seguida, serão usadas pela função ordenacoord de modo a obter uma lista de coordenadas ordenadas. Por fim, a função recorre à função constroiVazios, que dada a lista de coordenadas já ordenada e a lista de peças e coordenadas, retribui uma lista de peças que inclui os Vazio.

ordenacoord :: Coordenadas -> [Coordenadas]

A função ordenacoord ordena as coordenadas x e y que, recorrendo às funções auxiliares ordenaX e ordenaY, retribui uma lista de coordenadas ordenadas.

ordenaX :: Coordenadas -> [Coordenadas]

ordenaY :: Coordenadas -> [Coordenadas]

constroimapa :: [(Peca, Coordenadas)] -> [Peca] -> Mapa

A função constroiMapaAux retribui um mapa quando dada uma lista de peças e uma outra lista de peças e coordenadas. Dado que as listas que constituem o mapa são constituídas pelas peças que compõem as linhas da matriz, estas mesmas listas terão uma dimensão igual ao número de colunas da matriz, ou seja, o número de elementos dessas listas é igual ao x máximo, pelo que, uma das funções auxiliares que iremos utilizar, será a função maxX que devolve a abcissa máxima da matriz. A funçao constroiMapaAux recorre às funções auxiliares drop e take, pois através delas conseguimos controlar o tamanho e a formação das listas. Como a função take devolve os n primeiros elementos da lista, e como a função drop "elimina" esses n elementos da lista, através da construção desta função recursiva é nos possivel obter a estrutura do mapa.

listaX :: [Coordenadas] -> [Int]

A função listaX, dada uma lista de coordenadas, retribui uma lista composta apenas pelas abcissas.

listaY :: [Coordenadas] -> [Int]

A função listaY, dada uma lista de coordenadas, retribui uma lista composta apenas pelas ordenadas.

maxX :: [(Peca, Coordenadas)] -> Int

A função maxX devolve a abcissa máxima da matriz recorrendo às funções auxiliares listaX e retiracoor.

maxY :: [(Peca, Coordenadas)] -> Int

A função maxY devolve a ordenada máxima da matriz recorrendo às funções auxiliares listaY e retiracoor.

parcoormax :: [Coordenadas] -> Coordenadas

A função parcoormax devolve um par de coordenadas, sendo este par as coordenadas máximas da matriz.

removeVazio :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função removeVazio é responsável por remover os Vazio de uma lista de peças e coordenadas e devolver, de seguida, essa nova lista. Para isso a função verifica se a peça da _head_ da lista corresponde a Vazio e, se isto se verificar, a função retira esse conjunto de peças e coordenadas da lista e atua recursivamente sobre a função. Caso a peça da _head_ da lista não corresponda ao Vazio, então essa peça não é removida e a função atua de forma recursiva até não haverem mais elementos na lista.

desconstroi :: Mapa -> (Int, Int) -> [(Peca, Coordenadas)]

inverter :: Mapa -> Mapa

Dado que a função desconstroiMapa retribuía a lista de peças e coordenadas ao contrário, usamos a função inverter que, dado um mapa com uma determinada sequência, retribui a sequência inversa, compensando assim a função desconstroiMapa.