Tarefa2_2022li1g046

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> Int -> Mapa

Função final encarregue de juntar a nova linha gerada ao mapa já existente.

geraLinha :: Mapa -> Int -> (Terreno, [Obstaculo])

Função encarregue de gerar uma nova linha com a escolha de um terreno e uma lista de obstáculos de forma aleatória, onde esta aleatoriedade será criada através da seed que seja dada como input.

atribuiVelocidade :: Terreno -> Int -> Terreno

Função que atribui uma velocidade gerada ao terreno gerado.

gerarVelocidade :: Int -> Int

Função que gera um valor de velocidade.

gerarTerreno :: Mapa -> Int -> Terreno

Fução encarregue da escolha de um terreno aleatório para a criação de uma nova linha na função estendeMapa.

gerarObstaculos :: Int -> (Terreno, [Obstaculo]) -> Int -> [Obstaculo]

Função encarregue da escolha de uma lista de obstáculos aleatórios para a criação de uma nova linha na função estendeMapa.

getElemento :: Int -> [a] -> a

Função cujo propósito é escolher aleatóriamente um elemento de uma lista.

genRandN :: Int -> [a] -> Int

Função que gera de forma aleatória um número que esteja contido num intervalo de valores através do input de uma seed.

proximosTerrenosValidos :: Mapa -> [Terreno]

Possibilidades de terrenos que podem preencher nova linha do mapa.

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

Possibilidades de obstáculos que podem preencher nova linha do mapa.

proximosObstaculosValidosRio :: [Obstaculo] -> [Obstaculo]

Verifica as possibilidades de obstáculos no caso do terreno ser Rio.

contarTroncosRioSeguidos :: [Obstaculo] -> Int

Conta o número de troncos seguidos no final de uma lista de obstáculos.

validaTroncosSeguidos :: [Obstaculo] -> Bool

Valida a condição das 5 unidades ocupadas por um tronco.

proximosObstaculosValidosEstrada :: [Obstaculo] -> [Obstaculo]

Verifica as possibilidades de obstáculos no caso do terreno ser Estrada.

contarCarrosEstradaSeguidos :: [Obstaculo] -> Int

Conta o número de carros seguidos no final de uma lista de obstáculos.

validaCarrosSeguidos :: [Obstaculo] -> Bool

Valida a condição das 3 unidades ocupadas por um carro.

proximosObstaculosValidosRelva :: [Obstaculo] -> [Obstaculo]

Verifica as possibilidades de obstáculos no caso do terreno ser Relva.

carrosSeguidos :: [Obstaculo] -> [[Int]]

Torna uma lista de obstáculos em uma lista de listas de Int, em que Carro corresponde a 1 e qualquer outro obstáculo corresponde a 0

somaListas :: [[Int]] -> [Int]

Faz a soma de listas

somaListasF :: [Int] -> Int

Retorna o maior número da lista

troncosSeguidos :: [Obstaculo] -> [[Int]]

Torna uma lista de obstáculos em uma lista de listas de Int, em que Tronco corresponde a 1 e qualquer outro obstáculo corresponde a 0