Tarefa2_2022li1g061

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> Int -> Mapa

FUNÇAO QUE ACRESCENTA UMA NOVA LINHA AO MAPA

geraLista :: Int -> [Obstaculo] -> Int -> [Obstaculo]

FUNÇAO QUE GERA UMA NOVA LISTA DE OBSTACULOS AO MAPA

getRandom :: IO Int

FUNÇÃO QUE GERA UM NÚMERO ALEATÓRIO Esta função gera de maneira pseudo-aleatória um número que vai ser utilizado para a construção da próxima linha do mapa

proximosTerrenosValidos :: Mapa -> [Terreno]

FUNÇÃO QUE DETERMINA QUAIS OS TIPOS DE TERRENO VÁLIDOS PARA A NOVA LINHA GERADA Esta função recebe o mapa e retorna uma lista com todos os tipos de terreno aceitáveis para a continuação do mapa, Caso tenha 4 rios seguidos devolve as outras duas opções Caso tenha 5 estradas seguidas devolve as outras duas opções Caso tenha 5 relvas seguidas devolve as outras duas opções Se não cair em nenhum dos outros casos devolve todas as opções

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

FUNÇÃO QUE DETERMINA QUAIS OS OBSTÁCULOS VÁLIDOS PARA A POSIÇÃO SEGUINTE Esta função recebe a largura do mapa e um tuplo com o tipo de terreno e a lista de obstáculos contidos na linha Caso o comprimento da lista de obstáculos seja igual ou superior à largura, retorna uma lista vazia Caso contrário, invocará a sua função auxiliar para dar continuidade ao raciocínio

proximosObstaculosValidosAux :: (Terreno, [Obstaculo]) -> [Obstaculo]

FUNÇÃO QUE ENTREGA AS LISTAS DE OBSTÁCULOS VÁLIDOS Esta função recebe a largura e um tuplo com o terreno e a lista de obstáculos Se o terreno for Rio e o comprimento da lista de obstáculos for inferior à largura, e, por sua vez, o limite de Troncos seguidos não for atingido, retorna [Nenhum, Tronco], caso esta última condição não se verifique retornará apenas [Nenhum] Se o terreno for Estrada e o comprimento da lista de obstáculos for inferior à largura, e, por sua vez, o limite de Carros seguidos não for atingido, retorna [Nenhum, Carro], caso esta última condição não se verifique retornará apenas [Nenhum] Se o terreno for Relva e o comprimento da lista de obstáculos for inferior à largura, e, por sua vez, o obstáculo Nenhum fizer parte da lista de obstáculos, retorna [Nenhum, Arvore], caso esta última condição não se verifique retornará apenas [Nenhum]

troncosSeguidosT2 :: [Obstaculo] -> Int -> Bool

FUNÇÃO QUE VERIFICA SE JÁ FOI ATINGIDO O LIMITE DE TRONCOS SEGUIDOS Esta função recebe uma lista de obstáculos e utiliza um contador recursivo Se o contador ainda não tiver atingido 5 e o último obstáculo da lista for um Tronco, é incrementado e a função recomeça, se esta última condição não se verificar retorna False Se o contador chegar a 5, a função retorna True Caso a lista termine (significando que é menor ou igual a 5), se o contador tiver atingido 5, retorna True, caso não tenha atingido retorna False

carrosSeguidosT2 :: [Obstaculo] -> Int -> Bool

FUNÇÃO QUE VERIFICA SE JÁ FOI ATINGIDO O LIMITE DE CARROS SEGUIDOS Esta função recebe uma lista de obstáculos e utiliza um contador recursivo Se o contador ainda não tiver atingido 3 e o último obstáculo da lista for um Carro, é incrementado e a função recomeça, se esta última condição não se verificar retorna False Se o contador chegar a 3, a função retorna True Caso a lista termine (significando que é menor ou igual a 3), se o contador tiver atingido 3, retorna True, caso não tenha atingido retorna False

riosSeguidosT2 :: [(Terreno, [Obstaculo])] -> Int -> Bool

FUNÇÃO QUE VERIFICA SE JÁ FOI ATINGIDO O LIMITE DE RIOS SEGUIDOS Esta função recebe uma lista de tuplos com o terreno e uma lista de obstáculos e utiliza um contador recursivo Se o contador ainda não tiver atingido 4 e o terreno seguinte for Rio, é incrementado e a função recomeça, se esta última condição não se verificar retorna False Se o contador chegar a 4, a função retorna True Caso a lista termine (significando que é menor ou igual a 4), se o contador tiver atingido 4, retorna True, caso não tenha atingido retorna False

estradasSeguidasT2 :: [(Terreno, [Obstaculo])] -> Int -> Bool

FUNÇÃO QUE VERIFICA SE JÁ FOI ATINGIDO O LIMITE DE ESTRADAS SEGUIDAS Esta função recebe uma lista de tuplos com o terreno e uma lista de obstáculos e utiliza um contador recursivo Se o contador ainda não tiver atingido 5 e o terreno seguinte for Estrada, é incrementado e a função recomeça, se esta última condição não se verificar retorna False Se o contador chegar a 5, a função retorna True Caso a lista termine (significando que é menor ou igual a 5), se o contador tiver atingido 5, retorna True, caso não tenha atingido retorna False

relvaSeguidaT2 :: [(Terreno, [Obstaculo])] -> Int -> Bool

FUNÇÃO QUE VERIFICA SE JÁ FOI ATINGIDO O LIMITE DE RELVAS SEGUIDAS Esta função recebe uma lista de tuplos com o terreno e uma lista de obstáculos e utiliza um contador recursivo Se o contador ainda não tiver atingido 5 e os terrenos subsequentes forem ambos Relva, é incrementado e a função recomeça, se esta última condição não se verificar retorna False Se o contador chegar a 4, a função retorna True Caso a lista termine (significando que é menor ou igual a 4), se o contador tiver atingido 4, retorna True, caso não tenha atingido retorna False