Tarefa2_2022li1g115

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> Int -> Mapa

geraTerreno :: Mapa -> Int -> Terreno

A função geraTerreno gera novos terrenos com velodidades diferentes

gerarObstaculo :: Terreno -> [Obstaculo] -> Int -> Int -> [Obstaculo]

A função gerarObstaculo gera uma lista de novos obstaculos aleatorios, respeitando as restrições impostas pela função proximosObstaculosValidos

proximosTerrenosValidos :: Mapa -> [Terreno]

A função proximosTerrenosValidos recebe um mapa e devolve uma lista com os proximos terrenos possíveis respeitando o critério que limita o número máximo de de terrenos iguais seguidos, ou seja, não pode haver mais de 4 rios seguidos, nem 5 estradas seguidas, nem 5 relvas seguuidas

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

A função proximosObstaculosValidos, com a ajuda das 2 auxiliares, gera a lista dos seguintes obstaculos possíveis respeitando as restrições

pObsAux1 :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

A função pObsAux1 é a primeira função auxiliar da função proximosObstaculosValidos, esta garante que existe sempre um obstáculo Nenhum, para que seja o jogador consiga sempre progredir no jogo.

pObsAux2 :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

A função pObsAux2 é a segunda função auxiliar da função proximosObstaculosValidos, esta gera uma lista de obstaculos tendo em conta as seguintes condições: Se o terreno for Rio, apenas são permitidos os obstáculos Tronco, no máximo 5 seguidos, e Nenhum Se o terreno for Estrada, apenas são permitidos os obstáculos Carro, no máximo 3 seguidos, e Nenhum