Tarefa3

Documentation

basicCorr

Arguments

:: Corridor
Lista de peças que vão ser transformadas.

-> [(Int, Piece)]
Lista de tuplos na forma [(1, Piece)], em que cada tuplo corresponde a uma peça.

É uma função auxiliar que transforma um corredor de um labirinto, peça por peça, em uma lista de tuplos com o número 1 e a respetiva peça.

gatherBasic

Arguments

:: [(Int, Piece)]
Lista de tuplos não otimizadamente compacta.

-> [(Int, Piece)]
Lista de tuplos com os padrões horizontais compactados.

Função que junta os números, respetivos a peças iguais consecutivas, numa lista de tuplos [(Int, Piece)], dando uma lista mais compacta de tuplos (Int, Piece).

basic

Arguments

:: Maze
Conjunto de corredores que irão ser compactados num padrão horizontal.

-> Instructions
Instruções do labirinto compactas num padrão horizontal.

Função que transforma, primeiramente, um corredor em uma lista do tipo (Int, Piece), de seguida dá-se a compactação dessa lista através da função gatherBasic enunciada anteriormente, transformando cada lista de tuplos compactada numa instrução do tipo Instruct (Int, Piece), para cada corredor do labirinto, dando no final um conjunto de instruções :: Instructions.

elemIndices

Arguments

:: Eq a

=> a
Elemento que queremos verificar se ocorre na lista.

-> [a]
Lista onde queremos verificar a ocorrência do elemento.

-> [Int]
Lista de posições em que o elemento ocorre na lista.

Função auxiliar que calcula a lista de posições em que um dado elemento ocorre numa lista.

poeRepeatsaux

Arguments

:: Int
Inteiro n, que vai estar na consituição da Instruction, Repeat n, que vai subsituir uma determinada Instruction numa lista.

-> Int
Indice da instrução que pretendemos substituir.

-> Instructions
Lista de Instruction que queremos alterar.

-> Instructions
Lista de Instruction com a substituição pretendida.

Função auxiliar da função poeRepeats (função seguinte), que substitui a instrução (ordenada pelo inteiro recebido na 2ª entrada) numa lista de instruções, por uma instrução do tipo Repeat Int ,cujo Int é o fornecido no primeiro argumento.

poeRepeats

Arguments

:: Int
Inteiro n, que vai estar na consituição da Instruction, Repeat n, que vai subsituir certas Instruction de uma lista.

-> [Int]
Indices das instruções que pretendemos subsituir.

-> Instructions
Lista de Instruction que queremos alterar.

-> Instructions
Lista de Instruction com a substituição desejada.

Função que substitui as instruções (ordenadas pela lista de inteiros recebida no 2º argumento de entrada) de uma lista de instruções, por uma instrução do tipo Repeat Int, cujo Int é o fornecido no primeiro argumento.

repeater

Arguments

:: Instructions
Primeira Instructions em que vai ser verificado se cada elemento, tem algum igual à segunda Instructions.

-> Instructions
Segunda Instructions, que vai sofrer as alterações caso hajam elementos da primeira Instructions iguais a elementos desta segunda Instructions.

-> Instructions
Instructions com as alterações feitas à segunda Instructions.

Função auxiliar à função seguinte, que recebe duas instruções, e que verifica se cada elemento da primeira Instructions, tem algum elemento igual à segunda Instructions, e no caso de haver, esses elementos repetidos (excetuando o primeiro) que estão na segunda lista, vão ser substituídos por uma instrução do tipo Repeat n (em que n é o indice da primeira ocorrência desse elemento na primeira Instructions), através da funcão auxiliar anterior.

simpRepeater

Arguments

:: Instructions
Instructions na qual vai ser verificado se existem elementos iguais ao longo da lista e caso haja, vão ser substituidos (excetuando o primeiro elemento dos repetidos) por uma Instruction do tipo Repeat Int, em que o Int é o indice do primeiro elemento dos repetidos.

-> Instructions
Instructions final com os padrões verticais compactados.

Função que utiliza a anterior, para verificar elementos repetidos num mesmo labirinto.

compactMaze

Arguments

:: Maze
Labirinto que irá ser compactado.

-> Instructions
Instructions compactadas que correspondem ao labirinto.

Função final que compacta ao máximo um labirinto em instruções.

Funções-teste.

m2 :: [Maze]

Casos de labirintos para a função compactMaze.

testItT3 :: [Maze] -> [Instructions]

Testa os casos para a função compactMaze.

somatorio :: [(Int, Piece)] -> Int

Função auxiliar que faz o somatório do número de peças de uma lista de (Int, Piece).

validInstruction :: Maze -> Instruction -> Bool

Testa se a Instruction é válida (ou seja, se o somatório do número de peças é igual à largura) para um determinado labirinto.

validInstructions :: Maze -> Instructions -> Bool

Testa se as Instructions são válidas para um determinado labirinto.

decCorridor :: Instruction -> Corridor

Descompacta uma instrução num corredor.

decompact :: Instructions -> Maze

Descompacta Instructions básicas num labirinto.