Tarefa3_2018li1g101

Description

Este módulo define funções comuns da Tarefa 3 do trabalho prático.

Testes

testesT3 :: [Estado]

Testes unitários da Tarefa 3

Cada teste é um Estado.

Funções principais da Tarefa 3.

comprime :: Estado -> String

Comprime um Estado para formato textual.

Disparos

comprimeDisparos :: [Disparo] -> String

Comprime os disparos em formato textual.

Jogadores

comprimeJogadores :: [Jogador] -> String

Comprime os jogadores em formato textual.

posicaoString :: PosicaoGrelha -> String

Passa uma Posicao para uma String

comprimeMapa :: Mapa -> String

Comprime o Mapa em formato textual.

dimensoesMapa :: Mapa -> String

Dá as dimensões do Mapa

tupleToString :: [(Int, Char)] -> String

Passa [(Int,Char)] para uma String

ajuntaMapa :: [String] -> [(Int, Char)]

Condensa a informação do mapa quando há tipos de parede repetidos

comprimeBlocos :: Mapa -> Mapa -> Posicao -> String

Comprime os Tipos de Parede do Mapa numa String

descomprime :: String -> Estado

Descomprime um Estado no formato textual utilizado pela função comprime.

descodificaDisparos :: String -> Bool -> String

Descodifica os Disparos | Descodifica os Disparos

descodificaJogadores :: String -> Bool -> String

Descodifica os Jogadores

headIgualCarater :: Char -> String -> Bool

Verifica se a cabeça é igual ao caracter

descodificaDimensao :: String -> String

Descomprime a Dimensao do Mapa de forma a construir o Mapa

descomprimeMapa :: String -> Dimensao -> Mapa

Recebe a parte da String relativamente à parte do Mapa e descomprime o Mapa

descomprimeDimensao :: String -> Dimensao

Descomprime Dimensao

atualizaMapaComEstado :: String -> Posicao -> Mapa -> Mapa

Pega na String e constroi o mapa

trabalhaString :: String -> (Char, Int, String)

Trabalha a string de maneira a dar os caracteres para atualizar o mapa e o resto da string

atualizaMapaPorChar :: (Char, Int) -> Posicao -> Mapa -> (Mapa, Posicao)

Atualiza o Mapa com os tipos de parede do Estado descomprimido

descomprimeJogadores :: String -> Int -> [Jogador]

Funcao que recebe a string relativa aos jogadores e passa a descomprime Jogador para descomprimir os jogadores

descomprimeJogador :: String -> Jogador

Funcao que descomprime Jogador

descomprimePosicao :: String -> PosicaoGrelha

Funcao que descomprime a Posicao

stringToNumber :: [String] -> Char -> String

Passa uma string relativa ao jogador e obtem as strings com os valores das vidas , lasers , choques

descomprimeDisparos :: String -> Int -> [Disparo]

Descomprime a parte da string relativamente aos disparos de forma a passar a funcao descomprimeDisparo

descomprimeDisparo :: String -> Disparo

Cria o disparo atraves da string que recebe

getArma :: Char -> Arma

Recebe um char e retorna uma arma

criaDisparoCompresso :: (Arma, Int) -> PosicaoGrelha -> Direcao -> Ticks -> Disparo

Cria disparo da descompressao da String relativa a um disparo

stringAteChar :: String -> Char -> String

Retorna a string até encontrar o char