Tarefa3_2018li1g127

Description

Introdução Tarefa 3:

Neste módulo trabalhamos com 2 funções principais : comprime e descomprime

O objetivo desta tarefa é comprimir um estado, o máximo possível com a função comprime e em seguida, conseguir descomprimi-lo com sucesso.

A maior parte das dificuldades desta tarefa, foi à volta de arranjar forma de dividir o mapa comprimido, de modo a facilitar a descompressão. = Estratégias Utilizadas:

Compressão

Começamos por comprimir o mapa, a lista de jogadores e a lsita de disparos separadamente.

Juntamos tudo, usando a função principal comprime , separando o mapa da lista de jogadores por um "!" e a lista de jogadores por "&".

Compressão Mapa

Começamos por substituir um Bloco Destrutível por um "d", um Bloco Indestrutível por um "i" e um Bloco Vazio por um "v". As linhas são separadas por "/".

Apesar da redução ser significativa, nós quisemos ir à busca de mais, e por isso, se houvesse 2 ou mais "i""v""d" seguidos, subsituíriamos isso por número de blocos seguidos seguida da letra correspondente ao bloco. Assim, em mapas significativamente grandes, a nossa taxa de compressão iria ser alta.

Compressão Lista de Jogadores

Optámos por separar os vários jogadores por "|", a posição iria estar de forma a ser x-y seguida da direção seguida das vidas. para separar as vidas, dos lasers, dos choques, como eram todos ints, escolhemos usar "/" para separá-los.

Compressão Lista de Disparos

Escolhemos começar por identificar o tipo de disparo - "c" -> canhão; "q" -> choque ; "l" -> laser.

Seguido do tipo de disparo (laser ou canhão), vem o identificador (que apenas é 1 número) e logo a seguir vem a Linha, separada pela Coluna por um "-"; a seguir à coluna, vem uma letra - a direção.

No caso do disparo Choque, a seguir ao "q" vem o identificador (apenas 1 int) e logo a seguir vêm o numero de ticks.

A separar os vários disparos, escolhemos uma "/".

Descompressão

A descompressão foi onde tivemos mais dificuldades, pois tivémos de fazer funções para separar, além do mapa, da lista de jogadores e da lista de disparos, também os vários disparos e os vários jogadores.

Dividimos a descompressão em 3 fases: descompilaMapa , descompjog e descompldisparos .

Descompressão Mapa

Uma função importante foi a divstringm que dava logo a string correspondente ao mapa.

A partir dessa string, a descompilaMapa descompilava linha a linha de forma a que descompilasse tudo corretamente.

Descompressão Lista de Jogadores

Nesta fase, há uma outra função divstringj que retira a string correspondente à lista de jogadores.

Depois a função descompjog , com a ajuda das várias auxiliares que obtinham a posição das linhas, das colunas, a direção, as vidas, os lasers e os choques, comprime a lista de jogadores toda.

Descompressão Lista de Disparos

Em semelhança ás outras funções, esta usa uma função para obter apenas a lista de disparos : descompldisparos .

Como a descompressão da lista de jogadores, a dificuldade aqui passou em obter os identificadores, a posição e a direção ou os ticks - dependendo do tipo de disparo. Utilizou-se as várias auxiliares para ajudarem no "trabalho".

Conclusão:

Nesta tarefa a comprime/descomprime obteve uma boa compressão , mas sentimos algumas dificuldades na funcão desconprime.

Havia várias alternativas na comprime e tentamos que esta tivesse a maior taxa de compressão , tendo em conta a dificuldade que depois seria descomprimir.

Testes

testesT3 :: [Estado]

Testes unitários da Tarefa 3.

Cada teste é um Estado.

Funções principais da Tarefa 3.

comprime :: Estado -> String

Comprime um Estado para formato textual.

NB: A função show representa um Estado num formato textual facilmente legível mas extenso.

NB: Uma boa solução deve representar o Estado dado no mínimo número de caracteres possível.

compilaMapa :: Mapa -> String

se o mapa for constituido só por 1 linha, junta os blocos dessa linha, caso seja constituido por mais de uma linha, comprime os blocos dessa linha com a auxiliar "juntabc" e repete o processo para o restante mapa acrescentando um "/" no meio das linhas

juntabc :: [Peca] -> String

junta blocos consecutivos e iguais numa linha

auxjuntabc :: Int -> [Peca] -> String

se tiver pelo menos 2 blocos consecutivos, a função vai contar quantos e adicioná-los

mudabloco :: Peca -> String

recebe uma peça e converte-a numa String

compbloco :: Peca -> String

comprime peca para string

compldisparo :: [Disparo] -> String

comprime lista de disparos para string

compdisparo :: Disparo -> String

comprime disparo para string

comprimejog :: [Jogador] -> String

Passa cada Jogador para uma String

auxdirecao :: Direcao -> String

Dada uma direção, transforma-a numa String

descomprime :: String -> Estado

Descomprime um Estado no formato textual utilizado pela função comprime.

NB: A função comprime é válida de for possível recuperar o Estado utilizando a função descomprime, i.e.:

descomprime . comprime = id

NB: Esta propriedade é particularmente válida para a solução pré-definida:

read . show = id

divstringm :: String -> String

encontra a parte da string pertencente ao mapa

divstringj :: String -> String

divide string para lista de jogadores, eliminando a parte pertencente ao mapa, e depois a parte pertencente aos disparos

auxdivj :: String -> String

elimina toda a parte da string pertencente aos disparos

divstringd :: String -> String

divide string para lista de disparos eliminando tudo até à parte dos disparos

descompilaMapa :: String -> Mapa

descompila Mapa | inicia quando o primeiro caractere for parte do mapa

escolhelinhas :: String -> String

recebe a lista do mapa e devolve apenas uma linha do mapa

descompilaLinha :: String -> [Peca]

se for só 1 bloco, mete o bloco | caso apareça um número, vai "ler" o número e juntar esse número de elementos numa lista e continuar a função para os seguintes elementos da lista | chama a função para fazer uma linha, tendo como primeiro argumento uma linha do mapa e como segundo argumento, o caractere a repetir

auxlinha :: String -> String

junta os números

ajudadescomp :: String -> Char -> [Peca]

transforma a string num número; recebe o caracter pretendido para descompilar e chama a função para repetir esse elemento x vezes, descompilando assim uma linha do mapa

fazxlinha :: Int -> Char -> [Peca]

repete peça x vezes para construir uma linha

descompjog :: String -> [Jogador]

descomprime string para lista de jogadores

encontravida :: String -> String

vê quantas vidas o Jogador possui

encontralaser :: String -> String

encontra a partir de quando começam todos os argumentos pertencentes ao laser (depois do /)

auxencontralaser :: String -> String

junta todos os argumentos pertencentes ao laser

encontrachoque :: String -> String

vê quantos choques o Jogador possui | elimina parte das vidas

aux1encontrachoque :: String -> String

elimina número de lasers

aux2encontrachoque :: String -> String

encontra número de choques até que haja novo jogador ("|") ou até que acabe a lista de jogadores

descomplista :: String -> [Peca]

descomprime string para linha do mapa

descomppeca :: String -> Peca

descomprime string para peca

descompldisparos :: String -> [Disparo]

descomprime string para lista de disparos | se for & ou / remove da lista | descomprime o 1º disparo da lista, e repete o processo para os seguintes

takewhiledisp :: String -> String

1º disparo da lista

descdisparo :: String -> Disparo

descomprime string para disparo

takewhilepos :: String -> String

tira posição no formato x-y

takepos1 :: String -> String

retira a primeira posição

takepos2 :: String -> String

retira a segunda posição

desdir :: String -> Direcao

descobre Direção