Tarefa3_2019li1g085

Description

Este módulo define funções comuns da Tarefa 3 do trabalho prático.

Testes

testesT3

Arguments

:: [Mapa]
Mapa de teste

Testes unitários da Tarefa 3.

Cada teste é um Mapa.

m1 :: Mapa

Mapa simples, composto por 1 Pista de Recta Terra 0

m2 :: Mapa

Mapa com 2 Pista com Recta e Rampa de diferentes pisos

m3 :: Mapa

Mapa com 2 Pista iguais

m4 :: Mapa

Mapa com 4 Pista não iguais

m5 :: Mapa

Mapa com 4 Pista não iguais

m6 :: Mapa

Mapa com 4 Pista iguais com todos os tipos de Piso

m7 :: Mapa

Mapa com apenas 1 Pista e 2 Peca

m8 :: Mapa

Mapa gerado aleatoriamente a partir da Tarefa 1 (gera 5 6 1)

m9 :: Mapa

Teste

Funções principais da Tarefa 3.

desconstroi

Arguments

:: Mapa
Mapa atual.

-> Instrucoes
Mapa desconstruido em forma de Instrucoes.

Desconstrói um Mapa numa sequência de Instrucoes.

NB: Uma solução correcta deve retornar uma sequência de Instrucoes tal que, para qualquer Mapa válido m, executar as instruções '(desconstroi m)' produza o mesmo Mapa m.

NB: Uma boa solução deve representar o Mapa dado no mínimo número de Instrucoes, de acordo com a função tamanhoInstrucoes.

Funções de descontrução de mapa

inst

Arguments

:: Pista
Pista por descontruir.

-> Int
Indice da Pista.

-> Instrucoes
Pista desconstruida em forma de Instrucoes.

Transforma uma Pista em Instrucoes.

mapa2inst

Arguments

:: Int
Iteração ( == Indice da Pista).

-> Mapa
Mapa atual.

-> [Instrucoes]
Mapa desconstruido em forma de Instrucoes, dividido por Pista excluindo a primeira Peca.

Desconstroi o Mapa de Pista em Pista.

Funções de agrupar horizontalmente

horizontaisFromMatriz :: [Instrucoes] -> [Instrucoes]

Agrupa as sequencias horizontais de Instrucoes iguais. (Usada antes de agrupar verticalmente)

horizontais

Arguments

:: Instrucoes
Instrucoes com os padroẽs horizontais por agrupar.

-> Instrucoes
Instrucoes com os padroẽs horizontais agrupados.

Agrupa as sequencias horizontais de Instrucoes iguais.

Funções de agrupar verticalmente

verticais

Arguments

:: Instrucoes
Linha de Instrucoes por agrupar.

-> Instrucoes
Linha de Instrucoes agrupadas verticalmente.

Agrupa os padroẽs verticais de Instrucoes no mesmo "indice".

removeVerticais

Arguments

:: Instrucoes
Linha de Instrucoes com padrões verticais agrupados.

-> Instrucoes
Linha de Instrucoes apenas com as Instrucoes pretendidas.

Remove as Instrucoes que ficaram a mais após serem agrupadas pela função verticais.

aplicaVerticais

Arguments

:: Instrucoes
Linha de Instrucoes.

-> Instrucoes
Linha de Instrucoes.

Função que permite aplicar as funções verticais e removeVerticais e assim, agrupar verticalmente uma coluna da matriz.

lista2inst

Arguments

:: [Instrucoes]
Mapa desconstruido em forma de Instrucoes, dividido por Pista.

-> Instrucoes
Mapa inteiro em forma de Instrucoes com os padrões verticais já agrupados.

Transforma uma lista de Instrucoes orgazinadas por Pista numa Instrucoes apenas (Usada para ambas as sequencias de agrupar).

Funções de secundárias

instrucoesIguais

Arguments

:: Instrucao
Instrucao "atual".

-> Instrucao
Instrucao "inferior".

-> Bool
São iguais?.

Verifica se duas Instrucoes são iguais.

obtemIndice

Arguments

:: Instrucao
Instrucao analisada.

-> Int
Indice a que pertence.

Obtem o "indice" da Instrucao.

adicionaIndice

Arguments

:: Instrucao
Instrucao que onde vai ser agrupado ambos os indices

-> Instrucao
Instrucao que irá ser agrupada

-> Instrucao
Instrucao após agrupar

Adiciona o indice da segunda Instrucao à primeira (Só funciona para Instrucoes iguais).