Tarefa3_2021li1g056

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2021/22.

Documentation

criaJogo

Arguments

:: Jogador
recebe um jogador

-> Mapa
recebe um mapa

-> String
introduz o jogador e o mapa na mesma string

função que introduz o jogador e o mapa na mesma string

insereLinha

Arguments

:: Jogador
recebe um jogador

-> [Peca]
recebe uma lista de peças

-> Coordenadas
recebe um par de coordenadas que inicialmente vão ser (0,0)

-> String
introduz se possível o jogador numa linha (string) juntamente com peças

função que introduz se possível o jogador numa linha (string) juntamente com peças

mostraPeca

Arguments

:: Peca
recebe uma peça

-> String
transforma numa string

função que transforma uma peça numa string

mostraJogador

Arguments

:: Jogador
recebe um jogador

-> String
transforma numa string

função que transforma um jogador numa string

Orphan instances

Show Jogo
Esta tarefa consiste em representar o Jogo.

Para tal utiliza-se a função criaJogo que recebe um jogador e um mapa no qual vai criar, linha a linha (separado por "/n") através do insereLinha. O insereLinha consiste em pegar no jogador e pegar na primeira lista de peças do mapa e colocar o jogador na mesma string (linha) caso as coordenadas do jogador pertençam a essa linha se não avança para a próxima lista. Através da função mostraPeca e mostraJogador é possivel transformar uma valor do tipo peça ou jogador numa string.

Instance details

Methods

showsPrec :: Int -> Jogo -> ShowS

show :: Jogo -> String

showList :: [Jogo] -> ShowS