Tarefa3_2022li1g030

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Função animaJogo

Função principal

animaJogo :: Jogo -> Jogada -> Jogo

A função animaJogo, que recebe um jogo e uma jogada e que retorna um jogo, movimenta os obstáculos, de acordo com a velocidade do terreno em que se encontram, e o personagem, de acordo com a jogada dada.

A função animaJogo pode ser definida em função de outras funções auxiliares: a função ordenada, a função animaMapa, a função posicaoInicial, a função posicaoFinal e a função animaJogador. A função animaMapa contém ainda outra função auxiliar, a função animaLista que por sua vez contém outra função auxiliar, a função movObstaculos.

Assim, a função animaJogo pode ser definida da seguinte forma:

animaJogo (Jogo (Jogador (x,y)) (Mapa n l)) j | y == ordenada (animaJogador (Jogo (Jogador (x,y)) (Mapa n l)) j) = (posicaoFinal ((animaMapa (animaJogador (Jogo (Jogador (x,y)) (Mapa n l)) j))) (posicaoInicial (Jogo (Jogador (x,y)) (Mapa n l))))
                                              | otherwise = animaMapa (animaJogador (Jogo (Jogador (x,y)) (Mapa n l)) j)

Funções auxiliares

ordenada :: Jogo -> Int

A função ordenada, que recebe um jogo e retorna um número inteiro, descobre qual é a ordenada de um jogador num determinado jogo. Esta função é utilizada na função principal para percebermos se o jogador se movimentou horizontalmente ou verticalmente. Se o y inicial do jogador for igual ao y final do jogador após a função animaJogador, ou seja, após o seu movimento, então significa que o jogador apenas se moveu na horizontal (ou ficou parado). Assim, se apenas se moveu na horizontal então existe a possibilidade de ele acompanhar o movimento de um suposto tronco, isto se o jogador inicialmente se encontrava num tronco. Esta hipótese será confirmada ou rejeitada atráves da função posicaoInicial. Se o y inicial do jogador não for igual ao y final do jogador após a função animaJogador, ou seja, após o seu movimento, então significa que o jogador se moveu verticalmente, logo o movimento dos troncos não o poderá afetar.

Assim, a função ordenada pode ser definida da seguinte forma:

ordenada (Jogo (Jogador (x,y)) (Mapa n l)) = y  

movObstaculos :: Int -> (Terreno, [Obstaculo]) -> (Terreno, [Obstaculo])

A função movObstaculos, que recebe um inteiro e um par de terreno e de lista de obstáculos e que retorna um par de terreno e de lista de obstáculos, movimenta os obstáculos de um rio ou estrada de acordo com a velocidade do terreno. Numa estrada ou rio com velocidade v, os obstáculos devem mover-se |v| unidades na direção determinada. Se a velocidade v for igual a 0, os obstáculos não se movimentam, ou seja, a função irá retornar a mesma lista que recebeu, porque nada se alterou. Se a velocidade v for positiva, os obstáculos movem-se |v| unidades da esquerda para a direita. Se a velocidade v for negativa, os obstáculos movem-se |v| unidades da direita para a esquerda. Sabemos que ao deslocar os obstáculos de uma linha, estes desaparecerem por um dos lados do mapa e devem reaparecer no lado oposto. Assim, para mover os obstáculos apenas temos de mudar a ordem da lista de obstáculos. Quando a lista de obstáculos se move para a direita, quando a velocidade v é positiva, pegamos nos ultímos v elementos da lista e os colocamos na frente da lista. Para fazer isto podemos usar as funções last, init e (!!) pré-definidas no Haskell. Quando a lista de obstáculos se move para a esquerda, quando a velocidade v é negativa, pegamos nos primeiros v elementos da lista e os colocamos atrás da lista. Para fazer isto podemos usar as funções head, tail e (!!) pré-definidas no Haskell.

Assim, a função recursiva movObstaculos pode ser definida da seguinte forma:

movObstaculos n (Rio v, []) = (Rio v, [])
movObstaculos n (Estrada v,[]) = (Estrada v, [])
movObstaculos n (Rio v,l) | n==0 = (Rio v,l)
                          | n>0 = movObstaculos (n-1) (Rio v,[last l] ++ (init l))
                          | n<0 = movObstaculos (n+1) (Rio v,(tail l) ++ [head l])
movObstaculos n (Estrada v,l) | n==0 = (Estrada v,l)
                              | n>0 = movObstaculos (n-1) (Estrada v,[last l] ++ (init l))
                              | n<0 = movObstaculos (n+1) (Estrada v,(tail l) ++ [head l])

animaLista :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

A função animaLista, que recebe uma lista de pares de terreno e de lista de obstáculos e que retorna uma lista de pares de terreno e de lista de obstáculos, utiliza a função anterior (movObstaculos) para obter as listas de obstáculos dos rios e das estradas. Já numa relva, não existe velocidade e os obstáculos não se movimentam, ou seja, a função irá retornar a mesmma lista que recebeu, porque nada se alterou. Para além disso, se a função receber uma lista vazia irá retornar lista vazia, pois não existem obstáculos que se podem mover (a lista vazia não tem elementos logo não tem obstáculos).

Assim, a função recursiva animaLista pode ser definida da seguinte forma:

animaLista ((Rio n,l):t) = [movObstaculos n (Rio n,l)] ++ (animaLista t)
animaLista ((Estrada n,l):t) = [movObstaculos n (Estrada n,l)] ++ (animaLista t)
animaLista ((Relva,l):t) = [(Relva,l)] ++ (animaLista t)
animaLista [] = []

animaMapa :: Jogo -> Jogo

A função animaMapa, que recebe um jogo e retorna um jogo, utiliza a função animaLista para mover os obstáculos de todo o mapa do jogo.

Assim, a função animaMapa pode ser definida da seguinte forma:

animaMapa (Jogo (Jogador (x,y)) (Mapa n l)) = (Jogo (Jogador (x,y)) (Mapa n (animaLista l)))

posicaoInicial :: Jogo -> (Terreno, Obstaculo)

A função posicaoInicial, que recebe um jogo e retorna um par de terreno e de obstáculo, pega num jogo e calcula o terreno e o obstáculo onde o jogador se encontra naquele jogo. Para o fazer, utiliza a função (!!) pré-definida no Haskell.

Assim, a função posicaoInicial pode ser definida da seguinte forma:

posicaoInicial (Jogo (Jogador (x,y)) (Mapa n l)) = (fst (l !! y) , ((snd (l !! y)) !! x))

posicaoFinal :: Jogo -> (Terreno, Obstaculo) -> Jogo

A função posicaoFinal, que recebe um jogo e um par de terreno e obstáculo e retorna um jogo, calcula as próximas coordenadas do jogador, em função das coordenadas do jogador após as funções animaMapa e animaJogador e em função do terreno e do obstáculo em que o jogador se encontrava inicialmente. Se a função posicaoFinal receber o par de terreno e de obstáculo obtido pela função posicaoInicial e se receber o mesmo jogo obtido pela função animaMapa ao receber o jogo obtido pela função animaJogador, então podemos calcular as próximas coordenadas do jogador, as coordenadas finais do jogador.

Mesmo quando o jogador não efetua qualquer movimento (quando a sua jogada é Parado), se o personagem se encontrar em cima de um tronco, o jogador acompanha o movimento do tronco. Assim, se o jogador se encontrava num Rio com n velocidade e num Tronco, então o jogador mover-se-á n unidades numa determinada direção dependendo do valor de n. Assim, a abcissa do jogador passará de x para x+n. Se o jogador se encontrar num Rio e num Nenhum, ou se se encontrar noutro tipo de terreno, então o jogador não se irá mover com o mapa, logo irá apresentar as mesmas coordenadas.

Assim, a função posicaoFinal pode ser definida da seguinte forma:

posicaoFinal (Jogo (Jogador (x,y)) (Mapa n l)) (Rio v,Tronco) = (Jogo (Jogador (x+v,y)) (Mapa n l))
posicaoFinal (Jogo (Jogador (x,y)) (Mapa n l)) (_,o) = (Jogo (Jogador (x,y)) (Mapa n l))

animaJogador :: Jogo -> Jogada -> Jogo

A função animaJogador, que recebe um jogo e uma jogada e que retorna um jogo, movimenta um jogador num jogo.

Se a jogada for Parado então o jogo que retorna será o mesmo que recebeu, uma vez que o jogador não se moveu.

As jogadas Move Cima, Move Baixo, etc. fazem com que o jogador se mova 1 unidade para cima, baixo, etc, respectivamente. Se a jogada for Move Cima então a ordenada do jogador passará de y para y-1. Se a jogada for Move Baixo então a ordenada do jogador passará de y para y+1. Se a jogada for Move Esquerda então a abcissa do jogador passará de x para x-1. Se a jogada for Move Direita então a abcissa do jogador passará de x para x+1. Isto acontece porque o canto superior esquerdo representa as coordenadas (0,0) e porque todas as posições no mapa tem abcissas e ordenadas não negativas (ou seja, positivas ou nulas).

No entanto, temos que ter em atenção que o jogador não consegue escapar do mapa através dos seus movimentos. Por exemplo, se o jogador se encontrar na linha de topo do mapa, então mover-se para cima não tem qualquer efeito, uma vez que já se encontra no limite do mapa. Assim, quando y for igual a 0 e quando a jogada for Move Cima, o jogador não altera a sua posição. Seguindo o mesmo raciocínio, quando x for igual a 0 e a quando a jogada for Move Esquerda, o jogador não altera a sua posição. Quando x for igual á largura do mapa menos uma unidade (porque começamos a contar do 0) e quando a jogada for Move Direita, o jogador não altera a sua posição. Quando o y for igual ao comprimento da lista total do mapa menos uma unidade (porque começamos a contar do 0) e quando a jogada for Move Baixo, o jogador não altera a sua posição.

Por fim, ainda temos de ter em atenção que o jogador não ocupa posições onde existem Arvores. Ou seja, quando o jogador vai em direção a uma Arvore, ele acaba por se manter na mesma posição, uma vez que não pode ocupar a posição de uma Arvore. Assim, se uma Arvore se encontrar, por exemplo, em baixo do jogador, ou seja se posicaoInicial (Jogo (Jogador (x,y+1)) (Mapa n l)) resultar em (Relva,Arvore) e se a jogada for Move Baixo, então o jogador não se irá mecher. Isto também acontece quando tem uma Arvore em cima, à direita ou à esquerda do jogador e quando a jogada é respetivamente, Move Cima, Move Direita e Move Esquerda.

Assim, a função animaJogador pode ser definida da seguinte forma:

animaJogador (Jogo (Jogador (x,y)) (Mapa n l)) (Parado) = (Jogo (Jogador (x,y)) (Mapa n l))
animaJogador (Jogo (Jogador (x,y)) (Mapa n l)) (Move Cima) | y == 0 = (Jogo (Jogador (x,y)) (Mapa n l))
                                                           | posicaoInicial (Jogo (Jogador (x,y-1)) (Mapa n l)) == (Relva,Arvore) = (Jogo (Jogador (x,y)) (Mapa n l))
                                                           | otherwise = (Jogo (Jogador (x,y-1)) (Mapa n l))
animaJogador (Jogo (Jogador (x,y)) (Mapa n l)) (Move Baixo) | y == (length l) - 1 = (Jogo (Jogador (x,y)) (Mapa n l))
                                                            | posicaoInicial (Jogo (Jogador (x,y+1)) (Mapa n l)) == (Relva,Arvore) = (Jogo (Jogador (x,y)) (Mapa n l))
                                                            | otherwise = (Jogo (Jogador (x,y+1)) (Mapa n l))
animaJogador (Jogo (Jogador (x,y)) (Mapa n l)) (Move Esquerda) | x == 0 = (Jogo (Jogador (x,y)) (Mapa n l))
                                                               | posicaoInicial (Jogo (Jogador (x-1,y)) (Mapa n l)) == (Relva,Arvore) = (Jogo (Jogador (x,y)) (Mapa n l))
                                                               | otherwise = (Jogo (Jogador (x-1,y)) (Mapa n l))
animaJogador (Jogo (Jogador (x,y)) (Mapa n l)) (Move Direita) | x == (n-1) = (Jogo (Jogador (x,y)) (Mapa n l))
                                                              | posicaoInicial (Jogo (Jogador (x+1,y)) (Mapa n l)) == (Relva,Arvore) = (Jogo (Jogador (x,y)) (Mapa n l))
                                                              | otherwise = (Jogo (Jogador (x+1,y)) (Mapa n l))