Tarefa3_2022li1g046

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo :: Jogo -> Jogada -> Jogo

Função que atualiza o jogo, movendo os obstáculos dos terrenos e movendo o jogador de acordo com a jogada.

moveJogador :: Jogo -> Jogada -> Jogador

Função que move jogador tendo em conta as restrições impostas.

atualizaJogador :: Jogo -> Jogador

Função que atualiza a posição do jogador consoante a velocidade do terreno onde se encontra.

velocidadeTerreno :: (Terreno, [Obstaculo]) -> Int

Função que retira a velocidade de um dado terreno

linhaJogador :: Jogo -> (Terreno, [Obstaculo])

Função que retira a linha onde o jogador se encontra.

valPosTronco :: Jogador -> (Terreno, [Obstaculo]) -> Bool

Função que verifica se o jogador se encontra sob um tronco dado uma linha de um mapa.

valJogadorEmTronco :: Jogo -> Bool

Função que dado um qualquer mapa verifica se o jogador se contra sob um tronco.

invalMoveParaArvore :: Jogo -> Jogada -> Bool

Função que verifica dado um Jogo e uma Jogada se o jogador 'estará a ir em direção a uma árvore'.

atualizaTerreno :: (Terreno, [Obstaculo]) -> Jogador -> Int -> (Terreno, [Obstaculo])

Função que atualiza um terreno, tendo em conta a velocidade deste.

verificaAtropelamentoDireita :: Int -> (Terreno, [Obstaculo]) -> Jogador -> (Terreno, [Obstaculo])

Função que verifica se o jogador é atropelado quando os carros se movem para a direita

verificaAtropelamentoEsquerda :: Int -> (Terreno, [Obstaculo]) -> Jogador -> (Terreno, [Obstaculo])

Função que verifica se o jogador é atropelado quando os carros se movem para a esquerda

shiftDireitaLista :: Int -> [a] -> [a]

Função que movimenta uma lista de elementos n posições para a direita.

shiftEsquerdaLista :: Int -> [a] -> [a]

Função que movimenta uma lista de elementos n posições para a esquerda.

atualizaMapa :: [(Terreno, [Obstaculo])] -> Jogador -> Int -> [(Terreno, [Obstaculo])]

Função que atualiza o Mapa, tendo em conta a velocidade de cada Terreno.