Tarefa3_2022li1g061

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo :: Jogo -> Jogada -> Jogo

FUNÇÃO PRINCIPAL QUE FAZ O JOGO CORRER Esta função anima o jogo recebendo o jogo e a jogada jogadas, estando constantemente a gerar novos jogos, conforme as jogadas recebidas

gameToPlayer :: Jogo -> Jogador

FUNÇÃO AUXILIAR PARA TRANSFORMAR O JOGO APENAS NO JOGADOR Esta função serve para, conforme o jogo que for gerado na movePlayer, devolver apenas a posição em que o jogador se encontra

movePlayer :: Jogo -> Jogada -> Jogo

FUNÇÃO PARA CALCULAR OS MOVIMENTOS DO JOGADOR Esta função é um pouco complicada, mas no geral recebe um Jogo, a Jogada efetuada e devolve novamente o Jogo Caso a jogada seja Parado, caso o jogador esteja em cima de um Rio e de um Tronco, deve acompanhar esse tronco com auxilio da funçao auxiliar movePlayerObs. Caso nao esteja no tronco, fica parado. Caso a jogada seja "Move Direçao", tendo em conta as várias direções o jogador não pode ultrapassar arvores, nem pode sair do mapa com os seus movimentos.

movePlayerObs :: Jogador -> (Terreno, Obstaculo) -> Jogador

FUNÇÃO PARA MOVIMENTAR O JOGADOR DE ACORDO COM O TRONCO Esta função recebe um jogador, o par com o terreno e o obstaculo em que se encontra e devolve a posição do jogador com a movimentação do tronco Caso a velocidade seja maior que 0 o jogador move-se da esquerda para a direita à mesma velocidade que o tronco Caso a velocidade seja menor que 0 é o caso inverso Caso a velocidade seja 0 o jogador fica na mesma posição

moveObs :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

FUNÇÃO PARA MOVIMENTAR OBSTACULOS PRINCIPAL Esta funçao recebe uma lista de terrenos e obstaculos e devolve a mesma lista com as movimentações devidas Caso a lista recebida seja a lista vazia ele devolve a mesma lista vazia Caso seja um rio ou uma estrada (l.52 e 53) ele move o obstaculo de acordo com a funçao auxiliar Caso seja relva ele devolve a mesma lista, visto que relva nao tem velocidade

moveObsAux :: Velocidade -> [Obstaculo] -> Int -> [Obstaculo]

FUNÇÃO AUXILIAR PARA MOVIMENTAR OBSTACULOS Esta função recebe uma velocidade, uma lista de obstaculos e um contador e devolve uma lista de obstaculos com as movimentaçoes implementadas Caso a velocidade seja maior que 0 (mov da esq para dir) entao caso o contador seja menos que a velocidade (contador caso a velocidade seja maior que 1) Entao ele move o obstaculo ate a velocidade ser igual ao contador Caso a velocidade seja menor que 0 o caso é inverso