Tarefa4_2018li1g101

Description

Este módulo define funções comuns da Tarefa 4 do trabalho prático.

Testes

testesT4 :: [Estado]

Testes unitários da Tarefa 4.

Cada teste é um Estado.

Funções principais da Tarefa 4.

tick

Arguments

:: Estado
O Estado anterior.

-> Estado
O Estado após um Tick.

Avança o Estado do jogo um Tick de tempo.

NB: Apenas os Disparos afetam o Estado do jogo com o passar do tempo.

NB: Deve chamar as funçõs tickChoques, tickCanhoes e tickLasers pela ordem definida.

tickLasers :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos tiros de Laser disparados.

existelaser :: [Disparo] -> Bool

expandeLaser :: Estado -> Int -> Estado

andaLaser :: Disparo -> Int -> [Jogador] -> Estado -> Estado

elimina :: Disparo -> [Disparo] -> [Disparo]

novoDisparo :: Disparo -> Disparo

podeReplicar :: Disparo -> Estado -> Bool

isLaser :: Disparo -> Bool

explosaoLaser :: Estado -> [Jogador] -> (Estado, [Jogador])

newMapa :: [Disparo] -> Estado -> Estado

incrementaY :: Posicao -> Posicao

incrementaX :: PosicaoGrelha -> PosicaoGrelha

incrementaAmbos :: PosicaoGrelha -> PosicaoGrelha

tiroVertical :: Direcao -> Bool

newJogadores :: [Disparo] -> [Jogador] -> Estado -> (Estado, [Jogador])

atingiuAlgTank :: Posicao -> [Jogador] -> [Jogador] -> ([Jogador], [Jogador])

jogadorAtingido :: [Jogador] -> Jogador -> Bool

jogadoresIguais :: Jogador -> Jogador -> Bool

insereJog :: Jogador -> [Jogador] -> [Jogador]

insereEle :: Jogador -> ([Jogador], [Jogador]) -> ([Jogador], [Jogador])

insereList :: [Jogador] -> ([Jogador], [Jogador]) -> ([Jogador], [Jogador])

atualizaVida :: [Jogador] -> [Jogador]

newDisparos :: [Disparo] -> [Disparo] -> [Disparo]

comparatiros :: Disparo -> Disparo -> Bool

tickCanhoes :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos das balas de Canhao disparadas.

trabalhaCanhoes :: Estado -> Int -> Estado

eDisparoCanhao :: Disparo -> Bool

eliminaDisparo :: [Disparo] -> Int -> [Disparo]

atingiu :: Estado -> (Posicao, Direcao) -> Estado

destroiCanhoes :: (Posicao, Posicao) -> Estado -> Estado

destroiJogadores :: [Posicao] -> (Posicao, Posicao, Int) -> Estado -> Estado

acertouJogador :: Int -> Estado -> Estado

posTank :: [Jogador] -> [Posicao]

decreaseVidasJogador :: Int -> Int

destroiParede :: Peca -> Posicao -> Estado -> Estado

tickChoques :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos campos de Choque disparados.

trabalhaChoques :: [Disparo] -> [Disparo]