Tarefa4_2018li1g127

Description

Este módulo define funções comuns da Tarefa 4 do trabalho prático.

Testes

testesT4 :: [Estado]

Testes unitários da Tarefa 4.

Cada teste é um Estado.

Funções principais da Tarefa 4.

tick

Arguments

:: Estado
O Estado anterior.

-> Estado
O Estado após um Tick.

Avança o Estado do jogo um Tick de tempo.

NB: Apenas os Disparos afetam o Estado do jogo com o passar do tempo.

NB: Deve chamar as funções tickChoques, tickCanhoes e tickLasers pela ordem definida.

tickLasers :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos tiros de Laser disparados.

esclasers :: [Disparo] -> [Disparo]

cria lista apenas com os lasers que estao na lista de disparos

listjogadores :: [Disparo] -> [Jogador] -> Mapa -> [Jogador]

altera lista de jogadores de acordo com os efeitos dos lasers

afetajogador :: Disparo -> [Jogador] -> Mapa -> [Jogador]

altera jogador de acordo com o efeito de um laser

listdisparos :: [Disparo] -> [Disparo] -> Mapa -> [Disparo]

altera lista disparos de acordo com o efeito todos os lasers

afetadisparos :: Disparo -> [Disparo] -> Mapa -> [Disparo]

altera lista disparos de acordo com o efeito de um laser e elimina lasers da lista

mudamapa :: Mapa -> [Disparo] -> Mapa

altera mapa de acordo com o efeitos dos lasers

alteramapa :: Mapa -> Disparo -> Mapa

altera mapa de acordo com o efeito de um laser

blocosmapa :: Mapa -> Posicao -> Direcao -> Bool

verifica se há blocos na direcao do lasers

verifmapa :: Mapa -> Posicao -> Direcao -> Bool

verifica se ha blocos indestrutiveis no caminho do lasers

veriftank :: Mapa -> Posicao -> Posicao -> Direcao -> Bool

verifica se a posicao do tank afetada pelo laser

verifcanhao :: Mapa -> Posicao -> Posicao -> Direcao -> Bool

verifica se a posicao da balacanhao é afetada pelo laser

tickCanhoes :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos das balas de Canhao disparadas.

alteramapaC :: Mapa -> [Disparo] -> Mapa

recebendo o mapa e uma lista com disparos que chocam contra paredes, e , dependendo da direção, chama uma auxiliar para alterar o mapa corretamente

auxalteramapaCH :: Mapa -> Posicao -> Int -> Mapa

recebe o mapa, a posição onde os blocos se localizam, e um inteiro, que vai servir para destruir os 2 blocos que a bala tem contacto | se o bloco for destrutivel, então substitui o bloco por um bloco vazio e repete o processo para o bloco a baixo | se o bloco for indestrutivel, apenas repete o processo para o bloco a baixo

auxalteramapaCV :: Mapa -> Posicao -> Int -> Mapa

recebe o mapa, a posição onde os blocos se localizam, e um inteiro, que vai servir para destruir os 2 blocos que a bala tem contacto | se o bloco for destrutivel, então substitui o bloco por um bloco vazio e repete o processo para o bloco ao lado | se o bloco for indestrutivel, apenas repete o processo para o bloco ao lado

contraparede :: Mapa -> [Disparo] -> [Disparo]

Separa lista de disparos que chocam só contra uma parede

mataj :: [Jogador] -> [Disparo] -> [Jogador]

recebe uma lista de jogadores e uma lista de disparos | subtrai todas as vidas dos jogadores que intersetam disparos com a auxiliar subtraivida

subtraivida :: Jogador -> [Disparo] -> Jogador

Recebe um jogador e uma lista de disparos; se o jogador interceta com algum disparo, diminui 1 vida

contrajogador :: [Jogador] -> [Disparo] -> [Disparo]

reduz a lista de disparos apenas aos disparos que acertam em jogadores

verificabalatanquel :: Disparo -> [Jogador] -> Bool

verifica se uma bala acerta em algum jogador de uma lista

mantemdisp :: [Disparo] -> [Disparo]

recebe a lista de disparos iniciais e mantêm apenas os que não são canhão

separacanhao :: [Disparo] -> [Disparo]

Fica com uma lista só com disparos de canhão

modificadisparo :: Mapa -> [Jogador] -> [Disparo] -> [Disparo]

recebendo um mapa e uma lista de disparos, altera a lista de disparos, avançando aqueles em que a sua posição é vazia e retirando aqueles em que não é (choca com alguma coisa)

elimina :: [Disparo] -> [Disparo]

recebe uma lista de disparos e remove todos os disparos em que a posição é a mesma | se a lista for vazia, devolve a mesma, caso contrário, verifica se o primeiro disparo tem algum outro com a mesma posição; se sim, então chama uma auxiliar para eliminar todos os disparos com essa posição; caso contrário, mantêm o disparo e verifica o mesmo para todos os disparos seguintes

elemdisparos :: Disparo -> [Disparo] -> Bool

recebe um disparo e uma lista de disparos | se a posição do disparo for igual ao do primeiro da lista, então retorna Verdade; caso contrário, verifica para o resto da lista

eliminarepetidas :: Disparo -> [Disparo] -> [Disparo]

elimina todos os disparos com a mesma posição do disparo dado | se a posição do primeiro disparo for igual ao do primeiro da lista, então elimina o primeiro; caso contrário, verifica-o para o resto da lista

avancabala :: Disparo -> Disparo

recebendo um disparocanhao, avança esse disparo consoante a direção dada | Dependendo da direção, move a bala de canhão um bloco

verificadisparocostas :: Disparo -> [Disparo] -> Bool

recebe um disparo e uma lista de disparos e verifica se existe algum disparo nessa lista que esteja virado de costas para o mesmo

eliminacostas :: [Disparo] -> [Disparo]

Recebe uma lista de disparos e, usando uma auxiliar que elimina, numa lista de disparos, os disparos que estão nas costas, elimina todos os disparos numa lista que se encontram nas costas

auxeliminacostas :: Disparo -> [Disparo] -> [Disparo]

recebe um disparo e uma lista de disparos | se o disparo tem direção oposta, e está nas costas de outro disparo, ambos os disparos são eliminados

tickChoques :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos campos de Choque disparados.

alterachoques :: [Disparo] -> [Disparo]

altera os choques que estao na lista de disparos