Tarefa4_2019li1g181

Description

Introdução

O desafio da Tarefa 4 era calcular o efeito da passagem do tempo. Esta tarefa facilmente se dividia em duas partes: Atualizar a velocidade do ‘Jogador’ e Movimentar o ‘Jogador’.

Objetivos

Como mencionado na introdução, dividimos o trabalho em duas partes. Assim ficamos com dois objetivos. Numa primeira parte tínhamos que implementar a função ‘acelera’. Esta parte era muito teórica e envolvia na maior parte copiar fórmulas disponíveis no enunciado. Quando estava no chão a ‘velocidadeJogador’ era afetada pelo atrito das ‘Peca’s, pela sua velocidade inicial e tempo percorrido. No ar a ‘velocidadeJogador’ era afetada pela resistência do ar, pela velocidade inicial e tempo percorrido. A ‘gravidadeJogador’ era condicionada pela a gravidade inicial, pela aceleração da gravidade e pelo tempo decorrido.

Numa segunda fase, era necessário implementar a função ‘move’. Tal como a função ‘acelera’ esta também recorria a muitas fórmulas. No entanto, esta aborda a parte da trajetória. Para facilitar a abordagem deste problema, dividimos a função em três casos. Para quando estivesse no ar, no chão ou então morto. Quando estava no ar, era necessário saber se quando ele colidia com o chão se encontrava a mais de 45º de diferença de inclinação. Outro problema recorrente era também o facto, que se o ‘Jogador’ passa-se o fim da ‘Peca’ ele passava para o início da ‘Peca’ a seguir. Logo tivemos que arranjar uma forma de saber se o ‘Jogador’ passava o limite da ‘Peca’ e se passava ia simplesmente para o início da aseguir.

Conclusão

Esta tarefa não foi muito complexa, mas foi muito teórica. Resumidamente, era traduzir o que estava detalhado no enunciado para a linguagem de programação.

Testes

testesT4 :: [(Double, Mapa, Jogador)]

Testes unitários da Tarefa 4.

Cada teste é um par (tempo,Mapa,Jogador).

mapaTeste :: [[Peca]]

mapaTesteRampa :: [[Peca]]

Funções principais da Tarefa 4.

passo

Arguments

:: Double
O tempo decorrido.

-> Mapa
O mapa utilizado.

-> Jogador
O estado anterior do Jogador.

-> Jogador
O estado do Jogador após um passo.

Avança o estado de um Jogador um passo em frente, durante um determinado período de tempo.

passo' :: Double -> Mapa -> Jogador -> (Jogador, Double)

Função alternativa de passo que não desperdiça tempo

acelera

Arguments

:: Double
O tempo decorrido.

-> Mapa
O mapa utilizado.

-> Jogador
O estado anterior do Jogador.

-> Jogador
O estado do Jogador após acelerar.

Altera a velocidade de um Jogador, durante um determinado período de tempo.

calculaAtrito

Arguments

:: Peca
A Peca

-> Double
O atrito

Calcula o atrito de uma Peca

resAr :: Double

A resistência do ar

accelGravidade :: Double

A aceleração da gravidade

move' :: Double -> Mapa -> Jogador -> (Jogador, Double)

Função alternativa de move que não desperdiça tempo.

move

Arguments

:: Double
O tempo decorrido.

-> Mapa
O mapa utilizado.

-> Jogador
O estado anterior do Jogador.

-> Jogador
O estado do Jogador após se movimentar.

Altera a posição de Jogador, durante um determinado período de tempo.

moveAr :: Jogador -> Peca -> Ponto -> Ponto -> Jogador

ajustaVel

Arguments

:: Ponto
A posição atual

-> Ponto
A velocidade atual

-> Ponto
A velocidade ajustada

Diminui o módulo da velocidade para que esta não faça com que o jogador ultrapasse o limite da peça

colide

Arguments

:: Peca
A Peca a testar

-> Ponto
O Ponto inicial

-> Ponto
O vetor a somar ao Ponto

-> (Ponto, Bool)
A posição final e um Bool a indicar se ocorreu ou não uma colisão.

Calcula a posição final de um Ponto após este ser translatado por um vetor, sabendo que não pode acabar dentro de uma Peca.

Devolve a posição final e se ocorreu ou não uma colisão.

retaPeca

Arguments

:: Peca
A Peca para a qual calcular a Reta

-> Int
O índice da Peca no mapa

-> Reta
A Reta obtida

Calcula uma Reta que descreve a superfície de uma Peca.

Funções relativas ao caso Morto da função move

detMorto

Arguments

:: Double
O tempo decorrido.

-> Jogador
O estado antigo do Jogador.

-> Jogador
O estado atualizado do Jogador.

Determina se no intervalo de tempo o Jogador ainda está morto e se estiver decrementa um tempo

Funções relativas ao caso Chao na função move

moveChao

Arguments

:: Double
A inclinação da peça atual.

-> Double
A inclinação da próxima peça.

-> Double
A posição final.

-> Double
A altura final.

-> Double
O limite da peça atual.

-> Jogador
O Jogador a ser avaliado.

-> Jogador

checkMorto

Arguments

:: EstadoJogador
O EstadoJogador a verificar

-> Bool
O resultado