Tarefa4_2021li1g056

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2021/22.

Documentation

correrMovimentos :: Jogo -> [Movimento] -> Jogo

Esta tarefa consiste em definir os movimentos do jogador.

As funções eVazio, eCaixa e eCaixaOuBloco são usadas em vários movimentos com ou sem o jogador estar a carregar a caixa.

A função trocaCaixa é utilizada quando o jogador interage com uma caixa. Nas primeiras coordenadas que esta função recebe é inserida uma caixa no mapa através da função insereCaixa e nas segundas coordenadas que a função recebe é inserido um vazio no mapa através da função insereVazio.

A função gravidade determina até onde é possivel o jogador ou caixa caírem.

moveJogador :: Jogo -> Movimento -> Jogo

função que recebe um jogo e um movimento e altera o jogo

eVazio

Arguments

:: Coordenadas
recebe um par de coordenadas

-> Mapa
recebe um mapa

-> Bool
verifica se no mapa essas coordenadas pertencem a um vazio

função que verifica se no mapa essas coordenadas pertencem a um vazio

eCaixa

Arguments

:: Coordenadas
recebe um par de coordenadas

-> Mapa
recebe um mapa

-> Bool
verifica se no mapa essas coordenadas pertencem a uma caixa

função que verifica se no mapa essas coordenadas pertencem a uma caixa

eCaixaOuBloco

Arguments

:: Coordenadas
recebe um par de coordenadas

-> Mapa
recebe um mapa

-> Bool
verifica se no mapa essas coordenadas pertencem a uma caixa ou um bloco

função que verifica se no mapa essas coordenadas pertencem a uma bloco

trocaCaixa

Arguments

:: Coordenadas
recebe um par de coordenadas onde vai ser inserida a caixa

-> Coordenadas
recebe um par de coordenadas onde vai ser inserido o vazio

-> Mapa
recebe um mapa onde vão ser trocadas as peças

-> Mapa
constrói o mapa atualizado

função que constrói o mapa atualizado

insereCaixa

Arguments

:: Coordenadas
recebe um par de coordenadas onde vai ser inserida a caixa

-> Mapa
recebe um mapa

-> Mapa
insere a caixa nessas coordenadas do mapa

função que insere a caixa nessas coordenadas do mapa

insereVazio

Arguments

:: Coordenadas
recebe um par de coordenadas onde vai ser inserido o vazio

-> Mapa
recebe um mapa

-> Mapa
insere o vazio nessas coordenadas do mapa

função que insere o vazio nessas coordenadas do mapa

gravidade

Arguments

:: Coordenadas
recebe as coordenadas que vão sendo testadas até descobrir um obstáculo

-> Mapa
recebe um mapa

-> Coordenadas
recebe as coordenadas que vão ser anteriores ao obstáculo

-> Coordenadas
determina a posição que até onde foi possível descer

função que determina a posição que até onde foi possível descer