Tarefa4_2021li1g085

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2021/22.

Documentation

moveJogador :: Jogo -> Movimento -> Jogo

A função moveJogador recebe um Jogo e um Movimento retribuindo um novo Jogo, sendo este dependente do movimento aplicado.

correrMovimentos :: Jogo -> [Movimento] -> Jogo

A função correrMovimentos receber um Jogo e uma lista de Movimentos, retribuindo assim um novo Jogo, resultante da consecutiva aplicação de cada um dos movimentos pertencentes à lista.

posdirJog :: Jogador -> (Coordenadas, Direcao)

A função posdirJog recebe um Jogador e retribui as suas coordenadas e direção.

Exemplo: Se o jogador estiver na posição (0,3) e virado para Este e a carregar uma caixa (Jogador (0,3) Este True), através da aplicação da função recebemos ((0,3),Este).

desconstroiMapa :: Mapa -> [(Peca, Coordenadas)]

Para que o jogador execute um movimento é necessário saber se esse movimento é possível e, para isso, iremos recorrer à função desconstroiMapa, já utilizada na _Tarefa2_, que a partir de um mapa, devolve a sua respetiva lista de peças e coordenadas, incluindo as peças Vazio. Utilizamos esta função pois facilita a procura e verificação de peças.

desconstroi :: Mapa -> (Int, Int) -> [(Peca, Coordenadas)]

movimentoesquerda :: [(Peca, Coordenadas)] -> (Coordenadas, Direcao) -> Coordenadas

AndarEsquerda

A função movimentoesquerda verifica, a partir de uma lista de peças e coordenadas e das coordenadas do jogador, se este pode ou não executar o movimento _AndarEsquerda_.

Considerando _(x,y)_ as coordenadas do jogador:

Este movimento só é possível caso no espaço imediatamente à esquerda do jogador (x-1,y), exista uma peça Vazio, podendo ele, neste movimento, apenas deslocar-se no eixo das abcissas para a esquerda (caso se verifique a existência de um bloco ou caixa no espaço à sua esquerda e abaixo uma unidade, ou seja, em (x-1,y+1)) - 3ª guarda da função - ou sofrer uma situação de queda (caso não se verifique a existência de um bloco ou caixa no espaço à sua esquerda e abaixo uma unidade, ou seja, em (x-1,y+1)) - 5ª guarda da função (a 2ª guarda representa um caso de paragem para a situação de queda). Para a situação de queda recorremos à recursividade da função movimentoesquerda. Caso não se verifiquem estas condições, o jogador não altera a sua posição.

movimentodireita :: [(Peca, Coordenadas)] -> (Coordenadas, Direcao) -> Coordenadas

AndarDireita

A função movimentodireita verifica, a partir de uma lista de peças e coordenadas e das coordenadas do jogador, se este pode ou não executar o movimento _AndarDireita_.

Considerando _(x,y)_ as coordenadas do jogador:

Este movimento só é possível caso no espaço imediatamente à direita do jogador (x+1,y), exista uma peça Vazio, podendo ele, neste movimento, apenas deslocar-se uma unidade no eixo das abcissas para a direita (caso se verifique a existência de um bloco ou caixa no espaço à sua direita e abaixo uma unidade, ou seja, em (x+1,y+1)) - 3ª guarda da função - ou sofrer uma situação de queda (caso não se verifique a existência de um bloco ou caixa no espaço à sua direita e abaixo uma unidade, ou seja, em (x+1,y+1)) - 5ª guarda da função (a 2ª guarda representa um caso de paragem para a situação de queda). Para a situação de queda recorremos à recursividade da função movimentodireita. Caso não se verifiquem estas condições, o jogador não altera a sua posição.

treparsemcaixa :: [(Peca, Coordenadas)] -> (Coordenadas, Direcao) -> Coordenadas

TreparSemCaixa

Para que o personagem seja capaz de trepar sem caixa, temos de verificar a peça imediatamente à frente do jogador (tem de ser um obstáculo) e a peça acima desta (que tem de ser um Vazio).

Exemplo:

Considerando _(x,y)_ as coordenadas do jogador. Caso o personagem esteja voltado para Este e se verifique que a peça imediatamente à sua frente (x+1,y) é um obstáculo (uma caixa ou um bloco) e que a peça acima desta (x+1,y-1) é Vazio, então o jogador assume as coordenadas (x+1,y-1). Caso o personagem estivsse voltado para Oeste e se encontrasse dentro das mesmas condições, assumiria as coordenadas (x-1,y-1). Independentemente da direção, caso não se verificassem as condiçõesnecessárias para a realização do movimento, o personagem não alteraria a sua posição.

jogapanhacaixa :: Mapa -> (Coordenadas, Direcao) -> Mapa

Interagir com caixa

-| JogadorApanhaCaixa

Para que o jogador possa executar o movimento InterageCaixa é necessário que imediatamente à sua frente exista uma caixa e que a peça acima desta seja Vazio, sendo que quando o jogador apanha a caixa, esta é removida do mapa, sendo substituída por Vazio. Para remover a caixa do mapa construímos a função jogapanhacaixa que recorre a uma função auxiliar, a apanhacaixa, que recebe uma lista de peças e coordenadas e um conjunto de peça e coordenadas, retribuindo uma nova lista. Dado que a função jogapanhacaixa recebe um mapa e as coordenadas e direção do personagem e retribui um mapa, é necessário avaliar os casos em que o jogador pode apanhar uma caixa.

Exemplo:

Caso o jogador esteja voltado para Este e exista uma caixa imediatamente à sua frente (condição verificada através da função desconstroiMapa, que comprova se (Caixa,(x+1,y)) pertence à lista de peças e coordenadas correspondente ao mapa), a função recorre então à auxiliar apanhacaixa que utiliza essa mesma lista e o conjunto (Caixa,(x+1,y)) para substituir a caixa por uma nova peça, o Vazio. A função apanhacaixa atua recursivamente. De seguida, através da função constroiMapa, atribuindo lhe a nova lista gerada na função auxiliar, obtemos o novo mapa, isto é, sem a caixa.

apanhacaixa :: [(Peca, Coordenadas)] -> (Peca, Coordenadas) -> [(Peca, Coordenadas)]

jogpousacaixa :: Mapa -> (Coordenadas, Direcao) -> Mapa

JogadorPousaCaixa

O método utilizado para o jogador pousar a caixa é idêntico ao utilizado para o jogador apanhar a caixa.

Para que o jogador possa executar o movimento InterageCaixa é necessário que imediatamente à sua frente exista um Vazio, sendo que quando o jogador pousa a caixa, esta é inserida do mapa. Para acrescentar a caixa do mapa construímos a função jogapousacaixa que recorre a uma função auxiliar, a pousacaixa, que recebe uma lista de peças e coordenadas e um conjunto de peça e coordenadas, retribuindo uma nova lista. Dado que a função jogapousacaixa recebe um mapa e as coordenadas e direção do personagem e retribui um mapa, é necessário avaliar os casos em que o jogador pode pousar uma caixa.

Exemplo:

Caso o jogador esteja voltado para Este e exista um Vazio imediatamente à sua frente (condição verificada através da função desconstroiMapa, que comprova se (Vazio,(x+1,y)) pertence à lista de peças e coordenadas correspondente ao mapa), a função recorre então à auxiliar pousacaixa que utiliza essa mesma lista e o conjunto (Vazio,(x+1,y)) para substituir o Vazio por uma nova peça, uma Caixa. A função pousacaixa atua recursivamente. De seguida, através da função constroiMapa, atribuindo lhe a nova lista gerada na função auxiliar, obtemos o novo mapa, isto é, com a caixa.

pousacaixa :: [(Peca, Coordenadas)] -> (Peca, Coordenadas) -> [(Peca, Coordenadas)]

inverte :: [[Peca]] -> [[Peca]]

A função inverte é usada para inverter o mapa uma vez que a função desconstroiMapa aplicada nesta tarefa retribui os resultados invertidos.

podeavancardireita :: [(Peca, Coordenadas)] -> (Coordenadas, Direcao) -> Coordenadas

AndarDireita com caixa

A função podeavancardireita verifica, a partir de uma lista de peças e coordenadas e das coordenadas do jogador, se este pode ou não executar o movimento _AndarDireita_ enquanto carrega uma caixa.

Considerando _(x,y)_ as coordenadas do jogador:

Este movimento só é possível caso no espaço imediatamente à direita do jogador (x+1,y) e acima desse espaço (x+1,y-1), existam peças Vazio, podendo o jogador, neste movimento, apenas deslocar-se uma unidade no eixo das abcissas para a direita (caso se verifique a existência de um bloco ou caixa no espaço à sua direita e abaixo uma unidade, ou seja, em (x+1,y+1) e, caso exista um Vazio no espaço acima desse, ou seja, (x+1,y-1)) - 3ª guarda da função - ou sofrer uma situação de queda (caso se verifiquem os Vazios em (x+1,y) e em (x+1,y-1) e não se verifique a existência de um bloco ou caixa no espaço à sua direita e abaixo uma unidade, ou seja, em (x+1,y+1)) - 5ª guarda da função (a 2ª guarda representa um caso de paragem para a situação de queda). Para a situação de queda recorremos à recursividade da função podeavancardireita. Caso não se verifiquem estas condições, o jogador não altera a sua posição.

podeavancaresquerda :: [(Peca, Coordenadas)] -> (Coordenadas, Direcao) -> Coordenadas

AndarEsquerda com caixa

A função podeavancaresquerda verifica, a partir de uma lista de peças e coordenadas e das coordenadas do jogador, se este pode ou não executar o movimento _AndarEsquerda_ enquanto carrega uma caixa.

Considerando _(x,y)_ as coordenadas do jogador:

Este movimento só é possível caso no espaço imediatamente à esquerda do jogador (x-1,y) e acima desse espaço (x-1,y-1), existam peças Vazio, podendo o jogador, neste movimento, apenas deslocar-se uma unidade no eixo das abcissas para a esquerda (caso se verifique a existência de um bloco ou caixa no espaço à sua esquerda e abaixo uma unidade, ou seja, em (x-1,y+1) e, caso exista um Vazio no espaço acima desse, ou seja, (x-1,y-1)) - 3ª guarda da função - ou sofrer uma situação de queda (caso se verifiquem os Vazios em (x-1,y) e em (x-1,y-1) e não se verifique a existência de um bloco ou caixa no espaço à sua direita e abaixo uma unidade, ou seja, em (x-1,y+1)) - 5ª guarda da função (a 2ª guarda representa um caso de paragem para a situação de queda). Para a situação de queda recorremos à recursividade da função podeavancaresquerda. Caso não se verifiquem estas condições, o jogador não altera a sua posição.

treparcomcaixa :: [(Peca, Coordenadas)] -> (Coordenadas, Direcao) -> Coordenadas

TreparComCaixa

Para que o personagem seja capaz de trepar com caixa, temos de verificar a peça imediatamente à frente do jogador (tem de ser um obstáculo) e as duas peças acima desta (que têm de ser um Vazio).

Exemplo:

Considerando _(x,y)_ as coordenadas do jogador. Caso o personagem esteja voltado para Este e se verifique que a peça imediatamente à sua frente (x+1,y) é um obstáculo (uma caixa ou um bloco) e que as duas peças acima desta (em (x+1,y-1) e (x+1,y-1)) são Vazios, então o jogador assume as coordenadas (x+1,y-1). Caso o personagem estivesse voltado para Oeste e se encontrasse dentro das mesmas condições, assumiria as coordenadas (x-1,y-1). Independentemente da direção, caso não se verificassem as condições neessárias para realizar o movimento, o personagem não alteraria a sua posição.