Tarefa4_2022li1g046

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

jogoTerminou :: Jogo -> Bool

Função principal que emgloba as diversas funções auxiliares que verifica se o jogo terminou ou não.

foraDoMapa :: Jogo -> Bool

Função definida que verifica se as coordenadas da posição do jogador estão dentro dos limites do mapa.

larguraMapa :: Mapa -> Int

Função que retira a largura do mapa dado.

alturaMapa :: Mapa -> Int

Função que retira a altura do mapa dado.

invalPosRio :: Jogador -> [Obstaculo] -> Bool

Função que verifica se o jogador se encontra numa posição inválida caso o terreno seja ´Rio´.

invalPosEstrada :: Jogador -> [Obstaculo] -> Bool

Função que verifica se o jogador se encontra numa posição inválida caso o terreno seja Estrada.

invalPos :: Jogador -> (Terreno, [Obstaculo]) -> Bool

Função que verifica se o jogador se encontra numa posição inválida, tendo em conta todos os tipos de terrenos.

validaPosicao :: Jogo -> Bool

Função que verifica se o jogador se encontra numa posição inválida dependendo da linha onde se encontra e do tipo de terreno da mesma.