Tarefa4_2022li1g061

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

gameOver :: Jogo -> Bool

FUNÇÃO QUE INDICA QUE O JOGO ACABOU Esta função verifica todas as auxiliares e, caso algumas deles retorne true, então o jogador está morto ou fora do mapa, o que indica que o jogo acaba

offMap :: Jogo -> Bool

FUNÇÃO QUE INDICA SE O JOGADOR ESTÁ FORA DO MAPA Esta função recebe o jogo e, caso a posição do jogador seja fora do mapa, retorna True indicando que o jogador está fora.

dead :: Jogo -> Bool

FUNÇAO QUE VERIFICA SE O JOGADOR ESTÁ MORTO Esta função, pega nas funções atropelado e afogado e, caso alguma retorne que o jogador está morto, então esta função indica que ele está morto.

atropelado :: Jogo -> Bool

FUNÇÃO QUE VERIFICA SE O JOGADOR ESTÁ NUMA ESTRADA COM CARRO Esta função, como o nome indica, verifica se o jogador foi atropelado. Ou seja, a posição onde ele está é uma estrada e um carro. Retorna true caso ele esteja morto.

afogado :: Jogo -> Bool

FUNÇÃO QUE VERIFICA SE O JOGADOR ESTÁ NUM RIO SEM TRONCO Esta função, como o nome indica, verifica se o jogador está afogado. Ou seja, a posição onde ele está é um rio e não tem tronco. Retorna true caso ele esteja morto.

selectT :: [(Terreno, [Obstaculo])] -> Int -> (Terreno, [Obstaculo])

FUNÇÃO QUE DEVOLVE A LINHA DO MAPA ONDE O JOGADOR SE ENCONTRA Esta função recebe o mapa inteiro (exceto a largura) e a linha em que o jogador se encontra e devolve o par em que o jogador se encontra, ou seja a linha da matriz onde ele está

busca :: Jogo -> (Terreno, Obstaculo)

FUNÇÃO QUE PROCURA A POSIÇÃO DO JOGADOR Esta função, com auxilio da função acima, devolve o par terreno obstaculo daquela posição da matriz Ou seja, diz-nos com precisão em que terreno o jogador está