Tarefa4_2022li1g115

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

jogoTerminou :: Jogo -> Bool

A função jogoTerminou verifica quando o jogador perde, ou seja, uma das seguintes situações: O jogador saiu do limite do mapa, ou seja, ou as cordenas do jogador são negativas, ou a primeira cordenada do jogador é maior que o numero de obstáculos, ou então a segunda cordenada do jogador é maior do que a largura do mapa. O jogador foi atropelado (as cordenadas do jogador coincidem com as de um carro). O jogador caiu na água (as cordenadas do jogador não coincidem com as de um tronco quando o terreno é Rio).