Tarefa5

Description

Introdução

Esta tarefa consistiu em implementar um comportamento para os fantasmas. Quer para fugir do Pacman, quer para ir atrás do Pacman. A tarefa inicialmente aparentou ser um pouco intimidadora, uma vez que teríamos de ter várias possibilidades e variáveis em consideração para determinar a melhor jogada possível.

Objetivos

Inicialmente foi necessário fazer alterações na Tarefa 2 para admitir o movimento dos fantasmas o que provou ser um pouco trabalhoso, uma vez que implicou fazer várias alterações em várias funções da Tarefa 2 para assegurar o correto funcionamento do jogo.

Para saber que jogada fazer, é necessário saber as coordenadas do fantasma (em questão) e do Pacman. Para isso elaboramos novas funções e usamos também algumas previamente feitas para conseguir extrair as coordenadas dos Players corretamente. Após ter as funções de extração de coordenadas, era necessário funções que analisassem essas coordenadas e a sua contextualização no labirinto para que após a análise, devolvesse uma jogada a fazer. Para tal, foram feitas funções seguindo as sugestões do enunciado do trabalho prático, as quais sugeriam que tentassemos determinar a direção a tomar para o fantasma se aproximar o máximo possível do Pacman. Após os professores terem anunciado que tinhamos maior liberdade do que estava previsto no enunciado, decidimos melhorar um pouco mais a estratégia dos fantasmas adicionando mais algumas possibilidades, que não estavam a ser previstas pelo fantasma, tendo sido melhorada minimamente a eficácia dos fantasmas, relativamente à estratégia utilizada anteriormente.

Com a inteligência dos fantasmas quisemos que eles fossem desafiantes, mas também não queriamos que os fantasmas fossem extremamente eficazes a ir atrás do Pacman, algo que poderia ser frustrante para o jogador.

Discussão e conclusão

Foi divertido ver nesta tarefa o resultado das linhas de código que escrevemos em ação e fugirmos dos fantasmas, brincando com o jogo para o testar, tal como se estivessemos a jogar o jogo original do Pacman. Ficamos satisfeitos com o desempenho da inteligência artificial dos fantasmas, apesar de ter demonstrado ser uma tarefa trabalhosa e complicada.

Documentation

ghostIDCoords :: State -> Int -> Coords

Função que diz em que coordenadas está o Ghost com o determinado ID.

chCoord :: Coords -> Maze -> Orientation -> Coords

Função que dá a coordenada para a determinada orientação que é recebida como argumento, relativamente à coordenada inicial pedida.

nxtWall :: Coords -> Maze -> Orientation -> Bool

Função que testa se há uma parede no sitio para onde o Player iria tendo em conta a orientação recebida como argumento.

listCoords6 :: Coords -> Orientation -> [Coords]

Dada uma coordenada e uma orientação, devolve o conjunto das coordenadas, que se encontram 5 unidades de distancia relativamente à orientação fornecida.

filterCoords

Arguments

:: [Coords]
Conjunto de coordenadas criadas pela função anterior.

-> Maze
Labirinto onde se enquadram as coordenadas.

-> Coords
Coordenadas da posição do Pacman.

-> [Coords]
Resultado final filtrado pelas restrições programadas.

Recebe o conjunto de coordenadas, produzido pela função anterior e filtra as coordenadas que se encontram fora do labirinto e que incluam as paredes que cobrem o labirinto, e pára de construir a lista quando uma das coordenadas coincide com as do Pacman.

checkRetaWall :: [Coords] -> Maze -> Bool

Verifica (com uma função da Tarefa2) cada coordenada da lista filtrada anteriormente e se houver alguma coordenada em que haja uma Wall no labirinto, devolve False.

checkForWalls

Arguments

:: Coords
Coordenadas do Ghost.

-> Orientation
Orientação para a qual quer verificar se há paredes.

-> Maze
Labirinto onde se enquadram as coordenadas.

-> Coords
Coordenadas do Pacman.

-> Bool
Retornará True se o caminho estiver livre, caso contrário retornará False.

Função que utiliza todas as funções auxiliares anteriores, para verificar se o fantasma pode ir atrás do Pacman sem ir contra nenhuma parede.

coordsChase

Arguments

:: State
Estado do jogo.

-> Int
ID de um determinado fantasma.

-> Coords
Coordenadas do Pacman.

-> Coords
Coordenadas do fantasma com o determinado ID.

-> Play
Play a executar pelo fantasma com o respetivo ID com a intenção de ir atrás do Pacman.

Função auxiliar que dado um State, um ID, as coordenadas do Pacman e do Ghost com o determinado ID, decide a jogada a fazer para o Ghost ir atrás do Pacman.

chaseMode :: State -> Int -> Play

Função final que utiliza a função auxiliar anterior para dar um determinado ordem (Play) ao fantasma com o determinado ID, para ir atrás do Pacman.

coordsScatter

Arguments

:: State
Estado do jogo.

-> Int
ID de um determinado fantasma.

-> Coords
Coordenadas do Pacman.

-> Coords
Coordenadas do fantasma com o determinado ID.

-> Play
Play a executar pelo fantasma com o respetivo ID com a intenção de fugir do Pacman.

Função auxiliar que dado um State, um ID, as coordenadas do Pacman e do Ghost com o determinado ID, decide a jogada a fazer para o Ghost fugir do Pacman.

scatterMode :: State -> Int -> Play

Função final que utiliza a função auxiliar anterior para dar um determinado ordem(Play) ao fantasma com o determinado ID, para fugir do Pacman.

ghostPlayAux :: [Player] -> State -> [Play]

Função final adaptada para funcionar na Tarefa 4.

ghostPlay :: State -> [Play]

Função final que dado um State, retorna o conjunto de Plays a fazer pelos jogadores do State.