Tarefa5_2022li1g046

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2022/23.

Documentation

deslizaJogo :: Int -> Jogo -> Jogo

Função encarregue do "deslize" do mapa, adicionando uma nova linha e removendo a última.

atualizaMap :: Mapa -> Int -> Mapa

Função auxiliar que adiciona uma nova linha ao mapa dado, sem a última linha (uso do init)