Tarefa6

Description

Introdução

Esta tarefa consistiu em desenvolver um bot que fosse capaz de jogar sozinho sem nenhum input humano.

Objetivos

O objetivo desta tarefa era dar jogadas ao Pacman, tendo em conta o estado do jogo e do mapa. Definimos como objetivos ir atrás do fantasmas mortos, fugir de fantasmas vivos, e quando estiver perto, comer uma Big Food. Como objetivo principal estabelecemos que era fugir a fantasmas vivos, priorizando estar vivo a obter pontos, para tal o bot verifica se existe algum fantasma vivo a aproximar-se dele e caso haja, ele tenta ir na direção oposta para fugir do perigo. Secundariamente, ou seja, caso não houvesse perigo por perto, o bot iria atrás de fantasmas mortos, que dão uma maior pontuação e eliminam a ameaça do mapa (temporariamente). Caso não detete nenhuma destas circunstâncias, ele irá tentar ir atrás de comidas grandes para conseguir aumentar a sua pontuação e forçar os ghosts a ficarem Dead.

Discussão e conclusão

Apesar de não ser um bot extremamente inteligente, com uma inteligência artificial espetacular, dá para ver que é capaz de cumprir os seus objetivos fugindo do perigo e indo atrás de oportunidades para pontuar. Esta tarefa foi uma tarefa trabalhosa, mas, tal como na Tarefa 5, é muito satisfatório ver o código que se escreveu ser posto em prática ao testar e ver os fantasmas e o Pacman uns atrás dos outros, aplicando as estratégias que desenhamos para eles.

Documentation

neargL :: Coords -> [Coords] -> Bool

Função auxiliar que verifica as coordenadas do Pacman e dos fantasmas Alive, retornando True se houver algum fantasma Alive perto, a 4 unidades à esquerda.

neargR :: Coords -> [Coords] -> Bool

Função auxiliar que verifica as coordenadas do Pacman e dos fantasmas Alive, retornando True se houver algum fantasma Alive perto, a 4 unidades à direita.

neargU :: Coords -> [Coords] -> Bool

Função auxiliar que verifica as coordenadas do Pacman e dos fantasmas Alive, retornando True se houver algum fantasma Alive perto, a 4 unidades em cima.

neargD :: Coords -> [Coords] -> Bool

Função auxiliar que verifica as coordenadas do Pacman e dos fantasmas Alive, retornando True se houver algum fantasma Alive perto, a 4 unidades em baixo.

coordsMPl :: State -> Int -> Coords -> Maybe Play

Função auxiliar que dado um State, um Int, e as coordenadas do Pacman no labirinto, devolve a jogada a fazer pelo bot.

turnR :: Orientation -> Orientation

Função que dada uma orientação, dá a orientação seguinte, no sentido dos ponteiros do relógio.

chaseBF

Arguments

:: Coords
Coordenadas do Pacman

-> [Coords]
Conjunto de coordenadas das peças de comida grande

-> Int
ID do Pacman

-> Orientation
Orientação atual do Pacman

-> Maybe Play
Play a fazer tendo em conta a análise das coordenadas

Função que analisa as coordenadas do Pacman e dá a Play a realizar para ir atrás de uma das coordenadas das peças de comida grande que se encontra próximo. Caso não haja nenhuma peça próxima ele irá continuar continuar com a sua orientação.

bot :: Int -> State -> Maybe Play

Função final que recebe um ID e um State e utilizando as funções auxiliares anteriores, devolve a Play a fazer pelo Pacman bot.