Tarefa6_2018li1g127

Description

Introdução Tarefa 6:

Neste módulo trabalhamos com 2 funções principais, bot e reagedir que recebe já as informações do bot.

O objetivo deste módulo passou por criar um bot que, recebendo o estado, conseguisse mover-se no mapa, disparar contra os jogadores e defender-se.

Estratégias utilizadas:

As jogadas que são feitas, são todas graças à função reagedir que recebe o identificador do bot, o Estado e o estado do Bot, e que dependendo das várias posições, lista de disparos, etc... , escolhe a melhor opção para essa situação específica.

Uma função importante utilizada, senão a mais importante, é a verificaBloco que, dando 2 posições, verifica se existe blocos insdestrutiveis entre essas posições. Esta função é das mais importantes, pois se existir um bloco indestrutivel entre o bot e um jogador/disparo, o bot não vai se "preocupar" com eles.

Existe ainda outras duas funções essenciais: eixox e eixoy - O objetivo destas funções é verificar se existe algum jogador que esteja em linha em relação ao Bot seja numa linha (eixox) ou numa coluna (eixoy). Caso não exista, o Bot não se deve preocupar com esse Jogador.

Defesa:

Inicialmente começamos por implementar no bot, funções que fariam com que ele de certa forma se defendesse. Sempre que se deparasse com alguém que possuisse lasers, ele iria fugir para onde havia caminho livre.

Pode-se dizer que o coração da defesa está na função reagedir , porque é lá que tem as condições necessárias para que o Bot consiga fugir dos jogadores com sucesso.

Como já foi referido em cima, as funções eixox e eixoy têm um papel fundamental na Defesa.

Além disso, se houver disparos de canhão em direção ao bot, ele vai virar-se para os disparos e disparar também de forma a defender-se.

Uma função importante para esta situação é a reagedis que, conforme as condições impostas, exporta a melhor opção e a suscdisparo que verifica se existe algum disparo canhão em direção ao Bot.

Exemplo Defesa:

Um jogador que esteja virado horizontalmente para o bot com lasers, o bot iria fugir para cima, caso o caminho estivesse livre, senão iria para baixo. Em ultimo caso, iria virar-se para o jogador e disparar.

Ataque:

Depois dessas funções estarem implementadas, decidimos acrescentar algumas funções em que o bot iria atacar. Estas foram inseridas na função já feita reagedir .

A maior parte destas funções acontecem quando o Jogador em causa, não possui lasers. Caso isto se verifique, sempre que um Jogador estiver na linha/coluna do bot, o bot vai virar-se na direção em que se encontra o Jogador, e começar a disparar Lasers (caso os tenha) ou Canhões.

Mais tarde, acrescentamos a função em que, se o bot for afetado por um choque, ele vai usar o seu choque também para nenhum dos Jogadores se puder mexer, e assim o bot não fica em desvantagem.

Exemplos Ataque:

Um jogador que se encontre na mesma coluna/linha que o bot, caso o bot tenha lasers, ele vai virar-se para o jogador e disparar os seus lasers. Quando já não tiver lasers, ele dispara Canhões só, e só apenas se o Jogador não estiver virado para o Bot, ou se o Jogador estiver virado para ele, mas tiver 0 lasers.

Um jogador que utilize o choque que afete o bot, o Bot vai recontribuir com um choque caso os tenha.

Caso Neutro:

Este caso acontece quando não há jogadores a quem reagir, nem disparos. Neste caso, inicialmente pensámos em mantê-lo quieto, mas face aos resultados dos torneios, decidimos colocá-lo a andar pelas bordas no sentido dos ponteiros do relógio à procura de Jogadores.

Exemplos Neutro:

Caso não haja nenhum jogador para reagir, o Bot vai começar a andar para cima enquanto possível, depois para a direita, para baixo e para a esquerda conforme as condições.

Caso esteja com blocos destrutiveis à volta, vai destrui-los.

Conclusão:

No fim, pensamos que o Bot não ficou perfeito, mas consegue complicar a vida aos outros jogadores. Ele consegue esquivar-se dos lasers, atacar os jogadores, defender-se dos canhões, contra-atacar os choques e adaptar-se a mapas adversos.

Por isso, penso que a tarefa ficou completa, e apesar do Bot não jogar como um ser-humano, penso que ficou bem construido, adaptando-se tanto à posição dos vários jogadores, como à lista de disparos.

Funções principais da Tarefa 6.

bot

Arguments

:: Int
O identificador do Jogador associado ao robot.

-> Estado
O Estado para o qual o robot deve tomar uma decisão.

-> Maybe Jogada
Uma possível Jogada a efetuar pelo robot.

Define um robot capaz de jogar autonomamente o jogo.

reagedir :: Int -> Estado -> Jogador -> Maybe Jogada

recebe o identificador , o estado, o estado do Bot e aplica uma jogada consoante as condições

botchoque :: Int -> [Disparo] -> Bool

recebe o identificador do bot, uma lista de disparos, e verifica que existe algum choque pertencente ao bot

verificaBloco :: Posicao -> Posicao -> Mapa -> Bool

verifica se existe Bloco Indestrutivel entre duas posições

verificaLinhax :: Posicao -> Posicao -> Mapa -> Bool

verifica se existe algum bloco insdestrutivel entre a posição do bot e outra posição ( para y > b - Bot à esquerda)

verificaLinhaxi :: Posicao -> Posicao -> Mapa -> Bool

verifica se existe algum bloco insdestrutivel entre a posição do bot e outra posição ( para y < b - Bot à direita)

verificaColunay :: Posicao -> Posicao -> Mapa -> Bool

verifica se existe algum bloco insdestrutivel entre a posição do bot e outra posição ( para x > a - Bot em cima)

verificaColunayi :: Posicao -> Posicao -> Mapa -> Bool

verifica se existe algum bloco insdestrutivel entre a posição do bot e outra posição ( para x < a - Bot em baixo)

suscdisparo :: Posicao -> [Disparo] -> Bool

recebe a posição do Bot, uma lista de disparos, e verifica se existe algum disparo canhão a ir na direção do bot

verificaDisparos :: Posicao -> Disparo -> Bool

compara a posição do bot e a posição do disparo e verifica se está a dirigir-se na direção do bot

reagedis :: Direcao -> Posicao -> [Disparo] -> Mapa -> Maybe Jogada

recebe a direção a direção do Bot, a sua posição, a lista de disparos e o mapa e , consoante condições, executa uma jogada | se existir um bloco indestrutível entre as 2 posições ou se o disparo não estiver na direção do bot, executa a função para o resto da lsita de disparos

oposta :: Direcao -> Direcao

função que transforma uma direção na direção oposta

eixox :: Int -> Int -> Bool

dando a linha do jogador e a linha do bot, verifica se está em zona de perigo

eixoy :: Int -> Int -> Bool

dando a coluna do jogador e a coluna do bot, verifica se está em zona de perigo