Tarefa6_2019li1g181

Description

Introdução

Nesta tarefa o desafio que encontramos foi o de munir um robot com a capacidade de navegar pelo Mapa da forma mais eficiente. Este robot tinha que a partir do Estado mais atualizado tomar decisão sobre que Jogada iria fazer, ou então se simplesmente não fazia nada. O resultado final foi um robot que escolhe qual o caminho mais rápido e que nunca morre.

Objetivos

O objetivo final era conseguir que o robot determinasse o caminho mais rápido. No entanto, o Mapa pode-se alterar devido a ações dos outros Jogadores. Isto é, o Mapa pode ser diferente desde o ponto de partida até o ponto de chegada. Assim tivemos que avaliar como é que íamos determinar o melhor caminho. A nossa estratégia foi avaliar as peças da vizinhança do robot e comparar qual delas tinha o menor atrito. A Peca que tinha o menor era para onde o robot se movia. (Ainda quero arranjar uma forma de expandir a vizinhança para ele tomar melhores decisões).

A seguir, reparamos que era importante mudar a inclinação do robot. Isto porque, caso a diferença de inclinação fosse maior que 45º em relação à peça que ele aterrava, o robot morria e perdia tempo. Assim comparamos a inclinação do robot com a inclinação da Peca subjacente. Assim o robot alterava a sua inclinação para corresponder com essa Peca.

Por fim, reparamos que certas Jogadas resultavam na morte do robot. Assim, decidimos verificar se a Jogada feita ,pelas funções acima descritas, resultava em morte. Se resultasse o robot não fazia nada, se não fazia a Jogada.

Resultados

Por observação dos torneios que decorreram, reparamos que o robot funcionava e que muitas vezes ia pelo caminho mais curto, mas que nem sempre ia pelo caminho mais curto. Ou seja, tinha uma "visão" limitada. Devido a este facto o robot tendia a perder contra os outros Jogadores que conseguiam "ver" mais à frente.

Conclusão

Como resultado final, temos um robot que consegue “tomar” autonomamente decisões e que consegue navegar de forma eficiente até à chegada. No entanto, existe ainda margem para melhorar. O robot pode melhorar na decisão do melhor caminho.

Funções principais da Tarefa 6.

bot

Arguments

:: Int
O identificador do Jogador associado ao robot.

-> Estado
O Estado para o qual o robot deve tomar uma decisão.

-> Maybe Jogada
Uma possível Jogada a efetuar pelo robot.

Define um robot capaz de jogar autonomamente o jogo.

atritoPecasRaio1

Arguments

:: Int
A posição do robot.

-> Int
A Pista em que o robot se encontra.

-> Mapa
O Mapa a avaliar.

-> Maybe [Double]
A lista do atrito da próxima coluna de Pecas colineares.

caminhoMenosAtrito

Arguments

:: Int
A Pista em que o robot se encontra

-> Maybe [Double]
A lista do atrito da próxima coluna de Pecass colineares

-> Maybe Jogada
Uma possível Jogada a efetuar pelo robot .

mudaInclinacao

Arguments

:: Mapa
O Mapa a avaliar.

-> Jogador
O robot a ser testado

-> Maybe Jogada
Uma possível Jogada

proxJogadaMorto

Arguments

:: Maybe Jogada
A Jogada que se pretende testar.

-> Jogador
O robot que vai ser testado

-> Mapa
O Mapa a ser testado

-> Maybe Jogada
A Jogada avaliada