Tarefa6_2022li1g046

Description

Módulo para a realização da Tarefa 6 do projeto de LI1 em 2022/23.

Documentation

data Opcao

Opções que se encontram disponíveis no menu principal

Constructors

Jogar

Instrucoes

Sair

data Menu

Diferentes estados do jogo

Constructors

Opcoes Opcao

ModoJogo

PaginaInstrucao

VenceuJogo

type World = (Menu, Jogo, Images, Time, Pontuação)

Composição de um mundojogo e as suas propriedadeselementos

type Pontuação = Float

Tipo do elemento da pontuação

type Time = Float

Tipo do elemento de tempo

type Images = [Picture]

Tipo das imagens que compõem o jogo

win :: Display

Propriedades da janela do jogo

fps :: Int

frame rate que vai ser usado na função play

altura :: Float

altura da janela em pixeis

comp :: Float

comprimento da janela em pixeis

initialState :: Images -> World

estado incial do jogo

jogoInicial :: Jogo

jogo inicial que será utilizado para começar um novo jogo

drawOption :: String -> Picture

Função que desenha um certo texto

desenhaEstado :: World -> Picture

Função responsável por converter um estado do jogo para algo gráfico.

scoreboard :: Images -> Picture

Scoreboard que pode ser visível do lado direito durante o decorrer do jogo.

score :: Float -> Picture

Função que representa o score do jogador.

desenhaJogador :: Jogador -> Images -> Picture

Função responsável por converter um jogador e a sua posição para uma pictura.

desenhaTerreno :: Terreno -> Images -> Picture

Função que é utilizada para a conversão de um certo tipo de terreno numa picture.

desenhaObstaculo :: Obstaculo -> Images -> Picture

Função utilizada para converter um obstáculo para uma picture

desenhaObstaculos :: [Obstaculo] -> Float -> Images -> [Picture]

Função utilizada para converter uma lista de obstáculos para uma picture

desenhaObstaculosContrario :: [Obstaculo] -> Float -> Images -> [Picture]

Função auxiliar para desenhar os carros com uma rotação de 180 graus (utilizada quando a velocidade de um terreno do tipo estrada é negativa)

desenhaLinha :: (Terreno, [Obstaculo]) -> Float -> Images -> [Picture]

Função que desenha uma linha uma linha do mapa (converte para picture)

desenhaMapa :: Mapa -> Float -> Images -> [Picture]

Função que converte um dado para mapa para uma lista de pictures.

event :: Event -> World -> World

Função que reage a diversos tipos de eventos (utilização de teclas etc..)

step :: Float -> World -> World

Função que reage à passagem do tempo e que vai alterando o estado do jogo

checkPos :: Float -> World -> World

Função que verifica se o jogador se encontra na posição definida para que o mapa deslize.

seedRange :: Int -> Int

Função que converte uma seed que não esteja contida no intervalo (0,100) para um valor pertencente a este

checkTime :: Float -> World -> World

Função utilizada para atualizar o estado do jogo com o passar do tempo

exec :: IO ()

Função principal responsável pela execução e compilação de todo o jogo