Tarefa6_li1g081

Documentation

bot :: [String] -> Int -> Int -> Maybe Char

A Função bot recebe um Estado de Jogo e produz uma ação consoante esse mesmo Estado.

princ :: [String] -> Int -> Int -> Char

A Função princ é a função principal.

noSpiral :: [String] -> Int -> Char

A Função noSpiral é a Função Principal, antes da Espiral começar.

comand :: [String] -> (Int, Int) -> Int -> Int -> Char

A Função comand gera os Comandos do bot, quando este não está no raio de explosão de bombas.

chkNear :: [String] -> (Int, Int) -> Int -> Bool

A Função chkNear verifica se existe algum bloco destrutivel perto do jogador, caso exista, coloca uma bomba.

getComand :: [String] -> (Int, Int) -> Int -> Char

A Função getComand diz o comando que o bot irá executar, para este se aproximar de um bloco destrutível.

up :: [String] -> (Int, Int) -> Int -> Int -> Char

A Função up verifica se existem blocos na posição acima do bot.

right :: [String] -> (Int, Int) -> Int -> Int -> Char

A Função right verifica se existem blocos na posição à direita do bot.

down :: [String] -> (Int, Int) -> Int -> Int -> Char

A Função down verifica se existem blocos na posição abaixo do bot.

left :: [String] -> (Int, Int) -> Int -> Int -> Char

A Função left verifica se existem blocos na posição à esquerda do bot.

getExp :: [String] -> String -> Int -> [[String]]

A Função getExp obtem uma lista com elementos afetadas pelo raio da bomba.

bbDet :: [String] -> [String] -> String -> Int -> Char

A FunçãocbbDet é o Controlador Geral das funções que originam comandos quando existem bomas perto do bot.

plrCheck :: (Int, Int) -> [[String]] -> Bool

A Função plrCheck verifica se o bot se encontra dentro dos elementos afetados pelo raio de explosão de uma bomba.

plrCheckAux :: (Int, Int) -> [String] -> Bool

A Função plrCheckAux é uma função auxiliar da plrCheck.

moveChk :: [String] -> String -> String -> Char

A Função moveChk verifica a posição do bot face à bomba, e dá um comando com base nessa posição.

spotSchUp :: [String] -> (Int, Int) -> Int -> Char

A Função spotSchUp verifica se é possível mover para uma posição acima da atual.

spotSchRt :: [String] -> (Int, Int) -> Int -> Char

A Função spotSchRt verifica se é possível mover para uma posição à direita da atual.

spotSchDw :: [String] -> (Int, Int) -> Int -> Char

A Função spotSchDw verifica se é possível mover para uma posição abaixo da atual.

spotSchLf :: [String] -> (Int, Int) -> Int -> Char

A Função spotSchLf verifica se é possível mover para uma posição à esquerda da atual.

onPathChk :: [String] -> (Int, Int) -> Char -> Char

A Função onPathChk verifica se o lugar para onde o bot se pretende mover está dentro do raio de explosão de uma bomba, se sim ele este não muda de posição, se não tem permissão para se mover.

onPathChkAux :: [String] -> (Int, Int) -> Char -> [String] -> Int -> Char

A Função onPathChkAux é uma função auxiliar da onPathChk.

onPathChkAux2 :: [String] -> (Int, Int) -> Char -> [String] -> Int -> Char

A Função onPathChkAux2 é uma função auxiliar da onPathChkAux (Verifica se o bot se encontra no raio de explosão da bomba).

onPathChkAux3 :: [String] -> (Int, Int) -> [String] -> Int -> Bool

A Função onPathChkAux3 é uma função auxiliar da onPathChkAux2 (Verifica se o bot se encontrava inicialmente no raio de explosão da bomba).

bomblimit :: [String] -> (Int, Int) -> Int -> Char

A Função bomblimit cria um limite (raio 3 blocos) que impede que o bot fique preso entre duas bombas e/ou blocos.

bomblimitAux :: [String] -> [[String]] -> Bool

A Função bomblimitAux é uma função auxiliar da bomblimit.

getExp2 :: [String] -> (Int, Int) -> Int -> [[String]]

A Função getExp2 é uma função modificada da getExp para poder ser utilizada na função bomblimit.

yesSpiral :: [String] -> Int -> Char

A Função yesSpiral é a Função Principal depois da espiral ter começado.

goCenter :: [String] -> (Int, Int) -> Int -> Char

A Função goCenter tenta redirecionar o bot para o centro do mapa, para este não ser eliminado pela espiral.

posCount :: Int -> Int

A Função posCount conta o número de vezes que é preciso subtrair 2 à dimensão do mapa, para obter o valor da coordenada central.

spchk :: String -> Int -> String

Função definida numa Tarefa Anterior.

posGet :: String -> (Int, Int)

Função definida numa Tarefa Anterior.

pll :: [String] -> Int -> String

Função definida numa Tarefa Anterior.

charChk :: [String] -> (Int, Int) -> Char

Função definida numa Tarefa Anterior.

stringSch :: [String] -> Int -> String

Função definida numa Tarefa Anterior.

posSch :: String -> Int -> Char

Função definida numa Tarefa Anterior.

charGet :: [String] -> Int -> (Int, Int) -> Int -> [[String]]

Função definida numa Tarefa Anterior.

spup :: [String] -> Int -> (Int, Int) -> Int -> [String]

Função definida numa Tarefa Anterior.

spdw :: [String] -> Int -> (Int, Int) -> Int -> [String]

Função definida numa Tarefa Anterior.

splf :: [String] -> Int -> (Int, Int) -> Int -> [String]

Função definida numa Tarefa Anterior.

sprt :: [String] -> Int -> (Int, Int) -> Int -> [String]

Função definida numa Tarefa Anterior.

pathCheckSp :: [[String]] -> [[String]]

Função definida numa Tarefa Anterior.

pathCheckSpAux :: [String] -> [String]

Função definida numa Tarefa Anterior.

bombGet :: [String] -> [String]

Função definida numa Tarefa Anterior.

tkspe2 :: String -> Int -> Int -> String

Função definida numa Tarefa Anterior.