Tarefa6_li1g104

Documentation

bot :: [String] -> Int -> Int -> Maybe Char

Função principal do nosso bot, recebe o estado de jogo actual, o nº do jogador e o tempo de jogo e devolve uma jogada, B(plantar bomba) ou mover o jogador(U,D,L,R)

type Direccao = Char

type Mapa = [String]

type Grelha = [String]

type Elementos = [String]

type PUbomba = String

type PUflame = String

type Bomba = String

type Bombas = [Bomba]

type Raio = Int

type Jogador = String

type Linha = Int

type Coluna = Int

type Coordenada = (Coluna, Linha)

type Tempo = Int

type Dimensao = Int

plantarbomba :: Coordenada -> Raio -> Mapa -> Bool

verutil :: Coordenada -> Raio -> Mapa -> Bool

Função que verifica se colocar uma bomba pode eventualmente destruir algum tijolo

verefeito :: Grelha -> Coordenada -> Raio -> Bool

Função que verifica se a explosão duma bomba causa algum efeito

verfeitoaux :: Grelha -> Coordenada -> Raio -> Int -> Direccao -> Bool

Função que verifica se a explosão duma bomba causa algum efeito numa determinada direcção(auxiliar de verefeito)

modoseguro :: Grelha -> Bombas -> Jogador -> Bool

Função que verifica se o jogador está livre de explosões na posição actual

modorisco :: Grelha -> Bombas -> Jogador -> Bool

versegpo :: Grelha -> Bombas -> Coordenada -> Bool

Função que verifica se uma posição esta segura

bombaslincol :: [Coordenada] -> Coordenada -> [Raio] -> Bool

Função que calcula se uma determinada posição está em risco de ocorrer uma explosão

distancia :: Coordenada -> Coordenada -> Int

Funçao que calcula a distancia entre 2 posições

raiobombas :: Bombas -> [Raio]

Função que recebe a lista de bombas do estado actual e devolve a lista dos seus raios

moverjogador :: Grelha -> Coordenada -> Coordenada -> Bombas -> Maybe Char

Função que decida para onde se deve mover o jogador(ou não efectuar qualquer movimento)

posipumaisprox :: Coordenada -> [Coordenada] -> (Coordenada, Int) -> Coordenada

Função que descobre qual a posição do PowerUp visível mais proximo do jogador, caso exista a menos de 10 movimentos, ou o centro do mapa

procurarposseg :: Grelha -> Jogador -> Bombas -> Maybe Char

Função que procura uma posição segura com apenas 3 comandos

somapos :: Coordenada -> Coordenada -> Coordenada

Função que dadas 2 coordenadas devolve a coordenada resultante da soma das duas

metdim :: String -> Int

Função que dada a 1ª linha da grelha devolve um inteiro correspondente a metade da dimensão(utíl para descobrir a posição central do mapa)

verposgrelha :: Grelha -> Coordenada -> Coordenada -> Maybe Char

Função que devolve a ocupação duma determinada posição da grelha

verjogcolbom :: Int -> [String] -> Int

Função que conta quantas bombas um jogador já tem colocadas no mapa -}

tupler :: String -> Coordenada

Função para adquirir a posicao no mapa de um PowerUP, jogador ou um bomba colocada

tuplesdados :: [String] -> [Coordenada]

Função que utilizamos para transformar a lista de bombas ou flames na lista das suas coordenadas

grelha :: [String] -> [String]

Função que recebe o input e devolve apenas a grelha de jogo

bombas :: [String] -> [String]

Função que recebe o input e devolve a lista de bombas

flames :: [String] -> [String]

Função que recebe o input e devolve a lista de flames

bombacol :: [String] -> [String]

Função que recebe o input e devolve a lista das bombas colocadas

jogadores :: [String] -> [String]

Função que recebe o input e devolve a lista jogadores

jogador :: Char -> [String] -> String

Função que pega na lista de jogadores e devolve apenas um jogador

verposigrelha :: Grelha -> Coordenada -> Coordenada -> Char

Função que devolve a ocupação duma determinada posição da grelha