Types

Documentation

data State

Constructors

State

Fields

maze :: Maze

playersState :: [Player]

level :: Int

type Maze = [Corridor]

type Corridor = [Piece]

data Piece

Constructors

Food FoodType

PacPlayer Player

Empty

Wall

data Player

Constructors

Pacman PacState

Ghost GhoState

data Orientation

Constructors

L

R

U

D

Null

data PacState

Constructors

PacState

Fields

pacState :: PlayerState

timeMega :: Double

openClosed :: Mouth

pacmanMode :: PacMode

data GhoState

Constructors

GhoState

Fields

ghostState :: PlayerState

ghostMode :: GhostMode

type Coords = (Int, Int)

type PlayerState = (Int, Coords, Double, Orientation, Int, Int)

data Mouth

Constructors

Open

Closed

data PacMode

Constructors

Dying

Mega

Normal

data GhostMode

Constructors

Dead

Alive

data FoodType

Constructors

Big

Little

data Color

Constructors

Blue

Green

Purple

Red

Yellow

None

data Play

Constructors

Move Int Orientation

type Instructions = [Instruction]

data Instruction

Constructors

Instruct [(Int, Piece)]

Repeat Int

coloredString :: String -> Color -> String

placePlayersOnMap :: [Player] -> Maze -> Maze

printMaze :: Maze -> String

printPlayerStats :: Player -> String

getPlayerID :: Player -> Int

getPlayerPoints :: Player -> Int

setPlayerCoords :: Player -> Coords -> Player

getPieceOrientation :: Piece -> Orientation

getPacmanMode :: Player -> PacMode

getPlayerState :: Player -> PlayerState

getPlayerOrientation :: Player -> Orientation

replaceElemInMaze :: Coords -> Piece -> Maze -> Maze

replaceNElem :: Int -> a -> [a] -> [a]

Funções adicionadas

getPlayerCoords :: Player -> Coords

Função que dá as coordenadas do jogador.

getPlayerLives :: Player -> Int

Função que dá o número de vidas do jogador.

getPlayersCoords :: [Player] -> [Coords]

Função que devolve a lista de coordenadas duma lista de players.

deadGhosts :: [Player] -> [Player]

Função que junta os players que são dead ghosts.

ghostsDCoords :: State -> [Coords]

Função que dá as coordenadas dos Dead Ghosts que estão num determinado State.

aliveGhosts :: [Player] -> [Player]

Função que junta os players que são Alive Ghosts.

ghostsAlCoords :: State -> [Coords]

Função que dá as coordenadas dos ghosts que estão num determinado State.

isDeadGhost :: Player -> Bool

Função que verifica se um jogador é um Dead Ghost.

areDeadGhost :: [Player] -> Bool

Função que verifica se um conjunto de jogadores tem um Dead Ghost.

midCoord

Arguments

:: Maze
Labirinto.

-> Coords
Coordenadas da localização do interior da casa dos fantasmas no respetivo labirinto.

Devolve a coordenada do interior da casa dos fantasmas relativamente a um dado labirinto, usada para mandar os fantasmas comidos para lá.

ghostsCoords :: [Player] -> [Coords]

Dada uma lista de Players, devolve a lista das coordenadas de todos os fantasmas.

thePacman :: [Player] -> Player

Dada uma lista de Players devolve o Player que é Pacman.

isPac :: Player -> Bool

Se o Player recebido for Pacman, devolve True, caso contrário devolve False.

isGhost :: Player -> Bool

Se o Player recebido for Ghost, devolve True, caso contrário devolve False.

idPlayer :: [Player] -> Int -> Player

Função que de um conjunto de Players, devolve o Player com o ID pedido.

ghostPlayers :: [Player] -> [Player]

Função que devolve a lista de Ghosts que estão numa lista de Players.

pacCoords :: State -> Coords

Função que diz em que coordenadas o Pacman se encontra no labirinto.

coo1D :: Coords -> Coords

Função que devolve a coordenada imediatamente abaixo da coordenada recebida.

getPieceCoords :: Piece -> State -> [Coords]

Dado uma Piece e um State, vai utilizar a função auxiliar, tendo como acumulador inicial (0,0), para determinar a lista das coordenadas da Piece dada.

auxGetPcCoords :: Piece -> Maze -> Coords -> [Coords]

Função auxiliar da função anterior que dado uma Piece, um Maze e Coords (acumulador, que será (0,0) para que determine corretamente as coordenadas do labirinto), devolvendo a lista de coordenadas onde a respetiva peça se encontra no labirinto.