Ficha3

Description

Este módulo contém as definições Haskell, para o cálculo de funções recursivas sobre listas e sobre inteiros, pedidas na Ficha 3 de Laboratórios de Informática I.

Documentation

remove

Arguments

:: [String]

-> Char

-> [String]
Exercício 1

A função remove recebe uma lista de strings e remove todas as strings iniciadas por um dado caractere.

Notas:

Expand

Esta função é sensível a maiúsculas e minúsculas;

Esta função funciona para caracteres numéricos;

Esta função funciona para caracteres especiais;

Esta função funciona para listas vazias.

Exemplos de utilização:

Expand

>>> remove ["Ola", "Amor", "Comida"] 'A'
["Ola", "Comida"]
>>> remove ["Ola", "Amor", "Comida"] 'a'
["Ola", "Amor", "Comida"]
>>> remove [] 'a'
[]

Propriedades:

Expand

remove [] _ = []

addTuple :: [(Int, Int)] -> Int -> [(Int, Int)]

A função addTuple recebe uma lista de pares de inteiros e adiciona um valor dado à primeira componente de cada par.

Exemplos de utilização:

Expand

>>> addTuple [(1, 2), (-1, 1), (0, 0)] 5
[(6, 2), (4, 1), (5, 0)]
>>> addTuple [(1, 2), (-1, 1), (0, 0)] (-1)
[(0, 2), (-2, 1), (-1, 0)]
>>> addTuple [(1, 2), (-1, 1), (0, 0)] 0
[(1, 2), (-1, 1), (0, 0)]
>>> addTuple [] 9
[]

Propriedades:

Expand

addTuple [] _ = []

maxSndTuple :: [(Int, Int)] -> Int

A função maxSndTuple recebe uma lista, não vazia, de pares de inteiros e calcula qual o maior valor da segunda componente.

Exemplos de utilização:

Expand

>>> maxSndTuple [(1, 2), (-1, 1), (0, 0)]
2
>>> maxSndTuple [(1, 2), (-1, 1), (0, 3)]
3

Propriedades:

Expand

maxSndTuple [(a, b)] = b

type Nome = String

O type Nome é sinónimo de uma string (representativa de uma identidade).

type Coordenada = (Double, Double)

O type Coordenada é sinónimo de um par de doubles (representativo de um conjunto de coordenadas cartesianas).

data Movimento

O data type Movimento é a definição de um novo tipo de dados (representativo de um movimento segundo os pontos cardeais).

Constructors

N
(North) --> y + 1

S
(South) --> y - 1

E
(East) --> x + 1

W
(West) --> x - 1

type Movimentos = [Movimento]

O type Movimentos é sinónimo de uma lista do tipo Movimento (representativa de uma sequência finita de movimentos).

data PosicaoPessoa

O data type PosicaoPessoa é a definição de um novo tipo de dados (representativo de um conjunto de coordenadas cartesianas referentes a uma identidade).

Constructors

Pos Nome Coordenada

auxCoord :: PosicaoPessoa -> Movimento -> PosicaoPessoa

A função auxCoord recebe um data type PosicaoPessoa e um Movimento e calcula novas coordenadas.

Exemplos de utilização:

Expand

>>> auxCoord (Pos "Afonso" (0, 0)) N
(Pos "Afonso" (0, 1))
>>> auxCoord (Pos "Afonso" (0, 0)) S
(Pos "Afonso" (0, -1))
>>> auxCoord (Pos "Afonso" (0, 0)) E
(Pos "Afonso" (1, 0))
>>> auxCoord (Pos "Afonso" (0, 0)) W
(Pos "Afonso" (-1, 0))

posicoesM :: [PosicaoPessoa] -> Movimento -> [PosicaoPessoa]

A função posicoesM recebe uma lista de posições de pessoas e um movimento e atualiza as coordenadas de todas as posições.

Notas:

Expand

Esta função recorre à anterioremente definida auxCoord.

Assume-se que o movimento corresponde a uma unidade do referencial.

Exemplos de utilização:

Expand

>>> posicoesM [(Pos "Afonso" (0, 0)), (Pos "Bernardo" (5, 3))] N
[(Pos "Afonso" (0, 1)), (Pos "Bernardo" (5, 4))]
>>> posicoesM [(Pos "Afonso" (0, 0)), (Pos "Bernardo" (5, 3))] S
[(Pos "Afonso" (0, -1)), (Pos "Bernardo" (5, 2))]
>>> posicoesM [(Pos "Afonso" (0, 0)), (Pos "Bernardo" (5, 3))] E
[(Pos "Afonso" (1, 0)), (Pos "Bernardo" (6, 3))]
>>> posicoesM [(Pos "Afonso" (0, 0)), (Pos "Bernardo" (5, 3))] W
[(Pos "Afonso" (-1, 0)), (Pos "Bernardo" (4, 3))]

Propriedades:

Expand

posicoesM [] _ = []

posicao :: PosicaoPessoa -> Movimentos -> PosicaoPessoa

A função posicao recebe um data type PosicaoPessoa e uma lista de movimentos e calcula a posição da pessoa depois de executar todos os movimentos.

Notas:

Expand

Esta função recorre à anterioremente definida auxCoord.

Assume-se que o movimento corresponde a uma unidade do referencial.

Exemplos de utilização:

Expand

>>> posicao (Pos "Afonso" (0, 0)) [S, S, W]
Pos "Afonso" (-1, -2)
>>> posicao (Pos "Afonso" (0, 0)) [N, N, E, S]
Pos "Afonso" (1, 1)

Propriedades:

Expand

posicao pos [] = pos

posicoesMs :: [PosicaoPessoa] -> Movimentos -> [PosicaoPessoa]

A função posicoesMs recebe uma lista de posições de pessoas e uma lista de movimentos e calcula a posição de cada pessoa depois de executar todos os movimentos.

Notas:

Expand

Esta função recorre à anterioremente definida auxCoord.

Esta função recorre à anterioremente definida posicao.

Assume-se que o movimento corresponde a uma unidade do referencial.

Exemplos de utilização:

Expand

>>> posicoesMs [(Pos "Afonso" (0, 0)), (Pos "Bernardo" (5, -5))] [S, S, W]
[(Pos "Afonso" (-1, -2)), (Pos "Bernardo" (4, -7))]
>>> posicao [(Pos "Afonso" (0, 0)), (Pos "Bernardo" (5, -5))] [N, N, E, S]
[(Pos "Afonso" (1, 1)), (Pos "Bernardo" (6, -4))]

Propriedades:

Expand

posicoesMs [] _ = []

pessoaNorte :: [PosicaoPessoa] -> Maybe PosicaoPessoa

A função pessoaNorte recebe uma lista de posições de pessoas e indica qual a posição mais a Norte.

Notas:

Expand

Assume-se que a pessoa mais a Norte é aquela cuja ordenada é maior.

Exemplos de utilização:

Expand

>>> pessoaNorte [(Pos "Afonso" (0, 0)), (Pos "Bernardo" (5, -5))]
Just (Pos "Afonso" (0, 0))
>>> pessoaNorte [(Pos "Afonso" (0, 0)), (Pos "Bernardo" (5, 5))]
Just (Pos "Bernardo" (5, 5))

Propriedades:

Expand

pessoaNorte [] = Nothing

pessoaNorte [a] = Just a

pessoasNorte :: [PosicaoPessoa] -> [String]

A função pessoasNorte recebe uma lista de posições de pessoas e indica quais as pessoas mais a Norte. Ou seja, uma lista com o nome de todas as pessoas, exceto a(s) que se encontre(m) mais a Sul.

Notas:

Expand

Assume-se que a pessoa mais a Sul é aquela cuja ordenada é menor.

Caso exista apenas um elemento na lista, o resultado terá esse elemento.

Exemplos de utilização:

Expand

>>> pessoasNorte [(Pos "Afonso" (0, 0)), (Pos "Bernardo" (5, -5)), (Pos "Carlos" (5, -5))]
["Afonso"]
>>> pessoasNorte [(Pos "Afonso" (0, 0)), (Pos "Bernardo" (5, 5)), (Pos "Carlos" (5, -5))]
["Bernardo", "Afonso"]

Propriedades:

Expand

pessoasNorte [] = []

pessoasNorte [Pos name (x, y)] = [name]