Main

Description

Introdução Tarefa 5:

Nesta tarefa tinhamos como objetivo desenvolver a componente gráfica do nosso jogo utilizando a biblioteca Gloss.

É importante reconhecer que todas as tarefas foram desenvolvidas de acordo com a lógica da implementação do jogo.

Nesta tarefa tinhamos várias opções de desenhar e desenvolver a parte grafica do nosso jogo.

Estratégias utilizadas:

Começamos por tentar desenhar um mapa de jogo fixo , tendo primeiro criadado a função linhamapa que cria uma linha do mapa e a função mapa que usa a linhamapa para desenhar o mapa de jogo.

Despois criamos as funções desenhadisp e desenhajog e juntamos as 3 funções na funcao desenhaestado3 para conseguirmos desenhar todo um estado de jogo.

A partir daqui definimos as funções estadoinicialg , onde definimos o estado inicial do nosso jogo, reageEventoB que aplica as jogadas de cada jogador a um estado (função jogada definida na tarefa 2) e a função reageTempog que reage à passagem do tempo (usando a funçâo ticks definida na tarefa4).

Uma vez que já conseguiamos correr o nosso jogo começamos a criar os menus. Primeiro criamos um menu que desenhava um editor de mapas (função da Tarefa 1) que permite criar um mapa e usamos a função scale1 para redimensionar e posicionar o mapa à medida que este ficava maior. De seguida fomos redefinindo o nosso estadogloss de forma a poder receber mais informações relativas ao tipo de jogo pretendido pelo utilizador , tais como o número de jogador,o tipode jogo , e os mapas (definidos na tarefa 2).

Para concluir implementamos no nosso jogo a funçâo bot definida na Tarefa 6

Conclusão:

Ao longo desta tarefa sentimos algumas dificuldades principalmente por nunca termos trabalhado com gloss, mas conseguimos cria uma boa implementação gráfica do nosso jogo, criar menos varias funcionalidades.

Documentation

main :: IO ()

Função principal da Tarefa 5.

NB: Esta Tarefa é completamente livre. Deve utilizar a biblioteca gloss para animar o jogo, e reutilizar de forma completa as funções das tarefas anteriores.

listajogadores :: Int -> Mapa -> [Jogador]

funcao que devolve a lista de jogadores posicionados em cada canto mapa

mapaI :: Int -> Instrucoes

mapas guardados

imagem :: [Picture] -> Int -> Picture

escolhe imagem

estadoinicial :: Estado

dá o estado inicial

desenhaestado1 :: [Picture] -> Picture

desenha menu inicial

desenhaestado7 :: [Picture] -> Picture

desenha menu inicial

desenhaestado4 :: [Picture] -> Picture

desenha menu numero de jogadores

desenhaestado5 :: [Picture] -> Int -> Picture

desenha menu mapas

desenhaestado6 :: [[Picture]] -> Estado -> Picture

desenha vencedor

desenhaestado2 :: [[Picture]] -> Editor -> Picture

desenha criador de mapas

desenhaestadog3 :: Estadogloss -> Picture

desenha jogo

desenhaestado3 :: Estado -> [[Picture]] -> Picture

desenha jogo

mapa :: Mapa -> [Picture] -> Float -> [Picture]

desenha mapa

linhamapa :: [Peca] -> [Picture] -> Float -> Float -> [Picture]

desenha linha do mapa

desenhajog :: [Jogador] -> [Picture] -> [Picture]

desenha jogadores

rodaimagem :: Picture -> Direcao -> Picture

roda imagem de acordo com a sua direcao

posjog :: Posicao -> Picture -> Picture

posiciona o jogador

desenhadisp :: Estado -> [Picture] -> [Picture]

desenha disparos

expandelaser :: Posicao -> Direcao -> Mapa -> Picture -> [Picture]

expande o laser ate encontrar um bloco indestrutivel

desenhatabela :: [Jogador] -> [Picture] -> Float -> [Picture]

posiciona numeros na tabela de vidas, lasers e choques

posicoes :: Jogador -> [Picture] -> Float -> [Picture]

posiciona numero de vidas, lasers e choques

tempo :: Int -> Picture

desenha tempo

scale1 :: Picture -> Mapa -> Picture

funcao que posiciona e define o tamanho do mapa

reageEventomenu :: Event -> Estadogloss -> Estadogloss

funcao que define os controlos do menu inicial

reageEventotipojogo :: Event -> Estadogloss -> Estadogloss

funcao que define os controlos para voltar ao menu inicial no final do jogo

reageEventojogadores :: Event -> Estadogloss -> Estadogloss

funcao que define o numero de jogadores

reageEventoMapa :: Event -> Estadogloss -> Estadogloss

funcao que define o mapa de jogo

reageEventoVencedor :: Event -> Estadogloss -> Estadogloss

funcao que define os controlos para voltar ao menu inicial no final do jogo

reageEventoEditor :: Event -> Estadogloss -> Estadogloss

funcao que define os controlos do criador de mapas

reageEventojogo :: Event -> Estadogloss -> Estadogloss

aplica afuncao reageEvento para jogar multiplayer e a funcao reageEventoB para jogar contra bots

verificavencedor :: [Jogador] -> [Jogador]

verifica se existe apenas um jogador com vida dirente de 0

reageEvento :: Event -> Estadogloss -> Estadogloss

funcao que controla as jogadas de cada jogador

reageEventoB :: Event -> Estadogloss -> Estadogloss

funcao para controlar bots e um jogador

retiraevento :: Key -> [Key] -> [Key]

retira evento a uma lista de eventos

transjogada :: Maybe Jogada -> Estado -> Int -> Estado

aplica jogado do bot

reageTempo :: Float -> Estadogloss -> Estadogloss

reague à passagem do tempo alterando o estado

event1 :: [Key] -> Estado -> Estado

aplica a jogada correspondente a cada key

type Estadogloss = (Estado, [[Picture]], (Editor, Instrucoes), Int, Int, Int, Int, [Key], Int)

define um novo tipo de estado gloss ( tipo de estado | nº jogadores | mapas | tipo de jogo )

estadoinicialg :: [[Picture]] -> Estadogloss

estado inicial do jogo

desenhaestadog :: Estadogloss -> Picture

desenha estado do jogo

reageEventog :: Event -> Estadogloss -> Estadogloss

reage a um evento

reageTempog :: Float -> Estadogloss -> Estadogloss

reage à passagem do tempo

fr :: Int

frame rate

dm :: Display

display