Main

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2021/22.

Bibliografia

Para a realização desta tarefa recorri ao material disponibilizado nas aulas de Laboratórios de Informática 1 bem como a vídeos educativos sobre o funcionamento do Gloss.

Documentation

type EstadoGloss = (Menu, PecasPic, PictureJogador, PictureMenu)

O _type_ EstadoGloss contém informação sobre o estado do jogo à medida que este decorre.

type PecasPic = [(Peca, (Picture, (Float, Float)))]

O _type_ PecasPic cooresponde a uma lista de peças com a sua determinada imagem e coordenadas (as coordenadas são convertidas através de uma função para float, de modo a fazer o mapeamento entre as peças, a sua imagem e as suas coordenadas.

type PictureJogador = [Picture]

O type PictureJogador e o type PecasPic correspondem a uma lista de Pictures.

type PictureMenu = [Picture]

data Opcoes

Criamos data Opcoes e data Menu de modo a criar novos tipos que facilitam o controlo do menu e do jogo através dos eventos de teclado.

Constructors

Jogar

Sair

data Menu

Constructors

Controlador Opcoes

ModoJogo Jogo

YouWon

lvl1 :: Mapa

altura :: Float

As funções altura e comprimento indicam a respetiva altura e comprimento a que o mapa se encontra.

comprimento :: Float

ld :: Float

A função ld indica a distância entre cada célula da matriz.

estadoInicial1 :: Jogo

A função estadoInicial1 indica as condições iniciais do jogo.

estadoGlossInicial :: PecasPic -> PictureJogador -> PictureMenu -> EstadoGloss

Para executar esta parte do projeto foi necessário criar novos _types_ e _data_ capazes de conter a informação necessária de modo a aplicar as facilmente as capacidades gráficas que a biblioteca Gloss disponibiliza.

A função estadoGlossInicial recebe informação sobre as peças e as suas posições bem como as suas imagens, a imagem do jogador e as imagens utilizadas no menu, devolvendo assim informação sobre o estado inicial do jogo.

getMapa :: Jogo -> Mapa

Através da função getMapa, tendo o Jogo, é possível obter o mapa do jogo.

getJogador :: Jogo -> Jogador

Através da função getJogador, tendo o Jogo, é possível obter o Jogador, que contém informações sobre as suas coordenadas, direção e se carrega ou não caixa.

fr :: Int

dm :: Display

A função dm retribui a dimensão da janela do jogo que, neste caso, optamos por utilizar a funcionalidade FullScreen.

desenhaEstado :: EstadoGloss -> Picture

desenhaMapa :: Float -> Float -> Mapa -> PecasPic -> [Picture]

A função desenhaMapa recebe dois floats, recebe um mapa e recebe o _type_ PecasPic que através das funções auxiliares desenhaLinha e desenhaPeca faz o mapeamento entre a peça e as suas coordenadas com a sua respetiva imagem, retribuindo assim uma lista de Pictures. A função representada por linha recorre à função auxiliar desenhaLinha que desenha a _head_ do mapa. Por sua vez, a função representada por restomapa recorre recursivamente à funcão desenhaMapa. Assim, criando uma lista através da concatenação das funções linha e restomapa, é possível desenhar o resto da mapa.

Dificuldade: - Inicialmente senti alguma dificuldade na execução desta parte da tarefa, contudo, através de vídeos associados à utilização do Gloss, consegui perceber a sua lógica e aplicação.

desenhaLinha :: Float -> Float -> [Peca] -> PecasPic -> [Picture]

A função desenhaLinha atua do mesmo modo que a função desenhaMapa, no entanto em vez de receber um mapa recebe uma lista de peças. A função representada por peca recorre à função auxiliar desenhaPeca que desenha a _head_ da lista de peças. Por sua vez, a função representada por restopeca recorre recursivamente à funcão desenhaLinha. Assim, criando uma lista através da junção de ambas as funções, é possível desenhar o resto das linhas.

desenhaPeca :: Float -> Float -> Peca -> PecasPic -> Picture

A função desenhaPeca é semelhante às anteriores, sendo que a diferença reside que em vez de um mapa ou uma lista de peças, esta função recebe para além de dois floats uma só peça. Esta função transpõe para a posição x e y e imagem da peça que apresenta as coordenadas (x,y). As funções auxiliares floatX e floatY transformam as coordenadas (números inteiros) em floats, uma vez que as imagens podem precisar de translações individuais.

desenhaJog :: Jogador -> PictureJogador -> Picture

A função desenhaJog recebe informação sobre o jogador e uma lista de imagens (que variam na direção e se este carrega ou não caixa), reproduzindo assim a Picture do jogador dadas a sua posição, direção e se carrega ou não caixa. As funções auxiliares jogX e jogY transformam as coordenadas (números inteiros) em floats uma vez que as imagens podem precisar de translações individuais.

Exemplo: Se o jogador estiver voltado para este e não carregar caixa, então é dado _Translate_ da imagem do jogador (que se encontra na primeira posição da lista dada pelo _type_ PictureJogador [picj !! 0]) para as posições jogX e jogY.

jogX :: Int -> Float

jogY :: Int -> Float

reageEvento :: Event -> EstadoGloss -> EstadoGloss

A função reageEvento reage a eventos do teclado, isto é, dadas informações sobre o jogo (recebe _EstadoGloss_) esta função aplica determinadas ações quando premida uma determinada tecla.

Exemplo: Quando estamos em _ModoJogo_, se à sua frente o personagem encontrar uma obstáculo e premirmos a tecla Up, o personagem executa o movimento Trepar. O mesmo se aplica aos restantes movimentos quando estamos em _ModoJogo_, variando apenas as teclas a premir dependendo da ação e da possibilidade de esta se realizar.

Quando estamos no Menu, apenas utilizamos as teclas Up e Down para selecionar Jogar ou Sair.

Quando ganhamos o jogo, ao premir a tecla Enter após aparecer "You Won!", retornamos ao Menu, podendo escolher se queremos voltar a jogar ou sair do jogo.

reageTempo :: Float -> EstadoGloss -> EstadoGloss

vazio :: Picture

Esta função torna os _Vazio_ "transparentes". Aplicamo-la de modo a que o jogador não ficasse coberto por eles.

desenhaMenu :: EstadoGloss -> Picture

A função desenhaMenu é responsável por retribuir uma Picture quando dado um _EstadoGloss_, ou seja, quando tem informações sobre o estado do jogo.

Exemplo: Quando nos encontrámos no controlador Jogar, a função retribui, neste caso a primeira (background do menu), a quarta(botão Play) e a terceira (botão Exit) imagens da lista de imagens definada no _type_ _PictureMenu_. Quando nos encontrámos no controlador Sair a função retribui o mesmo que no controlador Jogar pois encontram-se no mesmo estado. Por sua vez, quando ganhamos o jogo, a função retribui apenas a primeira e quinta imagens.

De modo a ajustar as imagens, em alguns casos, foi necessário manipular a escala das mesmas.

Relativamente às imagens utilizadas tanto para o menu como para o próprio jogo, estas foram desenhadas e pensadas pelo grupo. Para dar load nas imagens recorremos a um conversor de png para bmp online

ganhaJogo :: Jogo -> Bool

A função ganhaJogo, através de função auxiliar coorPorta retribui um bool que indica se o jogador apresenta as mesmas coordenadas que a porta. Implementamos esta função pois é necessária para construir o caso em que se ganha o jogo, isto é, em que o jogador atinge a porta do mapa.

coorPorta :: [(Peca, Coordenadas)] -> Coordenadas

main :: IO ()

A função main recebe todas as informações do jogo, sendo ela a base do mesmo. Esta função recebe diversas informações tais como, a dimensão do display, a cor de fundo, a função estadoGlossInicial, a função desenhaEstado, a função reageEvento e a função reageTempo pelo que, através do conjunto destes dados, devolve então o nosso jogo. Para além disso, todas as imagens mencionadas nas funções anteriores têm de ser carregadas na main, através da funcionalidade _loadBMP_.