Main

Documentation

avanca :: [String] -> Int -> [String]

A função avanca devolve o estado de jogo após as alterações nesse instante

tiraSegundo :: [String] -> [String]

A função tiraSegundo retira 1 segundo ao tempo restante de cada bomba

verificaTempo :: [String] -> [String] -> [String]

A função verificaTempo verifica as bombas que têm 0 segundos restantes para explodir (vão explodir nesse instante)

resultadoBomba :: [String] -> (Int, Int) -> Int -> Int -> [String]

A função resultadoBomba recebe o mapa, a coordenada da explosão, o seu raio de explosão, e um acumulador e aplica as funções retiraBomba e retiraJogadores na coordenada da explosão e a função resultadoExplosão para cada direção

resultadoExplosao :: [String] -> (Int, Int) -> Int -> (Int, Int) -> Int -> [String]

A função resultadoExplosao recebe o mapa, a coordenada que a função vai afetar, o raio da explosão, o vetor da direção, e um acumulador e através de condições altera o estado de jogo

reduzContador :: [String] -> (Int, Int) -> [String]

A função reduzContador reduz o tempo restante das bombas atingidas por uma explosão para 1

daPosicao :: [String] -> (Int, Int) -> Char

A função daPosicao devolve o caracter que se encontra no mapa numa dada coordenada

retirarTijolo :: [String] -> (Int, Int) -> Int -> [String]

A função retirarTijolo remove o tijolo (?) atingido pela explosão

sePwrUp :: [String] -> (Int, Int) -> Bool

A função sePwrUp verifica se existe no estado de jogo um power up com as coordenadas dadas e devolve um Bool

seBomba :: [String] -> (Int, Int) -> Bool

A função seBomba verifica se existe no estado de jogo uma bomba com as coordenadas dadas e devolve um Bool

retiraBomba :: [String] -> (Int, Int) -> [String]

A função retiraBomba remove do estado de jogo a bomba que explode

retiraJogadores :: [String] -> (Int, Int) -> [String]

A função retiraJogadores remove do estado de jogo um jogador atingido pela explosão da bomba

tiraPwrUp :: [String] -> (Int, Int) -> [String]

A função tiraPwrUp remove do estado de jogo um powerup atingido pela explosão da bomba

espiral :: [String] -> (Int, Int) -> (Int, Int) -> Int -> Int -> [String]

A função espiral recebe o mapa, a posição inicial (1,1) da espiral, um vetor de movimento, um acumulador e o numero de blocos que vão ser substituidos

veDim :: [String] -> Int

A função veDim dá a dimensão do mapa

verificaBorda :: Int -> (Int, Int) -> (Int, Int) -> Int -> ((Int, Int), Int)

A função verificaBorda verifica se os blocos devem ser começado a ser substituidos noutra direção

mataTudo :: [String] -> [String] -> [String]

A função mataTudo remove os powerups e os jogadores do estado de jogo

mudaPos :: [String] -> (Int, Int) -> Char -> Int -> [String]

A função mudaPos recebe o mapa, a coordenada que vai ser substituida, o caracter que ser utilizado e encontra a linha correta do mapa e aplica a função mudaPosx a essa linha

mudaPosx :: String -> Int -> Char -> Int -> String

A função mudaPos recebe a linha do mapa, a coordenada que vai ser substituida, o caracter que ser utilizado e substituir esse caracter

main :: IO ()

Função que imprime o mapa