Main

Description

Modulo : Main

Descrição : Módulo que testa a validade de um dado ficheiro

Este programa verifica se um ficheiro é válido para representar um nível de LightBot. Como resultado desta verificação o programa devolve os valores <OK> quando o ficheiro é válido; <num> em que num é a primeira linha em que o ficheiro diverge das normas especificadas.

Documentation

type Tabuleiro = [String]

Tipo usado para representar tabuleiros. Cada linha do tabuleiro é composto por um conjunto de carateres.

type Comandos = String

Tipo usado para representar uma linha de comandos. Cada comando é associado a um carater.

type Posicao = String

Tipo usado para representar a posição do Robot. Uma posição válida deverá apresentar a seguinte estrutura @x y ori@.

main :: IO ()

main lê um ficheiro e inicializa a sua verificação, imprimindo o resultado final.

checkContent :: [String] -> Int

checkContent recebe um ficheiro, e verifica se a sua estrutura é válida para um nível de LightBot. Uma estrutura válida obdece às seguintes normas:

tem pelo menos 3 linhas;

pode ser dividida em três partes (cada uma delas válida) de acordo com o seu conteúdo: tabuleiro; posição; comandos;

não ter linhas após a secção que define os comandos;

Como resultado é devolvido um Int cujo valor varia entre -1 caso o ficheiro seja válido e erro em que erro é a primeira linha em que o ficheiro não cumpre as normas.

checkTab :: Tabuleiro -> (Int, Int) -> Int

checkTab verifica se a estrutura do Tabuleiro recebido está dentro das normas. Para isso faz uso de (Int,Int) que contém informação sobre as linhase e as colunas. O tabuleiro deve conter a seguinte estrutura:

o comprimento de todas as linhas é igual;

apenas contém letras.

Como resultado é devolvido um Int cujo valor varia entre -1 caso o tabuleiro esteja devidamente construido e n em que n é a primeira linha em que o tabuleiro não cumpre as normas.

checkCmd :: Comandos -> Bool

checkCmd verifica se os Comandos recebidos são válidos. Uma sequência de comandos válida deve conter apenas as seguinte os carateres 'A' (Avançar), 'S' (Saltar), 'E' (Esquerda), 'D' (Direita), 'L' (Luz).

Como resultado é devolvido um Bool cujo valor varia entre True se a sequência de comandos for válida; False se a sequência de comandos for inválida.

checkPos :: Posicao -> (Int, Int) -> Bool

checkPos recebe os limites do tabuleiro e verifica se a Posicao recebida está dentro do seguinte formato "x y ori", em que:

x corresponde à posição do Robot no eixo das abcissas num dado tabuleiro. O seu valor deve ser inteiro e variar entre 0 e o número de colunas do tabuleiro (exclusive);

y corresponde à posição do Robot no eixo das ordenadas num dado tabuleiro. O seu valor deve ser inteiro e variar entre o 0 e o número de linhas do tabuleiro (exclusive);

ori corresponde à orientação do Robot. Pode apenas variar nos seguintes valores: N (Norte), S (Sul), E (Este), O (Oeste).

Como resultado é devolvido um Bool cujo valor varia entre True se a posição for válida; False se a posição for inválida.

deepCheck :: Tabuleiro -> (Int, Int) -> [String] -> Int

deepCheck recebe um Tabuleiro inválido, as dimensões do tabuleiro, e o ficheiro que se encontra atualmente em teste.

Esta função reavalia qual a primeira linha em que o ficheiro divergiu das normas. Esta ação é necessária pois ficheiros com linhas após a sequência de comandos são divididos de forma incorreta.

Como resultado é devolvido um Int com o valor real da primeira linha onde o ficheiro diverge das normas especificadas.