Main

Description

O objectivo desta tarefa é implementar um mecanismo de geração de mapas. O input será a dimensão do mapa (número ímpar maior ou igual a 5) e um número inteiro positivo para usar como semente num gerador pseudo-aleatório. O output deverá ser um mapa impresso no formato:

Primeiro a grelha é descrita usando o caracter ‘#’ para representar pedra, ‘?’ para representar tijolo, e ‘ ‘ para representar uma célula vazia. Depois deverão ser listados os power ups escondidos, um por linha. Para cada power up deve ser indicado o seu tipo (o caracter ‘+’ representa Bombs e o caracter ‘!’ representa Flames) e as coordenadas onde se encontra (primeiro horizontal e depois vertical). Os power ups Bombs devem ser listados antes dos Flames, e devem ser ordenados aparecendo primeiro os que aparecem mais acima e mais à esquerda.

Para realizar esta tarefa decidimos criar o mapa em 3 fases

Inicialmente apenas com as suas paredes e o caracter v nos locais onde iria ser preenchido com espaços livres (' ') ou tijolos(?)

A seguir substituindo os vs pelos espaços, tijolos, bombas(+) ou flames(!) (a ideia de colocar inicialmente as bombas e flames no mapa era para facilitar a encontrar a localização das mesmas)

Por fim a grelha final substituindo as bombas(+) e flames(!) por tijolos

Documentation

main :: IO ()

mapa :: Int -> Int -> [String]

A principal função do nosso mapa, que é a junção de duas funções criadas, uma que gera a grelha e outra que gera a localição dos powerUPs

criarMapaV :: Int -> [String]

Função que cria a nossa grelha com vs nos quadrados a substituir por valores da lista

contV :: [[Char]] -> Int

Função que conta o total de vs no nosso mapa, ou seja os valores necessários na lista de inteiros gerada inicialmente

convtoMap :: Int -> Char

Funçao que converte de Int para um dos caracteres aceites no nosso mapa

subsVmapa :: [String] -> [Int] -> [String]

Funçao que substitui os vs do nosso mapa por ‘?’ para representar tijolo, ‘ ‘ para representar uma célula vazia, + para representar uma bomba e ! para representar flames

mapPU :: [String] -> [String]

Funçao que calcula as coordenadas dos powerUps

subPUmapa :: [String] -> [String]

Funçao que pega no mapa com + e ! a representar os powerUPs e os esconde atrás de tijolos (?)