Main

Documentation

type Mapa = [String]

type Grelha = [String]

type Elementos = [String]

type PUbomba = String

type PUflame = String

type Bomba = String

type Bombas = [Bomba]

type Raio = Int

type Jogador = String

type Linha = Int

type Coluna = Int

type Coordenada = (Coluna, Linha)

type Tempo = Int

type Dimensao = Int

type Direccao = Char

main :: IO ()

Função principal que invoca o jogo.

joga :: Estado -> (Estado -> Picture) -> (Event -> Estado -> Estado) -> (Float -> Estado -> Estado) -> IO ()

Função que cria um jogo.

type Estado = (Mapa, Tempo, [Picture])

Uma representação do estado do jogo.

estadoInicial :: [Picture] -> Estado

O estado inicial do jogo.

tupler :: String -> Coordenada

Função para adquirir a posicao no mapa de um PowerUP, jogador ou um bomba colocada

tuplesdados :: [String] -> [Coordenada]

Função que utilizamos para transformar a lista de bombas ou flames na lista das suas coordenadas

mapa :: Dimensao -> Int -> Mapa

A principal função do nosso mapa, que é a junção de duas funções criadas, uma que gera a grelha e outra que gera a localição dos powerUPs

criarMapaV :: Dimensao -> Mapa

Função que cria a nossa grelha com vs nos quadrados a substituir por valores da lista

contV :: [[Char]] -> Int

Função que conta o total de vs no nosso mapa, ou seja os valores necessários na lista de inteiros gerada inicialmente

convtoMap :: Int -> Char

Funçao que converte de Int para um dos caracteres aceites no nosso mapa

subsVmapa :: Grelha -> [Int] -> Grelha

Funçao que substitui os vs do nosso mapa por ‘?’ para representar tijolo, ‘ ‘ para representar uma célula vazia, + para representar uma bomba e ! para representar flames

mapPU :: [String] -> [String]

Funçao que calcula as coordenadas dos powerUps

subPUmapa :: [String] -> [String]

Funçao que pega no mapa com + e ! a representar os powerUPs e os esconde atrás de tijolos (?)

desenhaEstado :: Estado -> Picture

Função que desenha o jogo.

desenhargrelha :: [Picture] -> Grelha -> Float -> Float -> Float -> [Picture]

desenharespacoslivres :: [Picture] -> Grelha -> Float -> Float -> Float -> [Picture]

desenharespacoslivreslinha :: [Picture] -> String -> Float -> Float -> Float -> [Picture]

desenharlinha :: [Picture] -> String -> Float -> Float -> Float -> [Picture]

desenharelementos :: [Picture] -> Elementos -> [Coordenada] -> [Picture]

selecionaimagem :: [Picture] -> String -> Float -> Float -> Picture

reageEvento :: Event -> Estado -> Estado

Função que altera o estado do jogo quando acontece um evento.

move :: Mapa -> Int -> Char -> Mapa

grelha :: [String] -> [String]

Função que recebe o input e devolve apenas a grelha de jogo

bombas :: [String] -> [String]

Função que recebe o input e devolve a lista de bombas

flames :: [String] -> [String]

Função que recebe o input e devolve a lista de flames

bombacol :: [String] -> [String]

Função que recebe o input e devolve a lista das bombas colocadas

jogadores :: [String] -> [String]

Função que recebe o input e devolve a lista jogadores

jogador :: Char -> [String] -> String

Função que pega na lista de jogadores e devolve apenas um jogador

verposigrelha :: Grelha -> Coordenada -> Coordenada -> Char

Função que devolve a ocupação duma determinada posição da grelha

verposibomba :: String -> [String] -> Bool

Função que verifica se existe alguma bomba na posicao do jogador

verjogcolbom :: Int -> [String] -> Int

Função que conta quantas bombas um jogador já tem colocadas no mapa

criarbomba :: (Int, Int) -> Int -> Int -> Int -> String

Função que cria a nova string de uma bomba acabada de colocar

inserirbomba :: String -> [String] -> [String]

Função que insere a string de uma bomba acabada de colocar na nossa lista de bombas

maiortuple :: (Int, Int) -> (Int, Int) -> (Int, Int)

Função que copara dois tuples do tipo (coluna,linha) e devolve o maior

verPU :: String -> [String] -> Bool

Função que verifica se na posição destino do jogador existe algum PU

addPU :: String -> [String] -> String

Função que adiciona power up a um jogador

remPU :: String -> [String] -> [String]

Função que remove power up dada a um jogador

moverJogador :: String -> (Int, Int) -> String

Função que move o jogadaor na direçao do comando

removerjogadormovido :: Int -> [String] -> [String]

Função que remove o jogador movido da lista de jogadores inicial

reageTempo :: Float -> Estado -> Estado

Função que altera o estado do jogo quando o tempo avança n segundos.

avanca :: [String] -> Int -> [String]

bombasaexp :: [Bomba] -> [Bomba]

Função que seleciona as bombas que irão explodir no instante de tempo actual

decbombas :: [Bomba] -> [Bomba]

Função que remove bombas que irão explodir neste instante do estado actual e reduz 1 instante ao tempo em falta para estas explodirem

expbombas :: [Bomba] -> Mapa -> Mapa

Função que explode as bombas

expbomba :: Bomba -> Mapa -> Mapa

Função auxiliar de expbombas que explode cada bomba

removerjogpos :: Coordenada -> [Jogador] -> [Jogador]

Função que remove jogador que esteja na mesma posição que uma bomba a explodir

explodir :: Coordenada -> Coordenada -> Raio -> Mapa -> Mapa

Função que produz o efeito das chamas ao explodir uma bomba para um direcção

inserirflame :: Coordenada -> Grelha -> Grelha

função que insere a representação de uma flame na grelha

removerflames :: Grelha -> Grelha

Função que remove a representação de flames do instante anterior da grelha

removertijolo :: Coordenada -> Mapa -> Mapa

Função que remove um tijolo do mapa

tempobombaum :: Coordenada -> [Bomba] -> [Bomba]

Função que passa o tempo que falta para uma bomba explodir directamente para 1

eliminarjogatingidos :: Coordenada -> [Jogador] -> [Jogador]

Função que elimina os jogadores que estão numa determinada posição

fecharmapa :: Int -> Mapa -> Mapa

Função que coloca uma parede numa posição, eliminando todos os seus elementos, com base no tempo de forma espiral

removertudopos :: Coordenada -> Elementos -> Elementos

Função que remove os elementos que estão numa determinada posição

colocarparede :: Coordenada -> Grelha -> Grelha

Função que, na grelha de jogo, coloca uma parede numa dada posição

posicaoafechar :: Char -> Dimensao -> Coluna -> Linha -> Tempo -> Coordenada

Função que, dada a grelha de jogo e o tempo, indica qual posição se deve fechar nesse instante

reageBot :: Estado -> Estado

Função que recebe um estado e devolve um novo estado com os movimentos de cada Bot.

reageBot2 :: Estado -> Estado

reageBot3 :: Estado -> Estado

bot :: [String] -> Int -> Int -> Maybe Char

plantarbomba :: Coordenada -> Raio -> Mapa -> Bool

verutil :: Coordenada -> Raio -> Mapa -> Bool

Função que verifica se colocar uma bomba pode eventualmente destruir algum tijolo

verefeito :: Grelha -> Coordenada -> Raio -> Bool

Função que verifica se a explosão duma bomba causa algum efeito

verfeitoaux :: Grelha -> Coordenada -> Raio -> Int -> Direccao -> Bool

Função que verifica se a explosão duma bomba causa algum efeito numa determinada direcção(auxiliar de verefeito)

modoseguro :: Grelha -> Bombas -> Jogador -> Bool

Função que verifica se o jogador está livre de explosões na posição actual

modorisco :: Grelha -> Bombas -> Jogador -> Bool

versegpo :: Grelha -> Bombas -> Coordenada -> Bool

Função que verifica se uma posição esta segura

bombaslincol :: [Coordenada] -> Coordenada -> [Raio] -> Bool

Função que calcula se uma determinada posição está em risco de ocorrer uma explosão

distancia :: Coordenada -> Coordenada -> Int

Funçao que calcula a distancia entre 2 posições

raiobombas :: Bombas -> [Raio]

Função que recebe a lista de bombas do estado actual e devolve a lista dos seus raios

moverjogador :: Grelha -> Coordenada -> Coordenada -> Bombas -> Maybe Char

Função que decida para onde se deve mover o jogador(ou não efectuar qualquer movimento)

posipumaisprox :: Coordenada -> [Coordenada] -> (Coordenada, Int) -> Coordenada

Função que descobre qual a posição do PowerUp visível mais proximo do jogador, caso exista a menos de 10 movimentos, ou o centro do mapa

procurarposseg :: Grelha -> Jogador -> Bombas -> Maybe Char

Função que procura uma posição segura com apenas 3 comandos

somapos :: Coordenada -> Coordenada -> Coordenada

Função que dadas 2 coordenadas devolve a coordenada resultante da soma das duas

metdim :: String -> Int

Função que dada a 1ª linha da grelha devolve um inteiro correspondente a metade da dimensão(utíl para descobrir a posição central do mapa)

verposgrelha :: Grelha -> Coordenada -> Coordenada -> Maybe Char

Função que devolve a ocupação duma determinada posição da grelha

fr :: Int

Frame rate

dm :: Display

Display mode