Main

Description

O objectivo desta tarefa é, dada uma descrição do estado do jogo (idêntica à utilizada na tarefa2) implementar um mecanismo de compressão / descompressão que permita poupar caracteres e, desta forma, poupar espaço em disco quando o estado do jogo for gravado (permitindo, por exemplo, fazer pausa durante o jogo com o objectivo de o retomar mais tarde).

Para comprimir o estado actual do jogo decidimos:

Guardar a dimensao do mapa

Remover todas as paredes da nossa grelha

Manter as powerUps, bombas colocadas e os jogadores(sem os 2 primeiros caracteres de cada elemento e separados por vírgulas os elementos do mesmo tipo)

Para separar a dimensão da grelha sem paredes usamos o caracter d

Para separar a grelha sem paredes dos PU bomba usamos o caracter +

Para separar os powerUP bomba dos powerUP flames usamos o caracter !

Para separar os powerUP flames das bombas colocadas o caracter *

Para separar as bombascolocadas dos jogadores o caracter j

Documentation

main :: IO ()

encode :: [String] -> String

Função que transforma o estado actual do tipo [String] numa String em que se guarda a dimensão, a grelha sem paredes e a lista de powerUps, bombas colocadas e jogadores

decode :: String -> [String]

Função que devolve a String guardada ao seu estado inicial

guardardimensao :: [String] -> Int

Função que calcula a dimensão da grelha de jogo

removerParedes :: [String] -> String

Função que remove as paredes da grelha de jogo

recolocarParedes :: Int -> Int -> String -> [String]

Função que recoloca as paredes na grelha

separar :: [String] -> String

Função que transforma as bombas, flames, bombas colocadas ou os jogadores numa String com ',' a separar cada elemento

juntar :: String -> String -> [String]

Função que transforma as powerUps e bombas colocadas numa lista do tipo [String]

juntarjogadores :: String -> String -> [String]

Função que transforma os jogadores numa lista do tipo [String]

grelha :: [String] -> [String]

Função que recebe o input e devolve apenas a grelha de jogo

bombas :: [String] -> [String]

Função que recebe o input e devolve a lista de bombas

flames :: [String] -> [String]

Função que recebe o input e devolve a lista de flames

bombacol :: [String] -> [String]

Função que recebe o input e devolve a lista das bombas colocadas

jogadores :: [String] -> [String]

Função que recebe o input e devolve a lista jogadores