Main

Description

Documentation

mapa :: Int -> Int -> [String]

A função mapa gera a Arena de Batalha e as Coordenadas das Bombas e dos Power-Ups (mapa x y retorna uma Arena de Batalha de dimensão x e semente y e as Coordenadas das Bombas e dos Power-Ups).

Exemplos de Utilização:

>>> mapa 11 0
###########
#     ??  #
# # # #?# #
#  ? ?  ? #
#?# # #?# #
# ????   ?#
# #?#?#?#?#
#   ?  ?  #
# # # # # #
#         #
###########
+ 7 1
+ 7 2
+ 7 6
+ 9 6
! 7 4

>>> mapa 9 5
#########
#   ?   #
# #?#?# #
#  ??  ?#
# #?#?# #
#   ?? ?#
# #?# # #
#       #
#########
+ 5 4
! 4 5

mkRandom :: Int -> Int -> [Int]

A Função mkRandom gera aleatoriamente (x*x) números a partir de um número x (Inteiro) e de uma semente y (Inteiro).

chkPwUp :: Int -> String -> [String]

A Função chkPwUp gera as Coordenadas dos Power-Ups (chkPwUp recebe um Inteiro x (dimensão do mapa) e uma string a:as (mapa em que nas posições em que existam Power-Ups, os ' ? ' são substituidos pelo respetivo símbolo de Power-Up).

Exemplo de Utilização:

>>> chkPwUp (quando mapa 11 0)
+ 7 1
+ 7 2
+ 7 6
+ 9 6
! 7 4

>>> chkPwUp (quando mapa 9 0)
+ 5 4
! 4 5

sortStr :: [String] -> [String]

A Função sortStr ordena as Coordenadas dos Power-Ups (sortStr que recebe uma [String] proveniente da função chkPwUp e ordena as Coordenadas dos Power-Ups dessa mesma lista).

sortB :: [String] -> [String]

A Função sortB dá as Coordenadas das Bombas (sortB recebe uma [String] e produz uma outra [String] apenas com as Coordenadas das Bombas).

Exemplos de Utilização:

>>> sortB (quando mapa 11 0)
+ 7 1
+ 7 2
+ 7 6
+ 9 6

>>> sortB (quando mapa 9 5)
+ 5 4

sortF :: [String] -> [String]

A Função sortF dá as Coordenadas das Flames (/sortF/ recebe uma [String] e produz uma outra [String] apenas com as Coordenadas das Flames).

Exemplos de Utilização:

>>> sortF (quando mapa 11 0)
! 7 4

>>> sortF (quando mapa 9 5)
! 4 5

toString :: [(Char, (Int, Int))] -> [String]

A Função toString converte as Coordenadas dos Power-Ups em [String].

Exemplos de Utilização:

>>> toString [(+(5,4))]
+ 5 4

tradutor :: [(Char, Int)] -> Int -> [(Char, (Int, Int))]

A Função tradutor transforma as posições em Coordenadas.

taux :: Int -> Int -> Int

A Função taux dá a Coordenada Linha a partir de uma posição (Inteiro).

kaux :: Int -> Int -> Int

A Função kaux dá a Coordenada Coluna a partir de uma posição (Inteiro).

chkaux :: String -> Int -> [(Char, Int)]

A Função chkaux percorre uma String e sempre que encontra um Power-Up dá-nos a sua posição.

gama2 :: [Int] -> String

A Função gama2 percorre a [Int] fornecida pela mkRandom e atribui um caracter dependendo dos números fornecidos (A diferença entre esta Função e a Função gama, é que a gama não atribuí o devido caracter aos Power-Ups).

reString :: Int -> String -> [String]

A Função reString torna uma String numa [String].

replace :: String -> String -> String

A Função replace substitui . pelo ' ? ' ou ' '.

gama :: [Int] -> String

A Função gama percorre a [Int] fornecida pela mkRandom e atribui um caracter dependendo dos números fornecidos pela mkRandom.

mkmapGen :: Int -> [String]

A Função mkmapGen cria a partir de um número x (Inteiro) uma arena de dimensão x por x.

mkmap5 :: Int -> [String]

A Função mkmap5 cria uma arena de dimensão 5 por 5.

mkmap :: Int -> Int -> [String]

A Função mkmap cria uma arena de dimensão x por x.

oddaux :: Int -> Int -> String

A Função oddaux cria pedras nas linhas pares da Arena.

main :: IO ()

A Função main é a função de teste da Tarefa1.