Main

Description

Documentation

encode :: [String] -> String

A Função encode transforma uma [String] numa String através da função unlines.

enc :: [String] -> [String]

A Função enc "comprime" um Estado de Jogo num determinado momento, de maneira a poupar espaço de disco para quando esse mesmo jogo for novamente carregado.

deletc :: [String] -> [String]

A Função deletc utiliza a deletcaux para eliminar o caractér '#'.

deletcaux :: String -> String

A Função deletcaux procura e elimina o caractér '#' da [String] (Estado do Jogo).

deletp :: [String] -> [String]

A Função deletp remove o espaço e o símbolo representativo destes fatores : "* ","+ ","! ".

somaPb :: [String] -> Int

A Função somaPb percorre uma [String] e gera o resultado da soma dos Power-Ups (+) pertencentes a essa [String].

somaPf :: [String] -> Int

A Função somaPf percorre uma [String] e gera o resultado da soma dos Power-Ups (!) pertencentes a essa [String].

somaBb :: [String] -> Int

A Função somaBb percorre uma [String] e gera o resultado da soma das Bombas pertencentes a essa [String].

deletEp :: [String] -> [String]

A Função deletEp percorre uma [String] e elimina String vazias (exemplo de String vazia : " ").

deletsp2aux :: String -> String

A Função deletsp2aux elimina espaços pré-definidos.

deletsp1aux :: String -> String

A Função deletsp1aux elimina espaços pré-definidos.

deletsp2 :: [String] -> [String]

A Função deletsp2 utiliza a deletsp2aux para eliminar espaços pré-definidos.

deletsp1 :: [String] -> [String]

A Função deletsp1 utiliza a deletsp1aux para eliminar espaços pré-definidos.

deleteTo :: [String] -> [String]

A Função deleteTo utiliza as funções deletsp1 e deletsp2 (e as suas funções auxiliares), para eliminar espaços pré-definidos.

deletsp3 :: [String] -> [String]

A Função deletsp3 apaga espaços desnecessários que existem entre certos caracteres.

Exemplo de Utilização:

>>> Aquando da [String] com a Posição dos Jogadores
   "2 1 1" transforma-se em "21 1"

spToNb :: [String] -> [String]

A Função spToNb juntamente com as funções spnbgiv, stspnbgiv, ttnbgiv, spToNbaux enumeram os espaços que existem numa String antes e depois do caracter ' ? '.

spnbgiv :: String -> String

A Função spnbgiv juntamente com as funções spToNb, stspnbgiv, ttnbgiv, spToNbaux enumeram os espaços que existem numa String antes e depois do caracter ' ? '.

stspnbgiv :: String -> [String]

A Função stspnbgiv juntamente com as funções spToNb, spnbgiv, ttnbgiv, spToNbaux enumeram os espaços que existem numa String antes e depois do caracter ' ? '.

ttnbgiv :: [String] -> String

A Função ttnbgiv juntamente com as funções spToNb, spnbgiv, stspnbgiv, spToNbaux enumeram os espaços que existem numa String antes e depois do caracter ' ? '.

spToNbaux :: [String] -> [String]

A Função spToNbaux juntamente com as funções spToNb, spnbgiv, stspnbgiv, ttnbgiv enumeram os espaços que existem numa String antes e depois do caracter ' ? '.

decode :: String -> [String]

A Função decode "descomprime" um Estado de Jogo previamente guardado ("comprimido").

dec :: [String] -> [String]

A Função dec a partir da [String] obtida a partir da função del e remove-lhe as últimas 4 String, String essas que representam as variáveis utilizadas para "decifrar" o Estado de Jogo.

del :: [String] -> [String]

A Função del refaz novamente o mapa de jogo.

rePw :: [String] -> Int -> Int -> Int -> [String]

A Função rePw vai receber o número de Power-Ups e de bombas existentes e re-implementa os símbolos representativos ( "* ","+ ","! ").

rePl :: [String] -> [String]

A Função rePl percorre a [String] dos Jogadores e faz o processo inverso da função deletsp3.

Exemplo de Utilização:

>>> Aquando da [String] com a Posição dos Jogadores
   "21 1" transforma-se em "2 1 1"

reSpc :: [String] -> [String]

A Função reSpc divide uma [String] em String, String estas, que serão utilizadas pela função reSpcAux.

reSpcAux :: String -> String

A Função reSpcAux percorre as String enviadas pela função reSpc, e consoante o número existente antes ou depois de um ?, replica esse mesmo número em ' '.

simp :: [String] -> [String]

A Função simp separa a [String] (Estado do Jogo) numa [String] apenas com as Coordenadas dos Jogadores, dos Power-Ups e das Bombas.

simp2 :: [String] -> [String]

A Função simp2 separa a [String] numa [String] apenas com o mapa.

grabPu :: [String] -> [String]

A Função grabPu separa os Power-Ups da [String] (Estado de Jogo) criando apenas uma [String] com Power-Ups .

grabBb :: [String] -> [String]

A Função grabBb separa as Bombas da [String] (Estado de Jogo) criando apenas uma [String] com Bombas.

grabPl :: [String] -> [String]

A Função grabPl separa os Jogadores da [String] (Estado do Jogo) criando apenas uma [String] com Jogadores.

mkmapGen :: Int -> [String]

A Função mkmapGen cria a partir de um número x (Inteiro) uma arena de dimensão x por x.

mkmap5 :: Int -> [String]

A Função mkmap5 cria uma arena de dimensão 5 por 5.

mkmap :: Int -> Int -> [String]

A Função mkmap cria uma arena de dimensão x por x.

oddaux :: Int -> Int -> String

A Função oddaux cria pedras nas linhas pares da Arena.

reString :: Int -> String -> [String]

A Função reString torna uma String numa [String].

replace :: String -> String -> String

A Função reString torna uma String numa [String].

main :: IO ()

A Função main é a função de teste da Tarefa3.