Main

Documentation

retira :: String -> String

A Função retira separa a String com o Estado de Jogo, numa String apenas com o Mapa.

retira2 :: [String] -> [String]

A Função retira2 separa a [String] com o Estado de Jogo, numa [String] apenas com as Coordenadas dos Jogadores, Bombas e Power-Ups.

posGet2 :: String -> Float

A Função posGet2 dá-nos a Coordenada dos YY de um dado Jogador, Power-Up ou Bomba.

posGet3 :: String -> Float

A Função posGet2 dá-nos a Coordenada dos XX de um dado Jogador, Power-Up ou Bomba.

jogadorVivo :: [String] -> Bool

A Função jogadorVivo verifica se ainda existem Jogadores vivos.

type Estado = ((Float, Float), Float, [String], [Picture], Float)

Uma representação do estado do jogo.

estadoInicial :: IO Estado

O estado inicial do jogo.

desenhaEstado :: Estado -> Picture

Função que desenha o jogo.

desenha :: String -> [Picture] -> Float -> Float -> Float -> Float -> Float -> [Picture]

A Função desenha desenha o Mapa de Jogo, através da String fornecida pela retira.

desenha2 :: [String] -> [Picture] -> Float -> Float -> [Picture]

A Função desenha2 desenha os Jogadores e as Bombas, através da [String] fornecida pela retira2.

reageEvento :: Event -> Estado -> Estado

Função que altera o estado do jogo quando acontece um evento.

reageTempo :: Float -> Estado -> Estado

Função que altera o estado do jogo quando o tempo avança n segundos.

fr :: Int

Frame rate.

dm :: Display

Display mode.

joga :: Estado -> IO ()

Função principal que invoca o jogo.

main :: IO ()

Função principal que invoca o jogo.