Main

Description

Documentation

move :: [String] -> Int -> Char -> [String]

A Função move permite o manuseamento dos Jogadores, através da utilização de certos Comandos.

chkPlayer :: [String] -> Int -> Char -> [String]

A Função chkPlayer verifica qual o Número do Jogador que é pretendido que se mova e devolve a respetiva função desse mesmo Jogador.

fPl0 :: [String] -> [String] -> Int -> Char -> [String]

A Função fpl0 é uma função respetiva ao Jogador 0 que procura a String do Jogador 0 e dependendo do comando inserido leva-nos para a função comando (Se o comando inserido for B leva-nos para a função comandoB).

fPl1 :: [String] -> [String] -> Int -> Char -> [String]

A Função fpl1 é uma função respetiva ao Jogador 1 que procura a String do Jogador 1 e dependendo do comando inserido leva-nos para a função comando (Se o comando inserido for B leva-nos para a função comandoB).

fPl2 :: [String] -> [String] -> Int -> Char -> [String]

A Função fpl2 é uma função respetiva ao Jogador 2 que procura a String do Jogador 2 e dependendo do comando inserido leva-nos para a função comando (Se o comando inserido for B leva-nos para a função comandoB).

fPl3 :: [String] -> [String] -> Int -> Char -> [String]

A Função fpl3 é uma função respetiva ao Jogador 3 que procura a String do Jogador 3 e dependendo do comando inserido leva-nos para a função comando (Se o comando inserido for B leva-nos para a função comandoB).

comando :: [String] -> Char -> String -> String

A Função comando é uma função que verifica qual o comando que se pretenda que seja executado e verifica se é possível executar esse comando (Verifica se é possível mover-se para uma "dada posição", e caso seja verifica se existe algum tipo de powerUp e consequentemente move o respetivo Jogador).

chkmove :: [String] -> String -> Char -> Bool

A Função chkmove verifica se o movimento para uma determinada posição é válido (Se a posição para o qual pretendemos mover for '#' ou ' ? ', esta função retribui um False impedindo assim a ação).

posGet :: String -> (Int, Int)

A Função posGet recebe uma String e produz um (Int,Int)(O (Int,Int) fornecido por esta função corresponde às Coordenadas para as quais um Jogador se pretende mover).

charChk :: [String] -> (Int, Int) -> Char

A Função charChk utiliza o (Int,Int) produzido pela função posGet e distribui esses Int pelas funções stringSch e posSch.

stringSch :: [String] -> Int -> String

A Função stringSch devolve a String (linha) em que o Jogador se encontra.

posSch :: String -> Int -> Char

A Função posSch devolve o caracter presente na coluna onde o Jogador se encontra.

up :: String -> String

A Função up desloca as Coordenadas dos Jogadores uma entidade para cima.

dw :: String -> String

A Função dw desloca as Coordenadas dos Jogadores uma entidade para baixo.

lf :: String -> String

A Função lf desloca as Coordenadas dos Jogadores uma entidade para a esquerda.

rt :: String -> String

A função rt desloca as Coordenadas dos Jogadores uma entidade para a direita.

tkspe :: String -> Int -> Int -> String

A Função tkspe é uma função baseada na função take que funciona a partir do número de espaços (' ') com o objetivo de retirar certos Char.

tkspe2 :: String -> Int -> Int -> String

A Função tkspe2 é uma função baseada na função take que funciona a partir do número de espaços (' ') com o objetivo de retirar certos Char.

dpspe :: String -> Int -> Int -> String

A Função dpspe é uma função baseada na função drop que funciona a partir do número de espaços (' ') com o objetivo de retirar certos Char.

spchk :: String -> Int -> String

A Função spchk procura o Char (Representa a linha ou coluna) que vai sofrer alterações (Provenientes das Funções up , dw , lf , rt).

simp :: [String] -> [String]

A Função simp separa a [String] (Estado do Jogo) numa [String] apenas com as Coordenadas dos Jogadores, dos Power-Ups e das Bombas.

simp2 :: [String] -> [String]

A Função simp2 separa a [String] (Estado do Jogo) numa [String] apenas com o mapa.

chkPu :: [String] -> String -> String

A Função chkPu verifica se existe um Power-Up na posição para a qual o Jogador se pretende mover.

posPu :: [String] -> String -> String

A Função posPu gera a String correspondente aos Power-Ups existentes na posição para qual o Jogador se pretende mover.

givaux :: String -> Char -> String

A Função givaux adiciona o Power-Up ' + ' a uma String (String do Jogador).

auxgiv :: String -> Char -> String

A Função auxgiv adiciona o Power-Up ' ! ' a uma String (String do Jogador).

puGiv :: String -> String -> String

A Função puGiv adiciona Power-Ups à String correspondente a cada Jogador.

comandoB :: [String] -> Int -> String -> [String]

A Função comandoB verifica se é possível adicionar uma Bomba para uma dada posição.

chkBb :: [String] -> String -> Bool

A Função chkBb verifica se existe alguma Bomba no lugar onde um Jogador pretende colocar uma Bomba.

plBb :: [String] -> String -> Bool

A Função plBb verifica se o número de Bombas de um certo Jogador no mapa é superior ao número de Power-Ups Bomba que este possui.

nBb :: [String] -> Char -> Int

A Função nBb diz-nos o número de Bombas que um Jogador tem plantadas no mapa.

nPuB :: String -> Int

A Função nPuB diz-nos o número de Power-Ups Bombas que um Jogador possui.

nPuF :: String -> Int

A Função nPuF diz-nos o número de Power-Ups Flames que um Jogador possui.

grabPu :: [String] -> [String]

A Função grabPu separa os Power-Ups de uma [String] criando apenas uma [String] de Power-Ups.

grabBb :: [String] -> [String]

A Função grabBb separa as Bombas de uma [String] criando apenas uma [String] de Bombas.

grabPl :: [String] -> [String]

A Função grabPl separa os Jogadores de uma [String] criando apenas uma [String] de Jogadores.

ordn :: [String] -> String -> [String]

A Função ordn ordena as Bombas em relação às suas Coordenadas.

pl :: [String] -> Int -> String

A Função pl procura a String de um determinado Jogador na [String] do mapa.

upd :: [String] -> Int -> Char -> String -> [String]

A Função upd verifica se na posição para a qual um Jogador se movimentou existe algum Power-Up, e caso tenha, retira esse mesmo Power-Up da [String] do mapa.

main :: IO ()

A Função main é a função de teste da Tarefa2.