Main

Description

Documentation

avanca :: [String] -> Int -> [String]

A Função avanca faz decorrer um 1 instante de tempo no jogo

checktime :: [String] -> Int -> [String]

A Função checktime tem como objetivo verificar de o tempo de jogo é inferior a (n-2)^2(Sendo n a dimensão do mapa). Caso seja igual vai dar início à espiral, caso contrário remove 1 unidade de tempo.

chgchk :: [String] -> Int -> [String]

A Função chgchk testa se o mapa vai sofrer alterações depois de ocorrido 1 instante de tempo.

chkBomb :: [String] -> Bool

A Função chkBomb/ verifica se existem Bombas no mapa (Estas são as únicas afetadas pela alteração do tempo de jogo).

timeRemove :: [String] -> Int -> [String]

A Função timeRemove faz alterações à [String] contendo o Estado de Jogo (Remove 1 unidade de tempo e///ou explode Bombas).

xplodeChk :: [String] -> Bool

A Função xplodeChk percorre a [String] com o objetivo de saber se existe alguma Bomba (Cujo Timer seja Uma Unidade de Tempo). Se existir alguma Bomba nessas condições, essa mesma Bomba irá explodir e destruir tudo o que esteja no Raio de Ação desta (Destroí tudo excepto Pedras).

timeRemoveAux :: [String] -> [String]

A Função timeRemoveAux remove Uma Unidade de Tempo a todas as Bombas.

bombXplode :: [String] -> [String] -> Int -> [String]

A Função bombXplode faz uma bomba explodir.

bombXplodeAux :: [String] -> String -> Int -> [String]

A Função bombXplodeAux é uma Função Auxiliar da bombXplode.

bombGet :: [String] -> [String]

A Função bombGet percorre a [String] com o Estado de Jogo, criando uma [String] apenas com as Bombas cujo Timer seja de Uma Unidade de Tempo.

charGet :: [String] -> Int -> (Int, Int) -> Int -> [[String]]

A Função charGet percorre a [String] com o Estado de Jogo, devolvendo uma [String] com os diferentes Char que se encontrem no Raio de explosão da Bomba.

spup :: [String] -> Int -> (Int, Int) -> Int -> [String]

A Função spup gera os caracteres que se encontram na posição acima da Bomba.

spdw :: [String] -> Int -> (Int, Int) -> Int -> [String]

A Função spdw gera os caracteres que se encontram na posição abaixo da Bomba.

splf :: [String] -> Int -> (Int, Int) -> Int -> [String]

A Função splf gera os caracteres que se encontram na posição à esquerda da Bomba.

sprt :: [String] -> Int -> (Int, Int) -> Int -> [String]

A Função sprt gera os caracteres que se encontram na posição à direita da Bomba.

pathCheck :: [String] -> [[String]] -> [String]

A Função pathCheckVai procura Blocos (Destrutiveis ou Indestrutiveis) provinientes da charGet, com o objetivo de saber onde é que o raio de explosão da Bomba pára e que Blocos explodem

pathCheckAux :: [String] -> [String] -> String

A Função pathCheckAux é uma Função auxiliar da Função pathCheck.

augusto :: [String] -> [String] -> [String]

A Função augusto elimina os tijolos que foram afetados pelo raio de exlosão de uma bomba.

stringSchSp :: [String] -> (Int, Int) -> Char -> [String]

A Função stringSchSp procura String específicas no mapa.

posSchSp :: String -> Int -> Char -> String

A Função posSchSp procura o Char ? numa String e substitui esse mesmo Char por um espaço vazio.

deletEB :: [String] -> [String]

A Função deletEB elimina as Bombas que explodiram da [String] com o Estado de Jogo.

pathCheckSp :: [[String]] -> [[String]]

A Função pathCheckSp procura as posições pelo qual o raio de explosão de uma Bomba vai afetar, e verifica se alguma dessas posições coincide com uma posição de um Jogador.

pathCheckSpAux :: [String] -> [String]

A Função pathCheckSpAux é uma Função Auxiliar da pathCheckSp.

playerDChk :: [String] -> [[String]] -> [String]

A Função playerDChk verifica se existe algum Jogador no raio de explosão de uma Bomba.

playerDChkAux :: String -> [[String]] -> Bool

A Função playerDChk verifica se existe algum Jogador no raio de explosão de uma Bomba.

playerDChkAux2 :: String -> [String] -> Bool

A Função playerDChkAux2 é uma Função Auxiliar da playerDChkAux.

qChk :: String -> [String] -> [String]

A Função qChk verifica se existe algum Jogador na posição onde uma Bomba se encontra.

bombEchk :: [String] -> [[String]] -> [String]

A Função bombEchk verifica se existe alguma Bomba que esteja dentro do raio de explosão de uma outra Bomba, e caso exista diminui o tempo de explosão da Bomba afetada para 1.

bombEchkAux :: String -> [[String]] -> Bool

A Função bombEchkAux é uma Função Auxiliar da bombEchk.

bombEchkAux2 :: String -> [String] -> Bool

A Função bombEchkAux2 é uma Função Auxiliar da bombEchkAux.

toBombChg :: [String] -> [String] -> Int -> [String]

A Função toBombChg muda o timer, das bombas que se encontrem no raio de explosão de outras, para 1.

toBombChgAux :: [String] -> String -> Int -> [String]

A Função toBombChgAux é uma Função Auxiliar da toBombChg.

pwChk :: [String] -> [[String]] -> [[String]]

A Função pwChk verifica se existe algum Power-Up que se encontre no raio de explosão de uma Bomba.

pwChkAux :: [String] -> [String] -> [String]

A Função pwChkAux é uma Função Auxiliar da pwChk.

pwChkpos :: [String] -> String -> Bool

A Função pwChkpos é uma Função Auxiliar da pwChkAux.

delPw :: [String] -> [[String]] -> [String]

A Função delPw elimina os Power-Ups afetados pelo raio de explosão de uma Bomba.

delPwAux :: [String] -> [String] -> [String]

A Função delPwAux é uma Função Auxiliar da Função delPw.

delPwAux2 :: [String] -> String -> [String]

A Função delPwAux2 é uma Função Auxiliar da Função delPwAux.

fSpiral :: [String] -> Int -> Int -> [String]

A Função fSpiral é a Função responsável pela criação da Espiral.

forward :: [String] -> (Int, Int) -> Int -> Int -> Int -> Int -> [Int] -> [String]

A Função forward é responsável pela criação da parte superior da Espiral (Esquerda para a Direita).

downward :: [String] -> (Int, Int) -> Int -> Int -> Int -> Int -> [Int] -> [String]

A Função downward é responsável pela criação da parte latetal (Direita) da Espiral.

backward :: [String] -> (Int, Int) -> Int -> Int -> Int -> Int -> [Int] -> [String]

A Função backward é responsável pela criação da parte inferior da Espiral (Direita para a Esquerda).

upward :: [String] -> (Int, Int) -> Int -> Int -> Int -> Int -> [Int] -> [String]

A Função upward é responsável pela criação da parte latetal (Esquerda) da Espiral.

stsch :: [String] -> (Int, Int) -> [String]

A Função stsch procura uma certa linha no Estado de Jogo.

posSchSpacial :: String -> Int -> String

A Função posSchSpacial substitui um Char (na String obtida pela stsch) por um '#' (Efeito Espiral).

nbgiv :: Int -> [Int]

A Função nbgiv define o número de operações que cada função irá ter que executar, e também o número de vezez que cada função ira ser executada.

nbgivAux :: Int -> [Int]

A Função nbgivAux é uma Função Auxiliar da nbgiv.

dlt :: [String] -> (Int, Int) -> [String]

A Função dlt elimina qualquer Bomba, Power-Up ou Jogador que se encontre na mesma posição onde foi colocado um tijolo da espiral.

simp :: [String] -> [String]

A Função simp separa a [String] (Estado do Jogo) numa [String] apenas com as Coordenadas dos Jogadores, dos Power-Ups e das Bombas.

simp2 :: [String] -> [String]

A Função simp2 separa a [String] (Estado do Jogo) numa [String] apenas com o mapa.

spchk :: String -> Int -> String

A Função spchk procura o Char (Representa a linha ou coluna) que vai sofrer alterações.

tkspe :: String -> Int -> Int -> String

A Função tkspe é uma função baseada na função take que funciona a partir do número de espaços (' ') com o objetivo de retirar certos Char.

posGet :: String -> (Int, Int)

A Função posGet recebe uma String e produz um (Int,Int)(O (Int,Int) fornecido por esta função corresponde às Coordenadas para as quais um Jogador se pretende mover).

charChk :: [String] -> (Int, Int) -> Char

A Função charChk utiliza o (Int,Int) produzido pela função posGet e distribui esses Int pelas funções stringSch e posSch.

stringSch :: [String] -> Int -> String

A Função stringSch devolve a String (linha) em que o Jogador se encontra.

posSch :: String -> Int -> Char

A Função posSch devolve o Char presente na coluna onde o Jogador se encontra.

main :: IO ()

A Função main é a função de teste da Tarefa4.