Main

Description

Introdução

Esta tarefa consiste em implementar o jogo em si, utilizando o código desenvolvido nas tarefas anteriores e a biblioteca Gloss para criar uma interface gráfica para o jogo.

Conteudo

Quisemos criar um jogo simples que fosse no entanto divertido e completo, com gráficos mais modernos que os do jogo original da NES.

Para isso, decidimos implementar gráficos 3D para o mapa, mantendo no entanto os gráficos pixel-art presentes no jogo original.

Também quisemos criar vários níveis selecionáveis pelo jogador e implementar ainda funcionalidades multijogador, mas isto foi impossivel devido à falta de tempo.

Conclusão e discussáo

Gostamos do resultado final do nosso jogo, apesar de, devido a limitações do gloss, não ter a melhor performance.

Depois de fazer este projeto, percebemos que o Haskell não é uma linguagem ideal para o desenvolvimento de jogos, devido ao seu estilo funcional. Mais do que isso, achámos o Gloss bastante limitado, devido a limitações no que conta a manipulação de imagens e a falta de muitas funcionalidades presentes em bibliotecas de jogos para outras linguagens de programação, como por exemplo funcionalidades de áudio, suporte para comandos de consola e a habilidade para escrever texto utilizando fontes.

Documentation

fps :: Int

dm :: Display

drawEstado :: EstadoJogo -> IO Picture

updateEstado :: Float -> EstadoJogo -> IO EstadoJogo

eventEstado :: Event -> EstadoJogo -> IO EstadoJogo

main :: IO ()

Função principal da Tarefa 5.

NB: Esta Tarefa é completamente livre. Deve utilizar a biblioteca gloss para animar o jogo, e reutilizar de forma completa as funções das tarefas anteriores.