Main

Documentation

main :: IO ()

tarefa5 :: Picture -> (Float, Float)

A tarefa 5 tem como função receber uma Picture de input, e de output a largura e altura da Bounding Box do input . Esta função tem quatro variaveis locais :

ListaPontos : Vai executar a calculaPictures que nos da a XMax,XMin, YMax, Ymin de todas as Pictures do Input *largura - vai executar a choveBebados que vai selecionar o XMax e Xmin da listaPontos e faz a diferença dos Dois dando assim a a largura da BoundingBox que engloba o Input *altura - vai executar a choveCilindros que vai selecionar o YMax e Ymin da listaPontos e faz a diferença dos Dois dando assim a a altura da BoundingBox que engloba o Input *msg = é o output final da nossa tarefa , juntando assim a largura e altura num par de Floats

choveBebados :: [(Float, Float)] -> Float

choveBebados vai selecionar o XMax e Xmin da listaPontos e faz a diferença dos Dois dando assim a a largura da BoundingBox que engloba o Input O xMax da String de floats é selecionado pela renaMaiorX e o xMin da String de floats é selecionado pela renaMenorX .

choveCilindros :: [(Float, Float)] -> Float

choveCilindros vai selcionar o YMax e Ymin da listaPontos e faz a diferença dos Dois dando assim a altura da BoundingBox que engloba o Input o O xMax da String de floats é selecionado pela renaMaiorX e o xMin da String de floats é selecionado pela renaMenorX .

renaMaiorY :: [(Float, Float)] -> Float

renaMenorY :: [(Float, Float)] -> Float

intToFloat :: Int -> Float

inToFloat vai receber um int de input e transforma-o num Float. Esta função é utilizada no Bitmap para transformar a width e length dada neste Construtor (originalmente um int) em Float para ser mais fácil trabalhar com o mesmo.

calculaPicture :: Picture -> [(Float, Float)]

a função CalculaPicture é a função principal desta Tarefa , a qual utiliza pattern Matching e recursividade para aplicar os processos indicados a cada construtor de forma a obter uma lista de XMax,XMin, YMax, Ymin de Todas as Pictures do Input .

calculaPolygon :: [(Float, Float)] -> [(Float, Float)]

Calcula da lista de pontos dada no Construtor Polygon o XMax,XMin, YMax, Ymin , fazendo depois uma lista com emparelhados com o elemento 0 , uma vez que apenas interessa saber o valor Máximo e Mínimo de uma determinada figura para calcular a altura e largura do output final

calculaCircle :: Float -> [(Float, Float)]

Calcula através do raio o XMax,XMin, YMax, Ymin, fazendo depois uma lista com emparelhados com o elemento 0

calculaBitmap :: Float -> Float -> [(Float, Float)]

Calcula atravez do width e length o XMax,XMin, YMax, Ymin

calculaRotate :: Float -> Picture -> [(Float, Float)]

a funçao CalculaRotate é a função utilizada pela CalculaPictura para calcular o construtor Rotate, a qual utiliza pattern Matching e recurisvidade para aplicar os processos indicados a cada construtor de forma a obter uma lista de XMax,XMin, YMax, Ymin de Todas as Pictures do Input apos terem sofrido rotações.

roda :: Float -> [(Float, Float)] -> [(Float, Float)]

esta função recebe um ângulo de rotação e transforma-o em radianos, e após isso vai aplicar uma fórmula a todos os pontos de forma a saber qual é a posição do pontos após esta rotação.

paiNatalMaiorX :: [(Float, Float)] -> (Float, Float)

as funções paiNatal vão ter como função calcular um par de floats. Cada uma das funções paiNatal vai receber uma lista de floats do construtor Polygon e depois calcular pares com XMax, XMin, YMax, Ymin, respetivamente.

paiNatalMenorX :: [(Float, Float)] -> (Float, Float)

paiNatalMaiorY :: [(Float, Float)] -> (Float, Float)

paiNatalMenorY :: [(Float, Float)] -> (Float, Float)