Main

Funções do mooshak

inStr :: String -> [String]

Parte uma String numa lista de linhas

outStr :: [String] -> String

Junta uma lista de linhas numa String

Funções de teste

correTestes :: IO ()

Corre múltiplos testes para as três tarefas

correTeste :: ([String] -> [String]) -> String -> IO ()

Corre um teste para uma tarefa

main :: IO ()

tarefa1 :: [String] -> [String]

tarefa 1: retorna a mensagem da validaçao do codigo

Esta pode ser OK caso o codigo seja valido ou Número o menor numero da linha onde ocorre um erro :

txt corresponde ao mapa completo (tabuleiro+coordenada)

mapa corresponde ao tabuleiro

stringCoords corresponde as Coordenadas (lista de strings)

Validaçoes :

erroMapa : constituida por ValidaMapa

erroBordas : constituida por validaBordas

erroPontos : constituida por validaCaixas

erroCords : constituida por validaCoords e caso este nao apresente erros entao e executado o verificadas

verificadas : constituido por verificaCoords

dimx : comprimento do mapa

dimy : largura do mapa

linhaRepetida : constituida por tiralinha, zipada, repetidos e numeroDeLinhas

dividirMapa :: [String] -> ([String], [String])

divide o Mapa em listas de strings umas correspondem a um tabuleiro e a outra as coordenadas no mapa

carateresValidos :: Char -> Bool

parametros utilizados para dividir o mapa

carateresValidos1 :: String -> Bool

carateresValidos2 :: [String] -> Bool

espaco :: [String] -> [String]

juntaErros :: Int -> Int -> Int

calcula o menor inteiro entre dois obs : caso um destes seja 0 automaticamente dara o outro numero

validaMapa :: [String] -> Int

carateres validos dentro da secção do tabuleiro. obs : vai ser utilizada para dividir o tabuleiro e as coordenadas em duas partes e para validar o mapa

testa se o mapa (tabuleiro) tem todo o mesmo comprimento

validaBordas :: [String] -> Int -> Int

vai analisar em primeiro lugar as bordas horizontais caso cumpram as condiçoes vai analisar as bordasVerticais

validaVerticais :: [String] -> Int -> Int

validaCoords :: Int -> Int -> Int -> [String] -> Int

validaCoords : vai introduzir no testaCoords a lista de triplos (constiuido por uma coordenada e o contador) dada pelo converteCoords. caso seja igual a zero volta a repetir o processo para o resto da lista de triplos dada pelo converteCoords.

o testacoords confirma se o contador dado pelo coverteCoords é diferente de 0 , ou seja , se existem erros , e se as coordenadas estao dentro do mapa . obs : o Int dado por esta funçao é binario podendo ser intrepretado como um booleano .

converteCoords :: Int -> [String] -> [(Int, Int, Int)]

converte uma lista de Strings (stringCoords) em triplos(atravez do processaPosiçoes) . primeiro e segundo sao as cordenadas e o terceiro é o contador , ou seja , valor dado na contagem dos erros

algarismos' :: String -> Bool

analiseCoords :: String -> Bool

testa se as cordenadas sao contituidas por um par de numeoro/s

primeiros :: [(Int, Int, Int)] -> [(Int, Int)]

primeiro digito : vai retirar tirar enquanto é digit seguno digito : 1.vai deixar retirar enquanto e digito de forma a retirar o primeiro digito , 2. apos isso vai retirar o espaço vazio 3. retirar enquanto e digito

a funçao primeiros tem como input a funçao convertCoords e retira o contador , deixando apenas pares correspondentes a coordenadas do mapa

verificaCoords :: [(Int, Int)] -> [String] -> Int -> Int

verificaCoords :verifica se as coordenadas sao um espaço vazio ou um pontos, uma vez que correspondem a posiçao inicial das caixas e a posiçao do boneco

localizaCoords : recebe num par de cordenadas e localiza o caracter respetivo no mapa , confirmando que estas se encontram dentro do mapa (tabuleiro)

(!!!) :: Eq a => [a] -> Int -> a

validaCaixas :: [String] -> [String] -> Int

determina se o numero de pontos no mapa (posiçao final de caixas ) é igual ao numero de coordenadas das caixas

repetidos :: [(Int, Int)] -> [(Int, Int)] -> [Bool]

Verifica se nao existem coordenadas repetidas e no caso de existir ira dar a primeira linha repetida

numeroDeLinhas :: Int -> Int -> [Int]

zipada :: [Bool] -> [Int] -> [(Bool, Int)]

tiraLinha :: [(Bool, Int)] -> Int