Main

Documentation

inStr :: String -> [String]

outStr :: [String] -> String

correTestes :: IO ()

correTeste :: ([String] -> [String]) -> String -> IO ()

Corre um teste para uma tarefa

main :: IO ()

tarefa4 :: [String] -> [String]

pai :: [String] -> [(Int, Int)] -> [String] -> String

A funçao pai vai realizar as seguintes tarefas :

Aplicar a validaPontos que por sua vez vai analisar o tabuleiro de modo a contar o numero de pontos , ou seja, se ainda existem caixas que necessitam de ser movidas para a sua posição final .

Caso ja não exitam pontos na lista de strings (tabuleiros) entao é dado a mensagem Fim , se não existem duas opções : uma é caso ainda existam comandos então a lista de strings e processada de novo com as alterações feitas pelas valida.. ou seja as coordenadas sao alteradas e é dada uma nova lista de strings a funçao pai .

A funçao lerComando tem como objetivo receber as coordenadas do mapa , o tabuleiro e o comando , e conforme este comando este vai executar uma das validas , o qual vai alterar as coordenadas do boneco , se este se puder deslocar, e uma caixa caso o boneco desloque uma delas caso seja dado um comando nao aceitável ou reconhecido, este é ignorado

mae :: [String] -> [(Int, Int)] -> [String] -> Int -> Int

A função mae tem a mesma estrutura da função pai , no entanto o seu output é diferente na questão de em vez de devolver uma msg , vai devolver o número de comandos realmente executados pelo boneco , por meio de comparação da lista da lista de strings (tabuleiros), em que caso o tabuleiro dado de input e o dado pelo output do lerComando sejam diferentes então é incrementado +1 ao contador dos comandos executados (i). casos estes sejam iguais então mantem-se atual tick . o validaPontos para saber se o jogo está finalizado ou não. O tick (i) é dado em dois casos: caso os comandos acabem ou caso o mapa esteja finalizado

converterComandoMapa :: String -> [String]

a converte comandos serve para transformar a ultima string do ficheiro.txt que serve de input à tarefa , ou seja , os comandos que se pretendem serem executados , de uma string para uma lista de strings desta forma vão ser interpretados pela função pai e mae

validaUp :: (Int, Int) -> [String] -> [(Int, Int)] -> [(Int, Int)]

a validaUp vai validar se o boneco dentro do contexto do mapa pode ou não mover-se, e caso este se mova, se vai ou não alterar também a posição de uma caixa . Caso este seja alterado então vai ser executado a função modificaCoordsUp

Esta função vai receber de input a coordenada do elemento a ser modificado (boneco ou caixa) e vai neste caso adicionar +1 unidade à dimensão y da coordenada dada no input, ou seja de modo a mimicar o comando Up (U) no jogo

modificaCoordsUp :: (Int, Int) -> [(Int, Int)] -> [(Int, Int)]

validaDown :: (Int, Int) -> [String] -> [(Int, Int)] -> [(Int, Int)]

a validaDown validar se o boneco dentro do contexto do mapa pode ou não mover-se, e caso este se mova se vai ou não alterar tambem a posição de uma caixa . Caso este seja alterado entao vai ser executado a funçao modificaCoordsDown

Esta função vai receber de input a coordenada do elemento a ser modificado (boneco ou caixa) e vai neste caso adicionar -1 unidade à dimensao y da coordenada dada no input, ou seja de modo a mimicar o comando Down (D) no jogo

modificaCoordsDown :: (Int, Int) -> [(Int, Int)] -> [(Int, Int)]

validaLeft :: (Int, Int) -> [String] -> [(Int, Int)] -> [(Int, Int)]

a validaLeft vai validar se o boneco dentro do contexto do mapa pode ou não se mover , e caso este se mova se vai ou não alterar também a posição de uma caixa . Caso este seja alterado então vai ser executado a função modificaCoordsLeft

Esta função vai receber de input a coordenada do elemento a ser modificado (boneco ou caixa) e vai neste caso subtrair -1 unidade à dimensao x da coordenada dada no input, ou seja de modo a mimicar o comando Left (L) no jogo

modificaCoordsLeft :: (Int, Int) -> [(Int, Int)] -> [(Int, Int)]

validaRight :: (Int, Int) -> [String] -> [(Int, Int)] -> [(Int, Int)]

a validaRight vai validar se o boneco dentro do contexto do mapa pode ou nao se mover , e caso este se mova se vai ou nao alterar tambem a posição de uma caixa. Caso este seja alterado entao vai ser executado a funçao modificaCoordsRight

Esta função vai receber de input a coordenada do elemento a ser modificado (boneco ou caixa) e vai neste caso adicionar +1 unidade a dimensao x da coordenada dada no input, ou seja de modo a mimicar o comando Right (R) no jogo

modificaCoordsRight :: (Int, Int) -> [(Int, Int)] -> [(Int, Int)]

validaPontos :: [String] -> Bool

a validaPontos vai concatetar a Lista de string numa unica string, e após isso conta o número de pontos dentro desta lista, ou seja , dentro do tabuleiro e caso estes sejam 0, então todas as caixas estao na posição final dando assim um output booleano True, caso o número seja diferente de zero então o mapa ainda tem caixas que têm de estar na sua posição final e da assim um output False

toniCabrita :: [(Int, Int)] -> [(Int, Int)] -> Bool

Esta função compara as coordenadas de input e output de modo a verificar se estas são iguais ou se foram modificadas ou seja se o boneco e algum elemento do tabuleiro foi modificado

caracterCaixa :: [(Int, Int)] -> [String] -> [String] -> [String]

O caracterCaixa identifica, consoante as coordenadas das caixas, o respetivo espaço vazio correspondente a uma caixa Com isto, é selecionado esse espaço vazio, com a selecaoCaixa, que por sua vez chama a função transformaCaixa que irá transformar o espaço vazio num H que corresponde a uma caixa no mapa do jogo.

selecaoCaixa :: [String] -> Int -> Int -> [String]

transformaCaixa :: String -> Int -> String

caracterBoneco :: [(Int, Int)] -> [String] -> [String] -> [String]

o caracterBoneco recebe de input uma lista da coordenada inicial do boneco e duas vezes o novoMapa , uma vez que um irá servir para localizar a coordenada no mapa e outro para ter as o carateres alterado na respetiva posição inicial do boneco

selecaoBoneco :: [String] -> Int -> Int -> [String]

a selecaoBoneco seleciona a string do mapa (y da coordenada) onde se encontra a coordenada da inicial do boneco para a qual queremos alterar o carater para o

transformaBoneco :: String -> Int -> String

a transforamaBoneco seleciona dentro da string dada pela selecaoCaixa o carater correspondente a coordenada (x da coordenada) e substitui por o

carateresValidos :: Char -> Bool

Carateres válidos dentro da secção do tabuleiro. obs : vai ser utilizada para dividir o tabuleiro e as coordenadas em duas partes e para validar o mapa

converteCoords :: Int -> [String] -> [(Int, Int, Int)]

Converte uma lista de Strings (stringCoords) em triplos (através do processaPosicoes) . Em primeiro e segundo são as coordenadas e o terceiro é o contador, ou seja, valor dado na contagem dos erros

analiseCoords :: String -> Bool

Testa se as coordenadas são constituídas por um par de números

algarismos' :: String -> Bool

Primeiro dígito : vai retirar enquanto é digito Segundo digito : 1. Vai deixar retirar enquanto e digito de forma a retirar o primeiro digito , 2. Após isso vai retirar o espaço vazio 3. Retirar enquanto é dígito

primeiros :: [(Int, Int, Int)] -> [(Int, Int)]

A função primeiros tem como input a função converteCoords e retira o contador, deixando apenas pares correspondentes a coordenadas do mapa

localizaCoords :: (Int, Int) -> [String] -> Char

localizaCoords : recebe num par de coordenadas e localiza o caracter respetivo no mapa, confirmando que estas se encontram dentro do mapa (tabuleiro)

(!!!) :: Eq a => [a] -> Int -> a