Tarefa1

Documentation

ghostHouseUneven :: Maze

GHOST HOUSES MODELS

ghostHouseEven :: Maze

generateMaze :: Int -> Int -> Int -> Maze

Gera o labirinto final

generateCont

Arguments

:: Int
Número de peças por corredor

-> Int
Número de corredores

-> Int
Seed

-> [[Int]]
Matriz de inteiros que serão convertidos em peças

Gera o conteúdo do interior do labirinto

tuppleM

Arguments

:: Maze
Labirinto

-> (Int, Int)
Posição no labirinto

-> Maze
Labirinto Resultante

Põe um tupple em labirinto

Funções que validam um labirinto

validBorder

Arguments

:: Maze
Labirinto

-> Bool
Laterais válidas ou não

Verifica as laterais de um labirinto são válidas

validTunnel

Arguments

:: Corridor
Corredor

-> Bool
Corredor válido ou não válido

Verifica se o túnel é válido

validCorridorW

Arguments

:: Corridor
Corredor a testar

-> Bool
Valida ou não o corredor

Verifica se um corredor é válido não sendo nem a primeira nem a última lista do labirinto

validMiddle

Arguments

:: Maze
Labirinto

-> Bool
Meio válido ou não válido

Verifica se o meio é válido do labirinto é válido

validMaze

Arguments

:: Maze
Labirinto

-> Bool
Labirinto Válido ou não válido

Verifica a validade de um labirinto inteiro

Auxiliares para a casa dos fantasmas

middleFinder

Arguments

:: Maze
Labirinto

-> (Int, Int)
Par de corredores que serão os túneis

Diz qual o corredor ou corredores do meio para identificar o espaço onde fica a casa dos fantasmas | Se o labirinto tiver altura par são 2 corredores, se for ímpar, apenas o segundo elemento do par é preenchido

posGHLine

Arguments

:: Maze
Labirinto

-> Int
Corredor onde começará a casa dos fantasmas

Indica a posição da casa dos fantasmas em linha

posGHColumn

Arguments

:: Maze
Labirinto

-> Int
Indica o índex dos corredores onde começará a casa dos fantasmas

Indica a posição da casa dos fantasmas em coluna

putGhostHouse

Arguments

:: Maze
Labirinto original

-> Maze
Labirinto já com a casa dos fantasmas

Tendo um labirinto original, posiciona a respetiva casa de fantasmas

Funções que constroem o labirinto

firstLast

Arguments

:: Int
Número de peças por corredor

-> Maze
Labirinto com o primeiro e último corredores

Gera o primeiro e último corredores do labirinto(só walls)

addWall

Arguments

:: Maze
Labirinto

-> Maze
Labirinto com as paredes exteriores

Adiciona a primeira e última paredes nos corredores(exceto o primeiro e último) do labirinto

addTunnelCorridor

Arguments

:: Int
Index

-> Maze
Labirinto

-> Maze
Labirinto Resultante

Update do index num labirinto

middleKill

Arguments

:: Maze
Labirinto

-> (Int, Int)

-> Maze

Remove o corredor do meio das listas de corredores do meio (lista de corredores excluindo a primeira e ultima lista)

Auxiliares para matrizes e corredores

listInList

Arguments

:: [a]
Primeira Lista

-> Int
Índice

-> [a]
Segunda Lista

-> [a]
Lista Resultante

Pega numa lista e substitui uma segunda lista pelos elementos da lista anterior no índice desejado

listInMatrix

Arguments

:: [[a]]
Matriz

-> (Int, Int)
Posição da Matriz

-> [a]
Lista para substituir

-> [[a]]
Matriz Resultante

Pega numa matriz e numa posição da mesma, e coloca uma lista na posiçao dada

matrixInMatrix

Arguments

:: [[a]]
Primeira Matriz

-> (Int, Int)
Posição da primeira matriz

-> [[a]]
Segunda Matriz

-> [[a]]
Matriz Resultante

Insere uma matriz numa certa posição de outra matriz

Funções do Generator

generateRandoms :: Int -> Int -> [Int]

Given a seed returns a list of n integer randomly generated

subLists :: Int -> [a] -> [[a]]

convertPiece :: Int -> Piece

Converts an integer number into a Peca 3 = Comida Grande 0 n < 7 <= Comida Pequena 7 <= 9 <= Parede

convertCorridor :: [Int] -> Corridor

Converts a list of integers into a Corredor

convertMaze :: [[Int]] -> Maze

Converts a list of lists of integers into a Labirinto