Tarefa1

Testes tarefa 1

t1 :: (Int, Int, Int)

gera um labirinto de largura par e altura impar

t2 :: (Int, Int, Int)

gera um labirinto de largura impar e altura par

t3 :: (Int, Int, Int)

gera um labirinto de largura e altura par

t4 :: (Int, Int, Int)

gera um labirinto de largura e altura impar

geraAleatorios

Arguments

:: Int
número de numeros que queremos gerar

-> Int
número aleatório

-> [Int]
lista de números aleatórios

Given a seed returns a list of n integer randomly generated

convertePeca

Arguments

:: Int
número para converter a peça

-> Piece
peça resultante

Converts an integer number into a Peca 3 = Food Big 0 n < 70 <= Food Little 70 <= 99 <= Wall

printCorridor

Arguments

:: Corridor
corredor

-> String
corredor convertido numa string

Converts a Corridor to a string

printMaze

Arguments

:: Maze
labirinto

-> String
labirinto convertido numa string

Converts a Maze to a string

converteCorridor

Arguments

:: [Int]
lista de números aleatórios

-> Corridor
corredor resultante dos números aleatórios

Converts a list of integers into a Corridor

criaCorridor

Arguments

:: Int
largura do labirinto

-> Corridor
lista de peças

-> Maze
labirinto resultante dos corredores

Forma o labirinto com a lista de peças

generateMaze

Arguments

:: Int
largura do labirinto

-> Int
altura do labirinto

-> Int
random seed

-> Maze
labirinto obtido

Gerar o labirinto

geraMazeImpImp

Arguments

:: Maze
labirinto sem a casa dos fantasmas

-> Int
altura do labirinto

-> Maze
labirinto resultante com a casa dos fantasmas

Gerar a casa dos fantamas no labirinto com largura e altura impar

geraMazeParImp

Arguments

:: Maze
labirinto sem a casa dos fantasmas

-> Int
altura do labirinto

-> Maze
labirinto resultante com a casa dos fantasmas

Gerar a casa dos fantamas no labirinto com largura par e altura impar

geraMazeImpPar

Arguments

:: Maze
labirinto sem a casa dos fantasmas

-> Int
altura do labirinto

-> Maze
labirinto resultante com a casa dos fantasmas

Gerar a casa dos fantamas no labirinto com largura impar e altura par

geraMazeParPar

Arguments

:: Maze
labirinto sem a casa dos fantasmas

-> Int
altura do labirinto

-> Maze
labirinto resultante com a casa dos fantasmas

Gerar a casa dos fantamas no labirinto com largura e altura par

geraWallsSupInf

Arguments

:: Int
largura do labirinto

-> Corridor
corredor obtido de gerar as paredes inferiores e superiores

Cria as paredes superiores e inferiores do labirinto

geraWallsLaterais

Arguments

:: Maze
labirinto sem as paredes laterais

-> Maze
labirinto resultante de gerar as paredes laterais

Cria as paredes laterais do labirinto

geraTunel

Arguments

:: Corridor
corredor sem o tunel

-> Corridor
corredor resultante de colocar o tunel

Gerar o tunel no corredor respetivo

novoLab

Arguments

:: Maze
labirinto para colocar o tunel

-> Int
lugar onde colocar o tunel

-> Maze
labirinto resultante

Gerar labirinto com o tunel

novoLabPar

Arguments

:: Maze
labirinto para colocar o tunel

-> Int
lugar onde colocar o tunel

-> Maze
labirinto resultante

se a altura do Maze for par acrescenta outro tunel a linha de cima

geraCasaMeioPar

Arguments

:: Corridor
corredor sem a casa dos fantasmas

-> Corridor
corredor com a casa dos fantasmas

Gerar a parte do meio da casa nos Corredores nos labirintos cuja largura é par

geraCasaMeioAux :: Integral a => [Piece] -> a -> a -> [Piece]

funçao auxiliar para gerar a parte do meio da casa nos Corredores nos labirintos cuja largura é par

geraCasaBaixoPar

Arguments

:: Corridor
corredor sem a casa dos fantasmas

-> Corridor
corredor com a casa dos fantasmas

Gerar a parte de baixo da casa nos Corredores nos labirintos cuja largura é par

geraCasaBaixoAux :: Integral a => [Piece] -> a -> a -> [Piece]

funçao auxiliar para gerar a parte de baixo da casa nos Corredores nos labirintos cuja largura é par

geraCasaCimaPar

Arguments

:: Corridor
corredor sem a casa dos fantasmas

-> Corridor
corredor com a casa dos fantasmas

Gerar a parte de cima da casa nos Corredores nos labirintos cuja largura é par

geraCasaCimaAux :: Integral a => [Piece] -> a -> a -> [Piece]

funçao auxiliar para gerar a parte de cima da casa nos Corredores nos labirintos cuja largura é par

geraCasaMeioImpar

Arguments

:: Corridor
corredor sem a casa dos fantasmas

-> Corridor
corredor com a casa dos fantasmas

Gerar a parte do meio da casa nos Corredores nos labirintos cuja largura é impar

geraCasaMeioAux2 :: Integral a => [Piece] -> a -> a -> [Piece]

funçao auxiliar para gerar a parte do meio da casa nos Corredores nos labirintos cuja largura é impar

geraCasaBaixoImpar

Arguments

:: Corridor
corredor sem a casa dos fantasmas

-> Corridor
corredor com a casa dos fantasmas

Gerar a parte de baixo da casa nos Corredores nos labirintos cuja largura é impar

geraCasaBaixoImpAux :: Integral a => [Piece] -> a -> a -> [Piece]

função auxiliar para gerar a parte de baixo da casa nos Corredores nos labirintos cuja largura é impar

geraCasaCImpar

Arguments

:: Corridor
corredor sem a casa dos fantasmas

-> Corridor
corredor com a casa dos fantasmas

Gerar a parte de cima da casa nos Corredores nos labirintos cuja largura é impar

geraCasaCimparAux :: Integral a => [Piece] -> a -> a -> [Piece]

funçao auxiliar para gerar a parte de cima da casa nos Corredores nos labirintos cuja largura é impar

limpaPar

Arguments

:: Corridor
corredor

-> Corridor
corredor limpo a volta da casa

limpar a volta da casa dos fantasmas no labirinto cuja largura é par

limpaParAux :: Integral a => [Piece] -> a -> a -> [Piece]

funçao auxiliar para limpar a volta da casa dos fantasmas no labirinto cuja largura é par

limpar

Arguments

:: Corridor
corredor

-> Corridor
corredor limpo a volta da casa

limpar a volta da casa dos fantasmas no labirinto cuja largura é impar

limparAux :: Integral a => [Piece] -> a -> a -> [Piece]

função auxiliar para limpar a volta da casa dos fantasmas no labirinto cuja largura é impar

casaMeioLabPar

Arguments

:: Maze
labirinto

-> Int
lugar onde colocar a parte do meio da casa

-> Maze
labirinto resultante de colocar a casa

Colocar a parte do meio da casa no labirinto cuja largura é par

casaCimaLabPar

Arguments

:: Maze
labirinto

-> Int
lugar onde colocar a parte de cima da casa

-> Maze
labirinto resultante de colocar a casa

Colocar a parte de cima da casa no labirinto cuja largura é par

casaBaixoLabPar

Arguments

:: Maze
labirinto

-> Int
lugar onde colocar a parte de baixo da casa

-> Maze
labirinto resultante de colocar a casa

Colocar a parte de baixo da casa no labirinto cuja largura é par

casaMeioLabImpar

Arguments

:: Maze
labirinto

-> Int
lugar onde colocar a parte do meio da casa

-> Maze
labirinto resultante de colocar a casa

Colocar a parte do meio da casa no labirinto cuja largura é ímpar

casaCimaLabImpar

Arguments

:: Maze
labirinto

-> Int
lugar onde colocar a parte de cima da casa

-> Maze
labirinto resultante de colocar a casa

Colocar a parte de cima da casa no labirinto de largura impar

casaBaixoLabImpar

Arguments

:: Maze
labirinto

-> Int
lugar onde colocar a parte de baixo da casa

-> Maze
labirinto resultante de colocar a casa

Colocar a parte de baixo da casa no labirinto de largura impar

limparParLab

Arguments

:: Maze
labirinto

-> Int
lugar onde limpar

-> Maze
labirinto limpo à volta da casa dos fantasmas

Limpar o labirinto à volta da casa dos fantasmas de largura par

limparLab

Arguments

:: Maze
labirinto

-> Int
lugar onde limpar

-> Maze
labirinto limpo à volta da casa dos fantasmas

Limpar o labirinto à volta da casa dos fantasmas de largura par