Tarefa1

Description

O objetivo deste módulo e formar um mapa com um comprimento (x) uma altura (y) e uma semente (s). Um mapa é um conjunto de y corredores com cada um com tamanho x a semente serve de um numero arbritario para o gerador de pseudo random este modulo forma uma lista de inteiros pseudo random, transforma cada inteiro com a sua respectiva peça e vai formando um corredor. Um mapa do jogo pacman tem sempre um tunel e uma casa dos fantasma no meio do mapa que as suas entradas e dimensoes respectivamente, dependem de x e y.

Documentation

generateRandoms :: Int -> Int -> [Int]

Função que cria uma lista de n elementos de numeros random

convertePeca

Arguments

:: Int
recebe um inteiro

-> Piece
dá a peça respectiva ao inteiro

Função que converte um inteiro para a sua respectiva peca

converteCorredor

Arguments

:: [Int]
recebe uma lista de inteiros

-> Int
recebe o tamanho da lista (x)

-> Corridor
dá uma lista de peças de tamanho x

Função que converte uma lista de inteiros para um corredor utilizando a função anterior definida convertePeca

printCorridor

Arguments

:: Corridor
recebe um corredor com x peças

-> String
dá uma string com x caracteres

Função que transforma um corredor em string

mazetostring

Arguments

:: Maze
recebe uma lista de corridores

-> String
dá uma string do labirinto

Função que torna um Maze em string

printmaze :: Maze -> IO ()

função auxiliar que forma um maze com espacos

paredes

Arguments

:: Int
recebe um inteiro x

-> Corridor
o resultado e um corredor com x peças (Wall)

Função que forma as paredes, estas que sao corredores so com pecas Wall

corridorValido

Arguments

:: [Int]
recebe uma lista de int

-> Int
recebe um int x

-> Corridor
o output vai ser um corridor de tamanho x com wall no inicio e fim

Função que cria um corredor generico com uma lista de inteiros que forma todos os corredores do mapa com excecao das paredes e do tunel

juntaC

Arguments

:: [Int]
recebe uma lista de int

-> Int
recebe o tamanho dos corridores (x)

-> Int
recebe o numero de corridores (y)

-> Maze
gera uma labirinto com y corridores validos de tamanho x

Função que junta varios corredores validos

corredor1tunel

Arguments

:: [Int]
recebe uma lista de inteiros

-> Int
recebe o tamanho do corredor x

-> Corridor
dá o primeiro corredor do tunel de tamanho x

Função que forma o primeiro corredor do tunel dando drop nos numeros da lista que ja foram anteriormente utilizadas

corredor2tunel

Arguments

:: [Int]
recebe uma lista de inteiros

-> Int
recebe o tamando do corridor x

-> Corridor
dá o segundo corridor do tunel de tamanho x

Função que forma o segundo corredor do tunel dando drop nos numeros da lista que ja foram anteriormente utilizadas

corredor3tunel

Arguments

:: [Int]
recebe uma lista de inteiros

-> Int
recebe o tamanho do corridor x

-> Corridor
dá o terceiro corridor de tamanho x

Função que forma o segundo corredor do tunel no caso de o tunel ter só uma entrada dando drop nos numeros da lista que ja foram anteriormente utilizadas

corredor3tunelC

Arguments

:: [Int]
recebe uma lista de inteiros

-> Int
recebe o tamanho do corridor x

-> Corridor
dá o segundo corridor do tunel de tamanho x

Função que forma o segundo corredor do tunel no caso de o tunel ter duas entradas dando drop nos numeros da lista que ja foram anteriormente utilizadas

corredorpvazias

Arguments

:: [Int]
recebe uma lista de inteiros

-> Int
recebe o tamanho do corridor x

-> Corridor
dá um corridor de tamanho x

Função que forma o corredor anterior ao inicio do tunel dando drop nos numeros da lista que ja foram anteriormente utilizadas

juntatunel

Arguments

:: [Int]
recebe uma lista de inteiros

-> Int
recebe o numero de corridores x

-> Maze
da x corridores compactados

Função que forma o tunel completo no caso de ter apenas uma entrada

juntatunelC

Arguments

:: [Int]
recebe uma lista de inteiros

-> Int
recebe o numero de corridores x

-> Maze
da x corridores compactados

Função que forma o tunel completo no caso de ter duas entradas

generateMazeAux

Arguments

:: [Int]
recebe uma lista de inteiros

-> Int
recebe um inteiro para analisar se e par ou nao

-> Int
recebe um inteiro para analisar se e par ou nao

-> Maze
gera um labirinto com as respetivas funçoes

Função que respetivamente analiza se as variaveis são pares ou impares e manda para as funções corretas

generateMaze

Arguments

:: Int
recebe uma lista de inteiros

-> Int
recebe o comprimento x

-> Int
recebe o numero de corridores y

-> Maze
gera o labirinto final com as dimensoes x y

Função que gera a lista de numeros pseudo random e envia para a generateMazeAux com o comprimento e altura do labirinto para ser devidamente analizada