Tarefa2

Documentation

play :: Play -> State -> State

Função principal: | Testa se o jogador não está morto, com a orientação errada ou a entrar num túnel | Caso nenhuma destas situações ocorra, o estado será alterado pela função updateState

updateState

Arguments

:: Player
Jogador

-> State
Estado atual do jogo

-> State
Estado atualizado

Função que altera o estado(posição do jogador já atualizada)

Auxiliares para as condições em que o pacman come

eatFoodLittle :: [Player] -> [Player]

Muda os parâmetros dos jogadores quando o pacman come uma comida pequena

eatFoodBig :: [Player] -> [Player]

Muda os parâmetros dos jogadores quando o pacman come uma comida grande

eatFoodBigDie :: [Player] -> [Player]

Muda os parâmetros dos jogadores quando o pacman come uma comida grande mas | encontra um fantasma vivo

foodLittle :: Coords -> [Player] -> Maze -> [Player]

foodBig :: Coords -> [Player] -> [Player]

emptyPiece :: Coords -> [Player] -> Maze -> [Player]

Auxiliares da condição do túnel

findLeftTunnel

Arguments

:: Maze
Labirinto

-> [Coords]
Coordenadas do(s) tunel(eis) da esquerda

Encontra a posição do(s) túnel(eis) da esquerda

findRightTunnel

Arguments

:: Maze
Labirinto

-> [Coords]
Coordenadas do(s) tunel(eis) da direita

Encontra a posição do(s) túnel(eis) da direita

checkLeft

Arguments

:: Player
Jogador

-> Maze
Labirinto

-> Bool

Testa se o jogador se move para o túnel da esquerda

checkRight

Arguments

:: Player
Jogador

-> Maze
Labirinto

-> Bool

Testa se o jogador se move para o túnel da direita

telepLeft

Arguments

:: Player
Jogador

-> Maze
Labirinto

-> Player
Jogador depois de ter sido teletransportado para o lado esquerdo do tunel

Teletransporta o jogador para o túnel da esquerda

telepRight

Arguments

:: Player
Jogador

-> Maze
Labirinto

-> Player
Jogador depois de ter sido teletransportado para o lado direito do tunel

Teletransporta o jogador para o túnel da direita

eatGhosts

Arguments

:: Coords
Coordenadas do pacman

-> [Player]
Lista de jogadores

-> Maze
Labirinto

-> [Player]

Muda os parâmetros dos jogadores quando o pacman come um fantasma morto

ghosts

Arguments

:: Coords
Coordenadas do Pacman

-> [Player]
Lista de jogadores

-> [Int]
Lista de ID's dos fantasmas

Dadas as coordenadas do pacman, corre a lista de jogadores e devolve a lista | de ID's dos fantasmas que se encontram na mesma posição do pacman

Auxiliares dos fantasmas

coordGhostAlive

Arguments

:: [Player]
Lista de jogadores

-> [Coords]
Lista de coordenadas dos fantasmas que estão vivos

Dá as coordenadas das posições onde se encontram fantasmas vivos

coordGhostDead

Arguments

:: [Player]
Lista de jogadores

-> [Coords]
Lista de coordenadas dos fantasmas que estão mortos

Dá as coordenadas das posições onde se encontram fantasmas mortos

Auxiliar da condição da morte

dyingPlayer

Arguments

:: [Player]
Jogador

-> [Player]
Jogador depois de ter perdido uma vida ou ter morrido

Muda os parâmetros do jogador quando este encontra um fantasma vivo

Auxiliares Gerais para posições

coords

Arguments

:: Orientation
Orientação que indica para que lado o jogador quer

-> Player
Jogador

-> Coords
Coordenadas da posição para onde o jogador vai avançar

Dá as coordenadas para as quais é suposto o jogador avançar

poSeg

Arguments

:: Player
Jogador

-> Player
Jogador depois de ter andado uma posição

Move o jogador para a posição seguinte

actualPiece

Arguments

:: Player
Jogador

-> Maze
Labirinto

-> Piece
Peça onde o jogador se encontra

Identifica o tipo de peça onde o jogador se encontra

nextPiece

Arguments

:: Orientation
Orientação do jogador

-> Coords
Coordenadas do Jogador

-> Maze
Labirinto

-> Piece
Peça para onde o jogador irá avançar

Identifica o tipo de peça para o qual o jogador quer avançar

testOrient

Arguments

:: Orientation
Orientação da jogada

-> Player
Jogador

-> Bool
Falso se o jogador puder avançar

Testa se o jogador já está com a mesma orientação pedida pela jogada ou não

mudaOrient

Arguments

:: Orientation
Orientação da jogada

-> Player
Jogador

-> Player
Jogador depois de mudar a orientação

Muda a orientação de um jogador