Tarefa2

Testes

testesT2 :: [(Play, State)]

Testes da Tarefa 2

js :: [Player]

jogador1 :: Player

jogador2 :: Player

jogador3 :: Player

jogador4 :: Player

js2 :: [Player]

j1 :: Player

j2 :: Player

j3 :: Player

j4 :: Player

p1 :: Player

p2 :: Player

p3 :: Player

p4 :: Player

p5 :: Player

play

Arguments

:: Play
Jogada que contem o ID e a direção

-> State
Estado do jogo incial

-> State
Estado depois da jogada

Função principal que executa uma jogada de um jogador

playPacman

Arguments

:: Play
jogada

-> State
estado

-> State
estado depois da jogada

função play para o pacman

playGhost

Arguments

:: Play
jogada

-> State
estado

-> State
estado depois da jogada

play para os ghosts

aumentaNivel

Arguments

:: Int
nivel

-> State
state

-> State
state atualizado

Aumenta o nível do state

tempoMegaList

Arguments

:: [Player]
lista de jogadores

-> [Player]
lista de jogadores atualizada

atualiza o tempo mega do pacman na lista de players

tempoMega

Arguments

:: Player
pacman

-> Player
pacman atualizado

atualiza o tempo mega do pacman

pacmanOrGhost

Arguments

:: Player
jogador

-> Bool
True se for pacman

Verifica se é um pacman

mudaDirJogador

Arguments

:: Play
jogada

-> State
state

-> State
state atualizado

Muda a direção de um jogador num state

openOrCloseMouthList

Arguments

:: [Player]
lista de jogadores

-> [Player]
lista de jogadores atualizada

Atualiza a boca do pacman na lista de jogadores

openOrCloseMouth

Arguments

:: Player
player pacman

-> Player
pacman atualizado

Atualiza a boca do pacman

fantasmasVivos

Arguments

:: [Player]
Lista de jogadores

-> [Player]
Lista de jogadores com os fantasmas todos Vivos

Atualiza a lista de jogadores para os fantasmas ficarem vivos

temVidas

Arguments

:: [Player]
lista de players

-> Bool
devolve True se tiver vidas, False caso contrário

Verifica se o Pacman tem vidas

getPlayer

Arguments

:: Int
ID de um Player

-> [Player]
Lista de Players

-> Player
Player com o respetivo ID

Função que percorre uma lista de Players para encontrar o Player com o respetivo ID

changeDirection

Arguments

:: Play
Jogada que contem o ID e a direção

-> [Player]
Lista de jogadores

-> Bool
Retorna True se For preciso mudar de direçao

Verifica se é preciso mudar de orientação

alteraDirection

Arguments

:: Play
Jogada que contem o ID e a direção

-> [Player]
Lista de Jogadores

-> [Player]
Lista resultante de mudar a direçao do jogador

Altera a orientação de um certo jogador na lista de jogadores

alteraDirecaoPlayer

Arguments

:: Orientation
Orientação nova

-> Player
Jogador

-> Player
Jogador com a nova orientação

Altera a orientação de um player

sameDirection

Arguments

:: Play
Jogada que contem o ID e a direção

-> [Player]
Lista de jogadores

-> Bool
Retorna True se o jogador tem a mesma direção

Verifica se o jogador já está na mesma orientação que a jogada

alteraPosicao

Arguments

:: Play
Jogada que contem o ID e a direção

-> [Player]
Lista de Jogadores

-> [Player]
Lista de jogadores com a posição do jogador com o Id da play mudada

Altera a posição de um jogador na lista de jogadores

alteraPosicaoPlayer

Arguments

:: Orientation
Orientação

-> Player
Jogador

-> Player
Jogador com a posição alterada

Altera a posição de um certo jogador

somaCoords

Arguments

:: Coords
Coordendas

-> Orientation
Orientação

-> Coords
Coordenadas atualizadas dependendo da orientação

Calcula a posição seguinte dado uma orientação

fantasmaMazeDead

Arguments

:: Play
Jogada que contem o ID e a direção

-> Maze
Labirinto

-> [Player]
Lista de jogadores

-> Bool
Retorna True se estiver um fantasma morto na mesma posiçao que o jogador do Id da play

Verifica se o fantasma que está na posição da jogada do player está morto

fantasmaMazeAlive

Arguments

:: Play
Jogada que contem o ID e a direção

-> Maze
Labirinto

-> [Player]
Lista de jogadores

-> Bool
Retorna True se estiver um fantasma vivo na mesma posiçao que o jogador do Id da play

Verifica se o fantasma que está na posição da jogada do player está vivo

fantasmaPos

Arguments

:: [Player]
Lista de Jogadores

-> Coords
Coordenadas

-> Bool
Retorna True se estiver um fantasma vivo na posição das coordenadas

Verifica se está um fantasma vivo nas coordenadas dadas

fantasmaPosD

Arguments

:: [Player]
Lista de Jogadores

-> Coords
Coordenadas

-> Bool
Retorna True se estiver um fantasma morto na posição das coordenadas

Verifica se está um fantasma morto nas coordenadas dadas

contaFantasmas

Arguments

:: [Player]
Lista de Jogadores

-> Coords
Coordenadas

-> Int
Retorna o nº de fantasmas que se encontram em x posição

Conta quantos fantasmas se encontram em x coordenadas

fantasmaPosDTrack

Arguments

:: [Player]
Lista de Jogadores

-> Coords
Coordenadas

-> Int
ID do fantasma que está nas coordenas

Devolve o ID do fantasma morto nas coordenadas dadas

alteraPontuacao

Arguments

:: Int
Pontos

-> Play
Jogada que contem o ID e a direção

-> [Player]
Lista de jogadores

-> [Player]
Lista de jogadores com o jogador os pontos atualizados

Altera a pontuação de um jogador na lista de jogadores

alteraPontuacaoPlayer

Arguments

:: Int
Pontos

-> Play
Jogada que contem o ID e a direção

-> Player
Jogador

-> Player
Jogador com os pontos atualizados

Altera a pontuação de um jogador

alteraVidas

Arguments

:: Play
Jogada que contem o ID e a direção

-> [Player]
Lista de jogadores

-> [Player]

Altera as vidas de um jogador na lista de jogadores

alteraVidasPlayer

Arguments

:: Play
Jogada que contem o ID e a direção

-> Player
Jogador

-> Player
Jogador com as vidas alteradas

Altera as vidas de um jogador

atualizaFantasma

Arguments

:: [Player]
Lista de jogadores

-> [Player]
Lista de jogadores com os fantasmas todos mortos

Atualiza o mode dos fantasmas todos para mortos

meioFantasma

Arguments

:: [Player]
Lista de jogadores

-> Coords
Coordenadas

-> Maze
Labirinto

-> [Player]
Lista de jogadores com o fantasma depois de ser comido no meio

Coloca o fantasma na casa dos fantasmas quando é morto por um pacman

setGhostAlive

Arguments

:: Player
Ghost

-> Player
Ghost em mode Alive

Coloca o fantasma em modo alive

meio

Arguments

:: Maze
Labirinto

-> Coords
Coordenadas

Calcula o meio do labirinto

isMega

Arguments

:: [Player]
Lista de jogadores

-> Bool
Retorna True se o Pacman estiver Mega

Verifica na lista de jogadores se encontra o Pacman se encontra em estado Mega

isFantasmaAlive

Arguments

:: Piece
Peça do labirinto

-> Bool
Retorna True se o fantasma estiver vivo

Verifica se o fantasma se encontra vivo

isFantasmaDead

Arguments

:: Piece
Peça do labirinto

-> Bool
Retorna True se o fantasma estiver morto

Verifica se o fantasma se encontra morto

isComidaP

Arguments

:: Play
Jogada que contem o ID e a direção

-> Coords
Coordenadas

-> Maze
Labirinto

-> Bool
Retorna True se nas coordenanas estiver uma comida pequena

Verifica se o jogador vai jogar para a posição onde se encontra uma comida pequena

isComidaG

Arguments

:: Play
Jogada que contem o ID e a direção

-> Coords
Coordenadas

-> Maze
Labirinto

-> Bool
Retorna True se nas coordenanas estiver uma comida grande

Verifica se o jogador vai jogar para a posição onde se encontra uma comida grande

isWall

Arguments

:: Play
Jogada que contem o ID e a direção

-> Coords
Coordenadas

-> Maze
Labirinto

-> Bool
Retorna True se nas coordenanas estiver uma comida grande

Verifica se o jogador vai jogar para a posição onde se encontra uma parede

pieceMaze

Arguments

:: Coords
Coordenadas

-> Maze
Labirinto

-> Piece
Peça que esta nas coordenadas

Devolve a peça que se encontra nas coordenadas

indiceLista

Arguments

:: Int
Indice da peça do corredor

-> Corridor
Corredor

-> Piece
Peça que está no indice do corredor

Devolve a peça que está no indice n do corredor

isTunel

Arguments

:: Play
Jogada que contem o ID e a direção

-> Coords
Coordenadas

-> Maze
Labirinto

-> Bool
Retorna True se for entrar no labirinto

Verifica se o jogador se encontra num tunel

passarTunel

Arguments

:: Maze
Labirinto

-> Play
Jogada que contem o ID e a direção

-> [Player]
Lista de jogadores

-> [Player]
Lista de jogadores com a posição do jogador atualizada depois de passar o túnel

Altera na lista de jogadores o jogador que passou pelo tunel

passarTunelPlayer

Arguments

:: Maze
Labirinto

-> Play
Jogada que contem o ID e a direção

-> Player
Jogador

-> Player
Jogador na nova posição depois de passar o túnel

Altera as coordenadas de um jogador de maneira a passar pelo tunel

limparComida

Arguments

:: Coords
coordenadas onde se encontra a comida

-> Maze
labirinto

-> Maze
labirinto limpo

Limpa a comida que o Pacman comeu

atualizaIndiceLista

Arguments

:: Int
indice que é para alterar

-> Piece
peça para alterar

-> Corridor
corredor para alterar

-> Corridor
corredor alterado

Atualiza indice n do Corredor

acabouComida

Arguments

:: Coords
coordenadas para percorrer no labirinto

-> Maze
labirinto

-> Bool
lista de coordenadas onde se encontram as comidas

Retorna True se acabou a comida no labirinto

encontraFood

Arguments

:: Coords
coordenadas para percorrer o corredor

-> Corridor
corredor

-> Bool
lista de coordenadas de onde se encontram as comidas

Percorre um corredor à procura de Food Big e devolve a lista de coordenadas onde se encontram