Tarefa2

Description

O objetivo deste módulo é executar uma jogada de um pacman ou de um fantasma atraves do numero de identidade deste e de uma orientação. No caso do Pacman temos que verifica a peca para qual o pacman se move fazendo as respectivas mudancas no seu estado, na pontuação e na sua vida. o estado muda para mega no caso do Pacman comer uma big food que tambem provoca uma mudanca de estado para todos os fantasmas do jogo para Dead o estado muda para dying no caso do pacman ter 0 vidas e bater contra um fantasma no estado alive se isto acontecer mas o pacman tem mais vidas é lhe retirado uma a pontuacao muda em 3 casos: -come little food é lhe fornecido +1 pontos -come big food é lhe fornecido +5 pontos -come um fantasma em estado dying é lhe fornecido +10 pontos

Documentation

procuramaze

Arguments

:: Maze
recebe um labirinto

-> Int
recebe um indice

-> Corridor
retorna um corredor

Função que procura num Maze um corredor que esta no indice y

procuracorridor

Arguments

:: Corridor
recebe um corredor

-> Int
recebe um indice

-> Piece
retorna a peca que esta nesse indice

Procura uma peça que esta no indice x

substituicorredor

Arguments

:: Int
recebe um indice

-> Corridor
recebe um corredor

-> Corridor
retorna um corredor modificado

Função similar a procuracorridor que procura um indice de um corredor e substitui o seu conteudo por Empty pois o pacman comeu a peça

substituipeca

Arguments

:: Coords
recebe duas coordenadas

-> Maze
recebe um labirinto a ser modificado

-> Maze
retorna um labirinto modificado

Funçao similar a procuramaze que atraves de coordenadas procura o respectivo corredor e depois envia para a função substituicorredor para este modificar a a peca no caso (-1,_) acontece quando a peca nao precisa de ser substituida

estadojogador

Arguments

:: Piece
recebe uma peça

-> [Player]
recebe uma lista de Players

-> [Player]
retorna uma lista de players com as respectivas mudancas

Função que faz as respectivas mudancas no estado,pontuação e velocidade consoante a peca dada

identificaplayer

Arguments

:: Int
recebe um valor identificador

-> [Player]
recebe uma lista

-> Player
retorna o jogador que corresponde ao inteiro

Identifica um player atraves do seu ID

mudadead

Arguments

:: Int
recebe o comprimento do maze

-> Int
recebe a altura do maze

-> (Int, Int)
recebe dois ID o primeiro do pacman e o segundo do ghost a ser modificado

-> [Player]
recebe uma lista de jogadores

-> [Player]
retorna uma lista de jogadores com o par de pac e ghost modificado

Função que muda o estado do ghost e de pac se tiveram uma colisao o caso (-1,-1) é quando não existe colisao e o caso (p,-1) e quando apenas se modifica o pacman

verificacolisao

Arguments

:: Player

-> [Player]
recebe uma lista de jogador

-> (Int, Int)
retorna um par de ID (pacman,ghost) que teve em colisao

Verifica se o jogador que se moveu entra em colisao com algum outro jogador se não houve colisao é retornado (-1,-1) e se for apenas preciso modificar o pacman e retornado (pacman,-1)

playPlayer

Arguments

:: Maze

-> Play
recebe a jogada

-> [Player]
recebe a lista de jogadores

-> ([Player], Coords)
retorna a lista modificada e a coordenada modificada para pacman

vai ha lista de players identifica qual o player por o identificador faz os testes da parede e do tunel depois modifica as coordenadas do jogador e retorna tambem essas mesmas coordenadas para modificar o labirinto depois

play

Arguments

:: Play
recebe uma jogada

-> State
recebe um estado

-> State
retorna o estado modificado por a jogada

Funçao mãe que junta tudo e manda para as funções que modificam o mapa e tambem a lista de jogadores