Tarefa3

Documentation

testMazeEnunc :: [[Piece]]

compactMaze :: Maze -> Instructions

compactMaze

Função principal

Padrões Horizontais

horizontal

Arguments

:: Corridor
Corredor

-> Instruction
Instruções

horizontalP

Utilizando a patternH constroi uma Instruction que devolve a lista de instruções de um corredor

patternH

Arguments

:: (Int, Piece)
Primeiro par que vai ser comparado aos restantes da lista

-> Corridor
Corredor

-> [(Int, Piece)]
Lista de pares(repetidos estão compactados)

patternH

Auxiliar para verificar os padrões Devolve uma lista de instruções compactadas(junta as que são repetidas sucessivamente)

Padrões Verticais

compactMazeAux

Arguments

:: Instructions
Instruções do primeiro corredor

-> Maze
Labirinto

-> Instructions
Instruções finais

compactMazeAux

Compacta as instruções eliminando as repetições de corredores iguais sucessivos