Tarefa3

Description

O objetivo deste módulo e compactar o mapa em instruções que são uma lista de instruct que podem ser um conjunto [(Int,Piece)] que o int representa o numero de vezes que essa peça aparece no corredor de seguida ou repeat Int que so acontece quando a mesma instruct ja aconteceu anteriormente onde o Int representa o indice na lista. Utilizei par [(Int,Piece)] em vez de instruct na maior parte das funções para facilitar os testes.

Documentation

compactcorridoraux

Arguments

:: Corridor
recebe um corridor

-> [(Int, Piece)]
dá uma lista de pares (int,Piece)

Função que pega num corredor e tranforma de maneira simples para (1,piece)

juntaseguidos

Arguments

:: [(Int, Piece)]
recebe uma lista de pares

-> [(Int, Piece)]
dá uma lista de pares ordenados sem peças seguidas repetidas

Função que junta as peças seguidas como esta proposto na Tarefa

compactVertical

Arguments

:: Int
recebe um inteiro que acumula até a instrução ser igual

-> Instructions
recebe as instruçoes

-> Instructions
dá uma lista de instruçoes com os respetivos Repeats

Funçao que verifica se há corredores iguais para substituir os a seguir por Repeat int e usa um acumulador para sabermos qual e o indicie para usarmos na compactVerticalAux

compactVerticalAux

Arguments

:: Int
recebe uma posição

-> Instruction
recebe uma instrução

-> Instructions
recebe as instruçoes onde sabemos que existe uma instrução repetida

-> Instructions
Devolve as intruçoes com o respetivo Repeat

Função aux que recebe o indice do primeiro corredor que tem pelo menos um igual na lista e verifica na lista toda os que são iguais e substitui estes por Repeat indice

compactcorridor :: [(Int, Piece)] -> Instruction

Função que recebe uma lista de (Int,Piece) e transforma numa Instruction

compactMazeAux

Arguments

:: Maze
recebe um labirinto

-> Instructions
dá uma lista de instruçoes com os respetivos corredores compactados (

Função que compacta primeiro os corredores horizontais (muda para instructs e junta) e depois compacta os restantes corredores

compactMaze

Arguments

:: Maze
recebe um labirinto

-> Instructions
dá o labirinto em respetivas Instruções com sinalização nas repetiçoes.

Função final em que dá o labirinto compactado os corredores todos (Verticais) utilizando o 0 como valor inicial para o acumulador.