Tarefa4

Description

Introdução

A função principal da Tarefa4 é a passTime, que basicamente atualiza o State, fazendo as jogadas de todos os jogadores, ou seja esta Tarefa é a responsável pela passagem de tempo.

Objetivos

O objetivo desta Tarefa é fazer com que o Pacman e os Fantasmas joguem ao longo do tempo. Para isso criamos funções para o Pacman e para os Fantasmas, uma vez que estes últimos mudam de velocidade e por isso jogam de acordo com a mesma: se esta for 0.5 apenas jogam de 2 em 2 steps, se for 1 não há restrições.

Discussão e conclusão

Conseguimos alcançar os objetivos da Tarefa 4, no entanto contempla apenas 2 valores diferentes de velocidade: 1 e 0.5.

Documentation

defaultDelayTime :: Integer

Testes

testesT4 :: [(Int, State)]

passTime

Arguments

:: Int
step

-> State
state

-> State
state atualizado

atualizar o estado a jogar

jogar

Arguments

:: [Player]
lista de jogadores

-> State
state

-> Int
step

-> State
state atualizado

Atualiza o estado depois das jogadas de cada lista de players

jogarFant

Arguments

:: [Play]
lista de jogadas dos fantasmas

-> Player
fantasma

-> State
state

-> Int
step

-> State
state

Função para os fantasmas jogarem

getPacmanID

Arguments

:: [Player]
lista de jogadores

-> Int
id do pacman

Devolve o id do pacman