Tarefa4

Description

O objetivo desta tarefa é fazer com que o tempo seja um fator no nosso jogo. Por isto tivemos que ter em conta os seguintes fatores: -Fazer com que a boca do pac mude que foi resolvido com uma simples funcao que procurava o pacman na lista e mudava a boca -Com que o "tempo Mega" decresca e o estado do fantasma reagisse a isso -Executar as jogadas que foram fabricadas na tarefa 5 e tambem as do pacman

Documentation

defaultDelayTime :: Integer

Tempo default

mudaboca

Arguments

:: [Player]
recebe uma lista de jogadores

-> [Player]
devolve a lista de jogadores com a boca mudada

Funcao que muda a boca a cada iteracao

voltaNormal

Arguments

:: [Player]
recebe uma lista de jogadores

-> [Player]

Funcao que faz os ghost voltarem ao Alive quando o timemega acabar

baixaTempo

Arguments

:: [Player]
recebe uma lista de jogadores

-> ([Player], Bool)
retorna a lista com time-1 e um bool que identifica se o time mega chegou a 0

So acontece quando player esta em mega e retorna um par com a lista e com um bool que representa se o mega acabou ou nao

getOrientation

Arguments

:: [Player]
recebe uma lista de jogadores

-> (Orientation, Int)
retorna um par (orentacao,identificador) a orientacao e a que o identificador vai tomar

Funcao que percorre a lista até encontrar o pacman e depois envia a orientacao e a sua identicacao para executar a play do pacman

pacMove

Arguments

:: State
recebe um estado

-> State
retorna um estado com a jogada do pacman

Funcão que executa a jogado do Pacman para a orientacao que ele se encontra

changedirection

Arguments

:: Orientation
recebe uma orientacao

-> [Player]
recebe uma lista de jogadores

-> [Player]
retorna a lista com a orientacao do pacman mudada

Uma funcao que muda a orientacao para podermos usar na funcao na funcao do main updateControlledPlayer

applyghostplays

Arguments

:: Int
recebe o step

-> State
recebe um estado

-> [Play]
recebe uma lista de jogadores

-> State
retorna o state com as jogadas de todos os ghosts

Uma funcao que executa todas as plays calculadas na tarefa 5 analisando tambem o step

passTime

Arguments

:: Int
recebe um step

-> State
recebe um estado

-> State
retorna um estado com todas as alteracoes

Funcao principal que faz o tempo passar, verifica se ha time mega e envia os dados para as funcoes necessarias