Tarefa5

Description

Introdução

As estratégias implementadas para a realização desta tarefa passaram por fazer com que os fantasmas agissem de uma certa forma se estes estiverem no modo Alive e para outra no modo Dead.

Objetivos

O principal objetivo desta tarefa consistia em fazer jogadas para os fantasmas, para cada situação, mais especificamente para quando eles estão no modo Alive ou no modo Dead. No caso de estarem Alive teriam que perseguir o Pacman, para isso completamos a função chaseMode, caso contrário fugiam do mesmo e então criamos a função scatterMode. Resumidamente, o objetivo dos fantasmas é ficar na mesma posição que o Pacman.

Discussão e conclusão

Concluindo, a Tarefa 5 realiza aquilo que é pretendido para as diferentes situações, os fantasmas perseguem o Pacman de acordo com a sua posição.

Documentation

testesT5 :: [State]

ghostPlay

Arguments

:: State
Estado do jogo

-> [Play]
Lista de jogadas dos fantasmas

Devolve uma lista de jogadas para todos os fantasmas

ghostPlayAux

Arguments

:: State
Estado do jogo

-> Int
Indice da lista de jogadores

-> [Play]
Lista de jogadas dos fantasmas

Funcao auxiliar para percorrer a lista de jogadores de modo a todos os fantasmas jogarem

chaseMode

Arguments

:: State
Estado do jogo

-> Int
Id do fantasma

-> Play
Jogada em direçao ao Pacman

Modo de perseguir o Pacman

scatterMode

Arguments

:: State
Estado do jogo

-> Int
Id do fantasma

-> Play
Jogada na direçao contraria ao Pacman

Modo de fugir do Pacman

melhorJogadaChase

Arguments

:: Coords
Coordenadas do fantasma

-> Coords
Coordenadas do Pacman

-> Maze
Labirinto

-> Orientation
Melhor orientacao

Encontra a melhor orientaçao para perseguir o Pacman

melhorJogadaScatter

Arguments

:: Coords
Coordenadas do fantasma

-> Coords
Coordenadas do Pacman

-> Maze
Labirinto

-> Orientation
Melhor orientacao

Encontra a melhor orientaçao para fugir do Pacman

isWall2

Arguments

:: Coords
Coordenadas

-> Maze
Labirinto

-> Bool
Retorna True se nas coordenadas estiver uma parede

Verifica se o jogador vai jogar para a posição onde se encontra uma parede

getPacmanCoords

Arguments

:: [Player]
Lista de jogadores

-> Coords
Coordenadas do pacman

Devolve as coordenadas do pacman

isAlive

Arguments

:: Player
Jogador

-> Bool
Retorna True se estiver vivo e False se esta morto

Verifica se o fantasma está morto ou vivo