Tarefa5

Description

Documentation

mudaestadochase

Arguments

:: Player
recebe um Ghoststate

-> Player
recebe um Pacstate

-> [(Int, Orientation)]
retorna uma lista com as distancias e respetivas orientações

Função que devolve listas com as distancias e respetivas orientações para cada orientação

mudaorientacaochase

Arguments

:: Maze
recebe um maze

-> Player
recebe um ghostate

-> [(Int, Orientation)]
recebe uma lista das distâncias e respetivas orient provenientas da função mudaestadochase

-> Orientation
muda a orientação analisando a melhor distancia sem parede(recursivamente)

função que analisa o menor e verifica se está ou não numa parede

chaseMode

Arguments

:: State
recebe um estado de jogo

-> Int
recebe o id do Player

-> Play
retorna o estado com a jogada a seguir

função retorna uma jogada com a melhor orientaçao possivel

identificaPac

Arguments

:: [Player]
recebe a lista dos state

-> Player

Função para identificar o Pacman

mudaorientacaoscatter

Arguments

:: Maze
recebe um labirinto

-> Player
recebe um ghostate

-> Orientation
retorna a orientação a seguir

função que muda a orientação do ghost quando está a fugir sempre para a direita quando encontra uma parede

scattermode

Arguments

:: State
recebe um estado

-> Int
recebe um id

-> Play
retorna uma jogada com a respetiva orientação

função que retorna a jogada quando o ghost está a fugir

getPlay

Arguments

:: Player
recebe um ghostate

-> State
recebe um estado

-> Play
retorna a jogada com a orientação respetiva

função que analisa o modo do ghost e manda para a respetiva função

nextghost

Arguments

:: Int
recebe um acumulador

-> [Player]
recebe a lista de jogadores do state

-> (Player, Int)
devolve o jogador e um inteiro que nos da a informacao se a lista acabou ou nao ou se e pacman

Funcao que encontra o proximo fantasma a ser jogado e um inteiro para o identificar o caso de a fila acabar e tambem de ser o pacman

ghostPlayAux

Arguments

:: Int
recebe um acumulador

-> State
recebe um state

-> [Play]
retorna a lista de plays

Funcao auxiliar que fazem a lista de play

ghostPlay

Arguments

:: State
recebe um estado

-> [Play]
retorna uma lista de jogadas

Funcao principal que faz as plays