Tarefa6

Description

Introdução

Esta tarefa consistiu em implementar um bot do Pacman, que realizasse certas jogadas para cada situação de jogo, ou seja cada State. Nomeadamente um jogador que fugisse dos fantasmas, enquanto estes estivessem vivos, procurasse as comidas grandes, de modo a ficar no modo Mega e neste último caso perseguisse os fantasmas, de maneira a que os comesse.

Objetivos

As nossas estratégias para a realização desta tarefa resumem-se em verificar certas situações: - se o pacman está em modo Mega - se o pacman está perto de fantasmas - se o pacman está perto de comidas grandes E escrevemos funções para esses diferentes casos, de modo a que o bot do Pacman fizesse diferentes jogadas.

Discussão e conclusão

Concluindo, a Tarefa 6 realiza aquilo que é pretendido e o bot desloca-se para as diferentes situações, no entanto de uma forma rudimentar.

Documentation

testesT6 :: [(Int, State)]

bot

Arguments

:: Int
id do jogador

-> State
estado que contém o labirinto, lista de jogadores e nível

-> Maybe Play
possível jogada

Aplica jogadas ao bot dependendo do estado do jogo

distMan

Arguments

:: Coords
coordenadas

-> Coords
coordenadas

-> Int
distância calculada

Calcula a distância de Manhattan entre duas Coordenadas.

encontraFoodMaze

Arguments

:: Coords
coordenadas para percorrer no labirinto

-> Maze
labirinto

-> [Coords]
lista de coordenadas onde se encontram as comidas

Percorre o labirinto à procura de 'Food Big' e devolve a lista de coordenadas onde se encontram

encontraFoodBig

Arguments

:: Coords
coordenadas para percorrer o corredor

-> Corridor
corredor

-> [Coords]
lista de coordenadas de onde se encontram as comidas

Percorre um corredor à procura de Food Big e devolve a lista de coordenadas onde se encontram

coordMaisPerto

Arguments

:: [Coords]
lista de coordenadas a comparar

-> Coords
coordenadas a comparar

-> Coords
coordenadas mais proximas

Procura numa lista de coordenadas, a mais perto de outras coordenadas

coordsFantasmas

Arguments

:: [Player]
lista de jogadores

-> [Coords]
lista de coordenadas dos fantasmas

Devolve a lista de coordenadas dos fantasmas

chaseComida

Arguments

:: State
Estado que contem o labirinto, lista de jogadores e nível

-> Bool
True se a peça mais perto for 'Food Big'

Verifica se a peça mais perto do Pacman é 'Food Big'