Tarefa6

Description

O objetivo desta tarefa é calcular a melhor jogada para o Pacman, para isto, nos calculamos o ghost mais perto do pacman e dependendo do do seu modo,Normal ou Mega, fugia dele ou perseguia-o

Documentation

closestGhost

Arguments

:: Player
recebe o pacman

-> [Player]
recebe a lista de jogadores

-> [(Int, Int)]
retorna uma lista com o identificador e a distancia do pacman ao respetivo ghost

Calcula a distancia do pac a cada ghost dos players e faz uma lista com a respectiva distancia e identificador

moveBot

Arguments

:: Maze
recebe o labirinto

-> Player
recebe o pacman

-> Player
recebe o ghost mais perto do pacman

-> Orientation
retorna a melhor orientacao para o pacman tomar

Funcao que calcula a melhor jogada para o pacman

bot

Arguments

:: Int
recebe o identificador do pacman

-> State
recebe um estado

-> Maybe Play
retorna a melhor orientacao que o pacman pode tomar

Bot calcula uma lista com distancia dos ghost e ordenada do mais perto para o mais longe depois calcula a melhor jogada para o respectivo modo de pac