Types

Description

Este módulo define os tipos de dados comuns a várias tarefas.

Tipos de dados auxiliares.

data PlayAction

As várias ações dos jogadores.

Constructors

EatLittle
Come comida pequena.

EatBig
Come comida grande.

EatGhost
Come um fantasma.

Die
Morre.

MoveFront
Move na mesma direção.

Collide
Bate contra uma parede

data State

Um estado do jogo.

Constructors

State

Fields

maze :: Maze

Labirinto.

playersState :: [Player]

Lista de jogadores.

level :: Int

Nível do jogo.

type Maze = [Corridor]

Labirinto é um conjunto de corredores.

type Corridor = [Piece]

Corredor é um conjunto de peças.

data Piece

As várias peças de um jogo.

Constructors

Food FoodType
Peça do tipo comida.

PacPlayer Player
Peça do tipo jogador.

Empty
Peça do tipo vazio.

Wall
Peça do tipo parede.

data Player

Os dois tipos de jogador.

Constructors

Pacman PacState
Jogador do tipo pacman.

Ghost GhoState
Jogador do tipo fantasma.

data Orientation

As várias orientações.

Constructors

L
Esquerda.

R
Direita.

U
Para cima.

D
Para baixo.

Null
Fica no sítio.

data PacState

Um jogador do tipo pacman.

Constructors

PacState

Fields

pacState :: PlayerState

Estado do pacman.

timeMega :: Double

Tempo que resta na posição Mega.

openClosed :: Mouth

Maneira que se encontra a boca.

pacmanMode :: PacMode

Como se encontra o pacman (Normal, Mega ou Dying).

data Manager

Um manager do jogo.

Constructors

Manager

Fields

state :: State

Estado do jogo.

pid :: Int

Id do jogador a controlar.

step :: Int

Número que indica quantas iterações o jogo já passou.

before :: Integer

Intervalo de tempo entre jogadas.

delta :: Integer

Instante de tempo em que foi efetuada a última jogada.

delay :: Integer

Tempo decorrido desde a última jogada (milissegundos).

data GhoState

Um jogador do tipo ghost.

Constructors

GhoState

Fields

ghostState :: PlayerState

Estado do ghost.

ghostMode :: GhostMode

Modo do ghost (Alive ou Dead).

type Coords = (Int, Int)

Um par de coordenadas do tipo (linha,coluna).

type PlayerState = (Int, Coords, Double, Orientation, Int, Int)

Como é definido o estado de um jogador.

data Mouth

Tipos de boca de um pacman.

Constructors

Open
Boca encontra-se aberta.

Closed
Boca encontra-se fechada.

data PacMode

Tipos de modo de um pacman.

Constructors

Dying
Pacman está no modo morto.

Mega
Pacman está no modo mega.

Normal
Pacman está no modo normal.

data GhostMode

Tipos do fantasma.

Constructors

Dead
Fantasma encontra-se morto.

Alive
Fantasma encontra-se vivo.

data FoodType

Tipos de comida.

Constructors

Big
A comida é grande.

Little
A comida é pequena.

data Color

Tipos de cor.

Constructors

Blue
Azul.

Green
Verde.

Purple
Roxo.

Red
Vermelho.

Yellow
Amarelo.

None
Nenhuma.

data Play

O tipo de jogadas que vão ocorrer no jogo.

Constructors

Move Int Orientation

type Instructions = [Instruction]

Tipo de instruções.

data Instruction

Tipo de instruções.

Constructors

Instruct [(Int, Piece)]

Repeat Int

coloredString :: String -> Color -> String

Funções auxiliares usadas ao longo do trabalho.

placePlayersOnMap :: [Player] -> Maze -> Maze

Coloca os jogadores num labirinto.

printMaze :: Maze -> String

Dá print de um labirinto.

printPlayerStats :: Player -> String

Dá print das estatísticas de um jogo.

getPlayerID :: Player -> Int

Obtém o ID de um jogador.

getPlayerPoints :: Player -> Int

Obtém a pontuação de um jogador

setPlayerCoords :: Player -> Coords -> Player

Altera as coordenadas de um jogador.

getPacmanMode :: Player -> PacMode

Obtém o modo de um pacman.

getGhostMode :: Player -> GhostMode

Obtém o modo de um fantasma.

getPlayerState :: Player -> PlayerState

Obtém o estado de um jogador.

getPlayerOrientation :: Player -> Orientation

Obtém a orientação de um jogador.

getPlayerMouth :: Player -> Mouth

Obtém a boca de um jogador.

replaceElemInMaze :: Coords -> Piece -> Maze -> Maze

Troca uma peça de um labirinto pela peça que damos.

replaceNElem :: Int -> a -> [a] -> [a]

Auxiliar da função replaceNElemInMaze.

getPlayerCoords :: Player -> Coords

Obtém as coordenadas de um jogador.

isPacman :: Player -> Bool

Verifica se um jogador é do tipo pacman.

getPlayer :: Int -> [Player] -> Player

Devolve um jogador consoante o ID que recebe.

replacePlayer :: Player -> State -> State

Substitui um jogador num estado.

replacePlayerAux :: Player -> [Player] -> [Player]

Substitui um jogador numa lista de jogadores.

getPlayerFromList :: Int -> State -> Player

Remove um jogador de um estado.

removePlayer :: Player -> [Player] -> [Player]

Remove um jogador de uma lista de jogadores.

ordena :: [Player] -> [Player]

Ordena uma lista de jogadores para aparecer primeiro o pacman.

isPacmanMega :: Player -> Bool

Verifica se um pacman está em modo mega.

isGhost :: Player -> Bool

Verifica se o jogador é do tipo fantasma.

getTimeMega :: Player -> Double

Obtém o tempo mega de um pacman.

whereIsPacman :: [Player] -> Player

A função whereIsPacman utiliza a lista de jogadores e devolve o jogador do tipo Pacman.

whereAreGhosts :: [Player] -> [Player]

A função whereAreGhosts utiliza a lista de jogadores e devolve apenas os do tipo Ghost.

getTimeMegaFromState :: State -> Double

Obtém o tempo meda do pacman de um determinado estado.

getPlayers :: State -> [Player]

Obtém a lista de jogadores do estado.

getMaze :: State -> Maze

Obtém um labirinto de um estado.

removePlayerFromState :: Player -> State -> State

Remove um jogador de um estado.

removeAux :: Player -> [Player] -> [Player]

Função auxiliar da removePlayerFromState

nextPiece :: Orientation -> Player -> Maze -> Piece

Obtém a peça que se encontra à frente do jogador.

getPiece :: Coords -> Maze -> Piece

Esta pequena função serve para obter a peça que está na posição que damos como argumento.

rotate180Degrees :: Orientation -> Orientation

Muda a orientação do jogador em 180 graus.

changeMouth :: Mouth -> Mouth

Muda a boca do jogador.